% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUs.R
\name{add_dp_label}
\alias{add_dp_label}
\title{Modify Data Pack Org Unit list to add datapackr IDs.}
\usage{
add_dp_label(orgunits)
}
\arguments{
\item{orgunits}{Data frame of Data Pack org units produced by \code{\link{getDataPackOrgUnits}}.}
}
\value{
Data frame of Data Pack Org units with added Data Pack label, \code{dp_psnu}.
}
\description{
Adds Org Unit label used in Data Packs.
}
