% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUs.R
\name{add_dp_psnu}
\alias{add_dp_psnu}
\title{Modify PSNU list to add datapackr IDs.}
\usage{
add_dp_psnu(PSNUs)
}
\arguments{
\item{PSNUs}{Data frame of PSNUs produced by \code{\link{getPSNUs}}.}
}
\value{
Data frame of PSNUs with added Data Pack label, \code{dp_psnu}.
}
\description{
Adds PSNU label used in Data Packs.
}
