% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addPsnuid}
\alias{addPsnuid}
\title{addPsnuid}
\usage{
addPsnuid(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible
to one.}
}
\value{
A logical vector.
}
\description{
Tests whether a character string matches the regex of a DHIS2
11-digit UID. Vectorized over \code{string} and \code{pattern}.
}
