% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packFrame.R
\name{addValidationsSite}
\alias{addValidationsSite}
\title{Add validations tab and options to Site tool}
\usage{
addValidationsSite(wb)
}
\arguments{
\item{wb}{An Openxlsx workbook object}
}
\description{
Adds "Validations" tabe and all necessary validation options to \code{wb}.
}
