% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_import_file.R
\name{adorn_import_file}
\alias{adorn_import_file}
\title{Convert PSNU-level DATIM import file into analytics-friendly object}
\usage{
adorn_import_file(
  psnu_import_file,
  cop_year = getCurrentCOPYear(),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{psnu_import_file}{DHIS2 import file to convert}

\item{d2_session}{}
}
\value{
data
}
\description{
Convert PSNU-level DATIM import file into analytics-friendly
object, similar to the MER Structured Datasets
}
