% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{adornMechanisms}
\alias{adornMechanisms}
\title{adornMechanisms(d)}
\usage{
adornMechanisms(d)
}
\arguments{
\item{d}{}
}
\value{
Modified d object with d$data$analtyics
}
\description{
Join analytical dimensions with d$data$analtyics related
to partner, agency and mechanism information.
}
