% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSheet.R
\name{aggregateSheets}
\alias{aggregateSheets}
\title{Aggregates sheets based on sheet type in the Datapack.}
\usage{
aggregateSheets(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Aggregates sheets based on the sheet code. Currently handles
OVC, PMTCT_EID and KP.
}
