% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnalytics.R
\name{analyze_eid_2mo}
\alias{analyze_eid_2mo}
\title{Check Data Pack for <90\% PMTCT_EID from ≤02 months}
\usage{
analyze_eid_2mo(data)
}
\arguments{
\item{data}{Analytics object to analyze}
}
\value{
a
}
\description{
Check data gathered from Data Pack to identify cases where
the targeted percent of infants (<1 yr) born to HIV-positive women tested for
HIV (EID) between 0 and 2 months old is less than 90\%.
}
