% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnalytics.R
\name{analyze_indexpos_ratio}
\alias{analyze_indexpos_ratio}
\title{Check Data Pack data for low representation of HTS_INDEX_POS}
\usage{
analyze_indexpos_ratio(data)
}
\arguments{
\item{data}{Analytics object to analyze}
}
\value{
a
}
\description{
Check data gathered from Data Pack to identify cases where
the proportion of HTS_TST_POS represented by HTS_INDEX_POS is too low for the
given ART Coverage rate:

\tabular{cc}{
  \strong{ART Coverage}\tab  \strong{HTS_INDEX_POS \% of HTS_TST_POS}\cr
  <70\%\tab  30\%\cr
  70\% <= x < 80\%\tab  50\%\cr
  >= 80\%\tab  75\%\cr 
}
}
