% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnalytics.R
\name{analyze_linkage}
\alias{analyze_linkage}
\title{Check Data Pack data for linkage < 95\% or >100\%}
\usage{
analyze_linkage(data)
}
\arguments{
\item{data}{Analytics object to analyze}
}
\value{
a
}
\description{
Check data gathered from Data Pack to identify cases where
linkage rates are less than the standard of 95\% or >100\%.
}
