% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnalytics.R
\name{analyze_vmmc_indeterminate}
\alias{analyze_vmmc_indeterminate}
\title{Check Data Pack data for VMMC_CIRC Indeterminate Rate > 5\%}
\usage{
analyze_vmmc_indeterminate(data)
}
\arguments{
\item{data}{Analytics object to analyze}
}
\value{
a
}
\description{
Check data gathered from Data Pack to identify cases where
the expected number of VMMC_CIRC Indeterminate patients is greater than
5\% of the total number of VMMC_CIRC patients.
}
