% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R
\name{api_call}
\alias{api_call}
\title{Start DATIM API query and specify table}
\usage{
api_call(endpoint, d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{endpoint}{Character. DATIM API endpoint to query.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Web-encoded URL for DATIM API query.
}
\description{
Constructs URL for DATIM API query against specified table without paging.
}
