% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R
\name{api_fields}
\alias{api_fields}
\title{Select fields to return from a DATIM API query.}
\usage{
api_fields(api_call, fields)
}
\arguments{
\item{fields}{Fields to return. No need to include \code{&fields=}.}
}
\value{
Web-encoded URL for DATIM API query.
}
\description{
Specifies fields to return from DATIM API query and encodes for web.
}
