% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R
\name{api_filter}
\alias{api_filter}
\title{Filter a DATIM API query.}
\usage{
api_filter(api_call, filter)
}
\arguments{
\item{api_call}{Base DATIM API query, specifying API table and setting paging
as false.}

\item{filter}{Filter parameters. No need to include \code{&filter=}.}
}
\value{
Web-encoded URL for DATIM API query.
}
\description{
Adds filter to DATIM API query and encodes for web.
}
