% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R
\name{api_get}
\alias{api_get}
\title{Execute and return a DATIM API query.}
\usage{
api_get(api_call)
}
\arguments{
\item{api_call}{Base DATIM API query, specifying API table and setting paging
as false.}
}
\value{
Result of DATIM API query returned as dataframe.
}
\description{
Gets and flattens DATIM API query as dataframe.
}
