% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R
\name{api_sql_call}
\alias{api_sql_call}
\title{Query DATIM SQL View.}
\usage{
api_sql_call(sqlView, var = NA)
}
\arguments{
\item{sqlView}{uid of sqlView table to query.}

\item{var}{Variable to substitute into SQL query. Only supply if SQL view is
of type query.}
}
\value{
Web-encoded URL for DATIM API query.
}
\description{
Queries a DATIM SQL View and returns data object.
}
