% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R, R/utilities.R
\name{api_version}
\alias{api_version}
\title{Returns a version of the DHIS2 API for the current version of DATIM}
\usage{
api_version()

api_version()
}
\value{
API version.

Version of the API
}
\description{
Returns a version of the DHIS2 API for the current version of DATIM

Returns version of the API
}
