% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageQueue.R
\name{appendMessage}
\alias{appendMessage}
\title{Title appendMessage}
\usage{
appendMessage(x, message, level, tool)
}
\arguments{
\item{x}{A message queue}

\item{message}{A string or vector of strings of messages.}

\item{level}{A string or vector of strings of}

\item{tool}{A string vector of the tool (optional) description
message levels (ERROR, WARNING, INFO)}
}
\value{
A MessageQueue class.
}
\description{
Generic function to handle appending messages to a MessageQueue
}
