% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{appendUnallocatedData}
\alias{appendUnallocatedData}
\title{Title appendUnallocatedData}
\usage{
appendUnallocatedData(d)
}
\arguments{
\item{d}{Datapackr d object}
}
\value{
Datapackr d object
}
\description{
Performs a test for any data which has not been allocated to a mechanism
in the PNSUxIM tab. If there is any unallocated data, it will be appended
to the already allocated data using the default mechanism.
}
