% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_excel_limits}
\alias{apply_excel_limits}
\title{Caps row and cell counts to excel limits}
\usage{
apply_excel_limits(x)
}
\arguments{
\item{x}{List}
}
\description{
Caps row and cell counts to excel limits
}
