% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cache_is_fresh}
\alias{cache_is_fresh}
\title{Checks whether a cached file is stale}
\usage{
cache_is_fresh(cache, max_age = NULL)
}
\arguments{
\item{cache}{Filepath to the cached file to check.}

\item{max_age}{The maximum age allowed for a cache to be considered fresh.
Follows syntax of \code{lubridate} package.}
}
\value{
A dataframe containing only the fresh portion of the cache file
}
\description{
Checks whether a cached file is stale
}
