% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{can_read_file}
\alias{can_read_file}
\title{Test for file read access}
\usage{
can_read_file(path)
}
\arguments{
\item{path}{Filepath to test.}
}
\value{
Logical. \code{TRUE} if file can be read, \code{FALSE} if not.
}
\description{
Given a file path, will test whether the file can be read.
}
