% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{can_spawn}
\alias{can_spawn}
\title{Can Spawn}
\usage{
can_spawn()
}
\value{
Boolean True or false
}
\description{
Determines whether processes can be run in parallel.
This is used in indicator and validation rule evaluation, but
should not be run on Windows currently
}
