% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{canReadFile}
\alias{canReadFile}
\title{Test for file read access}
\usage{
canReadFile(path = NA, type = "standard")
}
\arguments{
\item{path}{Filepath to test. Default is NA.}

\item{type}{File type, either \code{standard} or \code{template}.}
}
\value{
Logical. \code{TRUE} if file can be read, \code{FALSE} if not.
}
\description{
Given a file path, will test whether the file can be read.
}
