% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{changePDAPJobApprovalStatus}
\alias{changePDAPJobApprovalStatus}
\title{Title}
\usage{
changePDAPJobApprovalStatus(org_unit_id, period_id, approval_status)
}
\arguments{
\item{org_unit_id}{UID of the organisation unit}

\item{period_id}{ISO8601 formatted period ID}

\item{approval_status}{One of 'submitted', 'cancelled', 'approved', 'rejected'}
}
\value{
Returns the response from the API when changing the approval status
}
\description{
Title
}
