% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{check_params}
\alias{check_params}
\title{Standardized package function parameter definitions}
\usage{
check_params(
  country_uids,
  PSNUs,
  cop_year,
  tool,
  season,
  schema,
  datapack_name,
  template_path,
  wb,
  model_data,
  snuxim_model_data,
  output_folder,
  results_archive,
  ...,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{PSNUs}{Dataframe of PSNUs to use in this function, containing at least
\code{psnu_uid}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{season}{Either \code{COP} or \code{OPU}.}

\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{datapack_name}{Name you would like associated with this Data Pack.
(Example: "Western Hemisphere", or "Caribbean Region", or "Kenya".)}

\item{template_path}{Local filepath to Data Pack template Excel (XLSX) file.
This file MUST NOT have any data validation formats present. If left
\code{NULL}, will select the default based on \code{cop_year} and \code{tool}.}

\item{wb}{Openxlsx workbook object.}

\item{model_data}{Data from DATIM needed to pack into a COP Data Pack.}

\item{snuxim_model_data}{Export from DATIM needed to allocate data across
mechanisms in the PSNUxIM tab}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export. If left as \code{NULL}, will output to
\code{Working Directory}.}

\item{results_archive}{If TRUE, will export compiled results of all tests and
processes to output_folder.}

\item{d2_session}{DHIS2 Session id}
}
\value{
params List of valid parameters to use in functions.
}
\description{
Standardized package function parameter definitions
}
