% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBlankPrioritizations.R
\name{checkBlankPrioritizations}
\alias{checkBlankPrioritizations}
\title{checkInvalidPrioritizations(d, sheet)}
\usage{
checkBlankPrioritizations(d, sheet, quiet = T)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Checks if there are invalid and missing prioritizations.
}
