% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColStructure.R
\name{checkColStructure}
\alias{checkColStructure}
\title{Check Column Structure}
\usage{
checkColStructure(d, sheet)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{Specified sheet within wb.}
}
\value{
d
}
\description{
Checks structural integrity of columns on critical sheets for
   submitted Data Pack.
}
