% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkComments.R
\name{checkComments}
\alias{checkComments}
\title{Detect troublesome comment types}
\usage{
checkComments(d)
}
\arguments{
\item{d}{datapackr object}
}
\value{
Logical whether there exists a comments issue in specified Excel file
}
\description{
Searches an Excel file to detect whether there are any comments that cause
corruption when executing openxlsx::saveWorkbook
}
