% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDecimalValues.R
\name{checkDecimalValues}
\alias{checkDecimalValues}
\title{Checks decimal values in a specific Data Pack sheet}
\usage{
checkDecimalValues(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Checks data pulled from a single sheet in a Data Pack for decimal values
alerts where there are invalid decimal values.
}
