% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackingChecks.R
\name{checkExistsIndexCols}
\alias{checkExistsIndexCols}
\title{checkExistsIndexCols}
\usage{
checkExistsIndexCols(d, sheets = sheets)
}
\value{
A vector showing which sheets have all index header columns
}
\description{
All data pack sheets have certain columns (PSNU, Age, Sex, etc )
which are essentially index columns on that sheet. In order to determine
whether there are duplicate rows in the sheet, all of these columns
must be presen
}
