% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkExternalLinks.R
\name{checkExternalLinks}
\alias{checkExternalLinks}
\title{Detect the presense of any external links in a workbook.}
\usage{
checkExternalLinks(d)
}
\arguments{
\item{d}{datapackr object}
}
\value{
datapackr d object
}
\description{
Searches an Excel workbook for any external links.
}
