% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFormulas.R
\name{checkFormulas}
\alias{checkFormulas}
\title{Check formulas used in specified sheet in submitted Data Pack.}
\usage{
checkFormulas(d, sheet)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{Specified sheet within wb.}
}
\value{
d
}
\description{
Checks formulas to make sure they are up to date and have not
been tampered with.
}
