% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{checkHasPSNUxIM}
\alias{checkHasPSNUxIM}
\title{Check that a Data Pack tool has a PSNUxIM tab}
\usage{
checkHasPSNUxIM(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
d object list with additional flags for PSNUxIM state.
}
\description{
Internal function to determine if a PSNUxIM tab exists
}
