% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInvalidIndicatorCodes.R
\name{checkInvalidIndicatorCodes}
\alias{checkInvalidIndicatorCodes}
\title{Check for invalid Indicator Codes}
\usage{
checkInvalidIndicatorCodes(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Check for invalid Indicator Codes
}
