% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInvalidOrgUnits.R
\name{checkInvalidOrgUnits}
\alias{checkInvalidOrgUnits}
\title{checkInvalidOrgUnits(d)}
\usage{
checkInvalidOrgUnits(
  d,
  sheet,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{Data Pack sheet.}

\item{d2_session}{DHIS2 Session id}
}
\value{
d
}
\description{
Checks data pulled from a single sheet in a Data Pack and
alerts where there are unallowed org units based on current DATIM PSNU level.
}
