% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMechanisms.R
\name{checkMechanisms}
\alias{checkMechanisms}
\title{Check Mechanisms}
\usage{
checkMechanisms(
  d,
  cached_mechs_path = paste0(Sys.getenv("support_files_directory"), "mechs.rds"),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{cached_mechs_path}{Local file path to an RDS file containing
a cached copy of the mechanisms SQL view.}

\item{d2_session}{DHIS2 Session id}
}
\value{
d
}
\description{
Tests mechanisms found in the Data Pack PSNUxIM tab
against the Mechanism View from DATIM.
}
