% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNegativeValues.R
\name{checkNegativeValues}
\alias{checkNegativeValues}
\title{Check a Data Pack sheet for negative values}
\usage{
checkNegativeValues(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Checks for negative values in a specific Data Pack sheet
}
