% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{checkNonEqualTargets}
\alias{checkNonEqualTargets}
\title{checkNonEqualTargets extracts all rows from the PSNUxIM tab
where the value of the DataPackTarget (currently column G)
does not match the corresponding value in the main DataPackTabs.}
\usage{
checkNonEqualTargets(d, original_targets)
}
\arguments{
\item{d}{Datpackr d object}

\item{original_targets}{original_targets object}
}
\value{
Modified d object
}
\description{
checkNonEqualTargets extracts all rows from the PSNUxIM tab
where the value of the DataPackTarget (currently column G)
does not match the corresponding value in the main DataPackTabs.
}
