% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{checkNonNumericPSNUxIMValues}
\alias{checkNonNumericPSNUxIMValues}
\title{Title checkNonNumericPSNUxIMValues}
\usage{
checkNonNumericPSNUxIMValues(d, header_cols)
}
\arguments{
\item{d}{Datapackr d object}

\item{header_cols}{header_cols object}
}
\value{
Datapackr d object
}
\description{
Checks for non-numeric data values in the PSNUxIM tab.
}
