% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNotPEPFARSupportedPSNUs.R
\name{checkNotPEPFARSupportedPSNUs}
\alias{checkNotPEPFARSupportedPSNUs}
\title{Title}
\usage{
checkNotPEPFARSupportedPSNUs(d)
}
\arguments{
\item{d}{Datapackr object with generated analytics table}
}
\value{
d Datapackr object
}
\description{
Checks whether any MER data has been allocated
to PSNUs which are prioritized as "Not PEPFAR supported".
}
