% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNumericValues.R
\name{checkNumericValues}
\alias{checkNumericValues}
\title{Check Data Pack sheet for non-numeric values}
\usage{
checkNumericValues(d, sheet, header_cols = NULL)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{Specified sheet within wb.}

\item{header_cols}{Header columns to check}
}
\value{
d
}
\description{
Checks data pulled from a single sheet in a Data Pack and
alerts where there are non-numeric values instead of valid data.
}
