% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPSNUData.R
\name{checkPSNUData}
\alias{checkPSNUData}
\title{Check PSNUxIM tab Data}
\usage{
checkPSNUData(
  d,
  validation_rules_path = paste0(Sys.getenv("support_files_directory"),
    "cop_validation_rules.rds"),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id}
}
\value{
d
}
\description{
Tests PSNUxIM data
}
