% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDisaggs.R
\name{checkPSNUxIM_Disaggs}
\alias{checkPSNUxIM_Disaggs}
\title{Title checkPSNUxIM_Disaggs}
\usage{
checkPSNUxIM_Disaggs(d)
}
\arguments{
\item{d}{Datapackr object.}
}
\value{
d Datapackr object.
}
\description{
Checks disagg combinations (Age/Sex/Key Pop) in the PSNUxIM
tab for invalid combinations. Removes data which is invalid.
}
