% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSNUxIM.R
\name{checkPSNUxIMDisaggs}
\alias{checkPSNUxIMDisaggs}
\title{Title checkPSNUxIMDisaggs}
\usage{
checkPSNUxIMDisaggs(d)
}
\arguments{
\item{d}{Datapack d object}
}
\value{
d Datapack d object
}
\description{
Title checkPSNUxIMDisaggs
}
