% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSiteTool.R
\name{checkSiteToolColStructure}
\alias{checkSiteToolColStructure}
\title{checkSiteToolColStructure(d)}
\usage{
checkSiteToolColStructure(d)
}
\arguments{
\item{d}{datapackr list object containing at least d$data$extract,
d$data$sheet & d$info$warningMsg.}
}
\value{
A datapackr list object, \code{d}, storing a warning message of all
   issues related to Data Pack or Site Tool columns names or order.
}
\description{
Checks structural integrity of columns on critical sheets for
   submitted  Site Tool.
}
