% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStructure.R
\name{checkStructure}
\alias{checkStructure}
\title{Check tab structure of tool submitted for validation.}
\usage{
checkStructure(d)
}
\arguments{
\item{d}{Datapackr object.}
}
\value{
d
}
\description{
Checks structural integrity of tabs for submitted tool.
}
