% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSiteTool.R
\name{colorCodeSites}
\alias{colorCodeSites}
\title{Color code sites by type}
\usage{
colorCodeSites(wb, sheet, cols, rows)
}
\arguments{
\item{wb}{An Openxlsx Workbook object.}

\item{sheet}{A name or index of a worksheet.}

\item{cols}{Columns to apply conditional formatting to.}

\item{rows}{Rows to apply conditional formatting to.}
}
\description{
Adds conditional formatting layers over site names to color code then by type:
\itemize{
  \item{Green}{Community sites.}
  \item{Blue}{Facility sites.}
  \item{Purple}{National-level.}
  \item{Khaki}{Military node.}
}
}
