% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMER_SNUxIM.R
\name{combineMER_SNUxIM}
\alias{combineMER_SNUxIM}
\title{combineMER_SNUxIM(d)}
\usage{
combineMER_SNUxIM(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Takes the output of the \code{\link{unPackSNUxIM}} and
    \code{\link{unPackSheets}} functions and delicatety combines these to create
    a single dataframe at the PSNU x IM level.
}
