% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapackVsDatim.R
\name{compareData_DatapackVsDatim}
\alias{compareData_DatapackVsDatim}
\title{Compare Data Pack Targets to DATIM}
\usage{
compareData_DatapackVsDatim(
  d,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id}
}
\value{
list object of diff result $psnu_x_im_wo_dedup, $psnu_w_dedup,
$updates (import to bring DATIM up to date with datapack), $deletes
(import to bring DATIM up to date with datapack)
}
\description{
Compares the data in a parsed data pack that would be
destined for DATIM with target data in in DATIM.
}
