% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapackVsDatim.R
\name{compareData_OpuDatapackVsDatim}
\alias{compareData_OpuDatapackVsDatim}
\title{compareData_OpuDatapackVsDatim}
\usage{
compareData_OpuDatapackVsDatim(
  d,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  datim_data = NULL
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}

\item{datim_data}{A data frame resulting from datimutils::getDataValueSets. If null, the data will be fetched
from DATIM.}
}
\value{
list object of diff result $psnu_x_im_wo_dedup, $psnu_w_dedup,
$updates (import to bring DATIM up to date with datapack), $deletes
(import to bring DATIM up to date with datapack)
}
\description{
Legacy function maintained for backwards compatibility. If
processing an Datapack (either Datapack or a standalone PSNU tab)
as an OPU, only MER targets should be considered for processing.
}
