% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePacks.R
\name{comparePacks}
\alias{comparePacks}
\title{Compare Data Pack and Site Tool data}
\usage{
comparePacks(datapack_path, sitetool_path)
}
\arguments{
\item{datapack_path}{Local filepath leading to Data Pack to analyze.}

\item{sitetool_path}{Local filepath leading to Site Tool to analyze.}
}
\value{
Log of differences in target totals.
}
\description{
Takes a Data Pack and a Site Tool and unpacks both for easy comparison of
target totals.
}
