% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{configFile}
\alias{configFile}
\title{Mapping of necessary Data Pack metadata}
\format{A data frame with metadata for each Operating Unit's Data Pack:
\describe{
  \item{DataPack_name}{Usually the same as Operating Unit, except in rare
  cases for some PEPFAR Regional Operating Units.}
  \item{model_uid}{Usually the same as Operating Unit, except in rare cases
  for some PEPFAR Regional Operating Units.}
  \item{countryName}{Name of country as stored in DATIM organization
  hierarchy, even within Regional Operating Units, with the exception of
  _Military nodes in Regional OUs, which are necessarily listed separately
  as if a distinct country.}
  \item{countryUID}{Usually the organisationunituid as described in the DATIM
  organization unit hierarchy, uniquely identifying the Country or _Military
  node listed in \code{countryName}. For countries that are new to PEPFAR 
  reporting in FY20 and that were not logged in DATIM prior to the initial
  release of Data Packs, an arbitrary uid is assigned in this column for
  later re-mapping to DATIM.}
  \item{milPSNU}{For all countries and Regions with a _Military node, this
  column lists the name}
  \item{milPSNUuid}{For all countries and Regions with a _Military node, this
  column lists the DATIM organisationunituid.}
  \item{level3name}{Lists the \code{name} as used in level 3 of DATIM's
  organization unit hierarchy. For Regional OUs, this is the name of the
  Regional OU instead of the Country Name.}
  \item{uidlevel3}{Lists the \code{uid} as used in level 3 of DATIM's
  organization unit hierarchy. For Regional OUs, this is the uid of the
  Regional OU instead of the Country uid.}
  \item{level4name}{For countries in Regional OUs, lists the \code{name} as
  used in level 4 of DATIM's organization unit hierarchy. For Regional OUs,
  this is the name of the Country instead of the Regional OU.}
  \item{uidlevel4}{For countries in Regional OUs, lists the \code{uid} as
  used in level 4 of DATIM's organization unit hierarchy. For Regional OUs,
  this is the uid of the Country instead of the Regional OU.}
  \item{Currently_in_DATIM}{Documents whether the Country listed in 
  \code{countryName} was documented in DATIM's organization hierarchy as of
  Jan 8, 2019}
  \item{Military_nested_underneath}{Documents (Y/N) whether the country listed in
  \code{countryName} has a _Military node nested beneath it in the DATIM
  organization hierarchy.}
  \item{milPSNU_in_DATIM}{Documents (Y/N) whether nested or explicitly listed
  _Military node was present in DATIM as of Jan 8, 2019. Where this is a
  \code{N}, datapackr generates the _Military node for Data Pack use with the
  uid listed in \code{milPSNUuid}.}
  \item{Current_Prioritization_Level}{Lists the level in the DATIM
  organization hierarchy that contains the "Prioritization" level for each
  country.}
  \item{Prioritizing_at_Natl_or_SNU}{Documents level at which a country
  prioritizes for faster processing in datapackr. Options include "SNU" for 
  Sub-National Unit-level prioritization, "Natl" for National-level
  prioritization, and "Regional" for units that prioritize at Regional Level
  (for FY20 this is only _Military nodes in some Regional OUs).}
  \item{isRegion}{Logical (1/0) field documenting whether a country listed in
  \code{countryName} is within a Regional OU.}
  \item{isMil}{Logical (1/0) field documenting whether the unit listed in 
  \code{countryName} is a Military node.}
}}
\usage{
configFile
}
\description{
A data frame containing Data Pack metadata required for executing
datapackr processes.
}
\keyword{datasets}
