% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cop20_data_pack_schema}
\alias{cop20_data_pack_schema}
\title{Schema describing correct structure of the COP20 Data Pack template.}
\format{

}
\usage{
cop20_data_pack_schema
}
\description{
This schema describes the correct structure of a COP20 Data Pack
file.It is generated from the template used to produce Data Packs, and is
also useful in validating Data Packs passed through datapackr.
}
\keyword{datasets}
