% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cop20OPU_data_pack_schema}
\alias{cop20OPU_data_pack_schema}
\title{Schema describing correct the structure of the COP20 OPU Data Pack
 template.}
\format{

}
\usage{
cop20OPU_data_pack_schema
}
\description{
This schema describes the correct structure of a COP20 OPU
Data Pack file, generated from the template used to produce Data Packs and
 useful in validating Data Packs passed through datapackr.
}
\keyword{datasets}
