% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cop22_data_pack_schema}
\alias{cop22_data_pack_schema}
\title{Schema describing the correct structure of teh COP22 Data Pack
 template.}
\format{

}
\usage{
cop22_data_pack_schema
}
\description{
This schema describes the correct structure of a COP22 Data Pack
file, generated from the template used to produce Data Packs and useful in
validating Data Packs passed through datapackr.
}
\keyword{datasets}
