% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cop_validation_rules}
\alias{cop_validation_rules}
\title{COP Validation Rules}
\format{
The following COP years are included in this dataset:
\describe{
\item{2021}{A list object containing the validation rules for COP21/FY22.}
\item{2022}{A list object containing the validation rules for COP22/FY23.}
}
}
\source{
\url{https://www.datim.org/}
}
\usage{
cop_validation_rules
}
\description{
A nested list of validation rules for both current and past
COP years.
}
\section{Structure for COP21 data set}{

The data set for 2021 conforms to the following structure:
\itemize{
\item \code{description}: A description of the DATIM validation rule, showing the
relationship required between two indicators. Synonymous to
the \code{name} and \code{instruction} columns.
\item \code{id}: The DATIM UID for the rule.
\item \code{importance}: Category showing the relative importance of the validaiton
rule. For COP20 and COP21, this is always listed as \code{MEDIUM}.
\item \code{instruction}: A description of the DATIM validation rule, showing the
relationship required between two indicators. Synonymous to the \code{description}
and \code{name} columns.
\item \code{name}: A description of the DATIM validation rule, showing the
relationship required between two indicators. Synonymous to the \code{description}
and \code{instruction} columns.
\item \code{operator}: The operator used in the validation rule. This must be either
\code{<=}, \code{>=}, or \code{|}.
\item \code{periodType}: A string indicating whether the indicator is reported
quartery or annually. The value is either \code{Quarterly} or \code{FinancialOct}.
\item \code{ruletype}: The type of rule being applied. This value is
always \code{VALIDATION}.
\item \code{leftSide.dataElements}: A nested list containing a single DATIM data
element defining the indicator on the left-hand side of the equation.
\item \code{leftSide.description}: A description of the indicator on the left-hand
side of the validation rule equation.
\item \code{leftSide.expression}: An expression defining how to calculate the value
of the left-hand side of the validation rule equation.
\item \code{leftSide.missingValueStrategy}: A string that states whether this rule
should be skipped if the value of the left-hand side of the equation is
missing. Value is either \code{NEVER_SKIP} or \code{SKIP_IF_ALL_VALUES_MISSING}.
\item \code{rightSide.dataElements}: A nested list containing a single DATIM data
element defining the indicator on the right-hand side of the equation.
\item \code{rightSide.description}: A description of the indicator on the right-hand
side of the validation rule equation.
\item \code{rightSide.expression}: An expression defining how to calculate the value
of the right-hand side of the validation rule equation.
\item \code{rightSide.missingValueStrategy}: A string that states whether this rule
should be skipped if the value of the right-hand side of the equation is
missing. Value is either \code{NEVER_SKIP} or \code{SKIP_IF_ALL_VALUES_MISSING}.
\item \code{rightSide.ops}:
\item \code{leftSide.ops}:
}
}

\section{Structure for COP22 data set}{

The data set for COP22 conforms to the following structure:
\itemize{
\item \code{name}: A descriptive name of the DATIM validation rule, showing the
relationship required between two indicators. Synonymous to
the \code{description} column.
\item \code{id}: The DATIM UID for the rule.
\item \code{periodType}: A string indicating whether the indicator is reported
quartery or annually. The value is either \code{Quarterly} or \code{FinancialOct}.
\item \code{description}: A description of the DATIM validation rule, showing the
relationship required between two indicators. Synonymous to
the \code{name} column.
\item \code{operator}: The operator used in the validation rule. This must be either
\code{<=}, \code{>=}, or \code{|}.
\item \code{leftSide.expression}: An expression defining how to calculate the value
of the left-hand side of the validation rule equation.
\item \code{leftSide.missingValueStrategy}:  A string that states whether this rule
should be skipped if the value of the left-hand side of the equation is
missing. Value is either \code{NEVER_SKIP} or \code{SKIP_IF_ALL_VALUES_MISSING}.
\item \code{rightSide:expression}: An expression defining how to calculate the value
of the right-hand side of the validation rule equation.
\item \code{rightSide.missingValueStrategy}:  A string that states whether this rule
should be skipped if the value of the right-hand side of the equation is
missing. Value is either \code{NEVER_SKIP} or \code{SKIP_IF_ALL_VALUES_MISSING}.
\item \code{rightSide.ops}:
\item \code{leftSide.ops}:
}
}

\keyword{datasets}
