% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{countryUIDs_homeCell}
\alias{countryUIDs_homeCell}
\title{Location of Country UIDs on Home tab.}
\usage{
countryUIDs_homeCell()
}
\value{
Cell reference where Country UIDs should be on Home tab.
}
\description{
Location of Country UIDs on Home tab.
}
