% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_play_spectrum_output.R
\name{create_play_spectrum_output}
\alias{create_play_spectrum_output}
\title{Create fake Spectrum output for testing}
\usage{
create_play_spectrum_output(
  country_uids,
  cop_year,
  output_folder = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
Fake Spectrum dataset
}
\description{
Creates a fake Spectrum Data Pack export for use in testing with Data Packs.
Requires login to DATIM.
}
