% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDATAExport.R
\name{createDATAExport}
\alias{createDATAExport}
\title{createDATAExport}
\usage{
createDATAExport(d, export_type = NULL)
}
\arguments{
\item{export_type}{paw or datim are passed as export types}
}
\value{
A data frame in standardized DHIS2 export format for DATIM or PAW
}
\description{
createDATAExport
}
