% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDATIMExport.R
\name{createDATIMExport}
\alias{createDATIMExport}
\title{Title}
\usage{
createDATIMExport(d)
}
\arguments{
\item{d}{}
}
\value{
A data frame in standardized DHIS2 export format
}
\description{
Title
}
