% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKeychainInfo.R
\name{createKeychainInfo}
\alias{createKeychainInfo}
\title{Create Keychain Info for use in datapackr sidecar.}
\usage{
createKeychainInfo(
  submission_path = NULL,
  tool = NULL,
  country_uids = NULL,
  cop_year = NULL,
  d2_session = NULL
)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{d2_session}{DHIS2 Session id}
}
\description{
Creates Keychain info needed for use across most datapackr unPack functions.
}
