% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPAWExport.R
\name{createPAWExport}
\alias{createPAWExport}
\title{createPAWExport}
\usage{
createPAWExport(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
A data frame in standardized DHIS2 export format for PAW
}
\description{
createPAWExport
}
