% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestDataset.holistic.R
\name{createTestDataset.holistic}
\alias{createTestDataset.holistic}
\title{Create a multi-country, holistic test Targets dataset}
\usage{
createTestDataset.holistic(
  cop_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{COP year for dating as well as selection of
templates.}

\item{d2_session}{datimutils d2Session object}
}
\value{
wb with specified sheet packed with data
}
\description{
Compiles a test DATIM import file that tests all major import DATIM assumptions.
}
