% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestDataset.R
\name{createTestDataset}
\alias{createTestDataset}
\title{Create a test Targets dataset}
\usage{
createTestDataset(
  country_uids,
  cop_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{UIDs for Countries to tailor data for.}

\item{cop_year}{COP year for dating as well as selection of
templates.}

\item{d2_session}{datimutils d2Session object}
}
\value{
wb with specified sheet packed with data
}
\description{
Compiles a test DATIM import file as would be output by a Data Pack.
}
