% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createWorkbook}
\alias{createWorkbook}
\title{Create a clean Data Pack Workbook shell}
\usage{
createWorkbook(
  datapack_name = NULL,
  country_uids,
  template_path = NULL,
  cop_year = getCurrentCOPYear(),
  tool = "Data Pack",
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{datapack_name}{Name you would like associated with this Data Pack.
(Example: "Western Hemisphere", or "Caribbean Region", or "Kenya".)}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{template_path}{Local filepath to Data Pack template Excel (XLSX) file.
This file MUST NOT have any data validation formats present. If left
\code{NULL}, will select the default based on \code{cop_year} and \code{tool}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Either "Data Pack" or "OPU Data Pack"? Default is "Data Pack".}

\item{d2_session}{DHIS2 Session id}
}
\value{
wb Data Pack Workbook shell
}
\description{
Create a clean Data Pack Workbook shell
}
