% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{dataPackName_homeCell}
\alias{dataPackName_homeCell}
\title{Location of Name of the DataPack on the Home tab}
\usage{
dataPackName_homeCell()
}
\value{
Cell reference where the name of the datapack can be found.
}
\description{
Location of Name of the DataPack on the Home tab
}
