% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapackr.R
\docType{package}
\name{datapackr}
\alias{datapackr}
\title{Data Pack on R}
\description{
datapackr: a grammar of Data Pack manipulation


datapackr is the engine behind PEPFAR Data Packs, a tool used annually to
aid PEPFAR Country Teams in setting targets for their annual Country
Operating Plan (COP).

It has three main goals:

\itemize{
\item Generate Data Packs, either at District or Site Level as directed.
\item Process and validate Data Packs submitted for review
\item Prepare Data Pack data for import into DATIM, the PEPFAR data collection database.
}
}
