% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{datapackrSupports}
\alias{datapackrSupports}
\title{Supported Tools & COP Years}
\usage{
datapackrSupports()
}
\value{
Tibble of supported tools mapped to their supported COP Years
}
\description{
Supported Tools & COP Years
}
