% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{default_catOptCombo}
\alias{default_catOptCombo}
\title{Returns `default` categoryOptionCombo uid.}
\usage{
default_catOptCombo()
}
\value{
`Default` categoryOptionCombo uid.
}
\description{
Returns `default` categoryOptionCombo uid.
}
