% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_approval_memo.R
\name{default_memo_font}
\alias{default_memo_font}
\title{Default Memo Font}
\usage{
default_memo_font()
}
\value{
String with default font for memo target tables
}
\description{
Default Memo Font
}
