% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_approval_memo.R
\name{default_memo_style_header}
\alias{default_memo_style_header}
\title{Default Memo Header Style}
\usage{
default_memo_style_header()
}
\value{
Officer list object defining default style for memo table headers
}
\description{
Default Memo Header Style
}
