% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_approval_memo.R
\name{default_memo_style_para}
\alias{default_memo_style_para}
\title{Default Memo Paragraph Style}
\usage{
default_memo_style_para()
}
\value{
Officer list object defining default style for memo table bodies
}
\description{
Default Memo Paragraph Style
}
