% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemo.R
\name{defaultMemoStylePara}
\alias{defaultMemoStylePara}
\title{Default Memo Paragraph Style}
\usage{
defaultMemoStylePara()
}
\value{
Officer list object defining default style for memo table bodies
}
\description{
Default Memo Paragraph Style
}
