% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunctDisaggs.R
\name{defunctDisaggs}
\alias{defunctDisaggs}
\title{Check for Defunct Disaggregates}
\usage{
defunctDisaggs(d, sheet)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{Specified sheet within wb.}
}
\value{
d
}
\description{
Checks data extracted from a sheet in a submitted Data Pack
   identify cases where invalid Disaggregate combinations have
   been used.
}
