% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{deleteExistingPDAPJobs}
\alias{deleteExistingPDAPJobs}
\title{Title}
\usage{
deleteExistingPDAPJobs(org_unit_id, period_id, job_type)
}
\arguments{
\item{org_unit_id}{UID of the organization unit}

\item{period_id}{ISO8601 formatted period ID}

\item{job_type}{Type of job to delete, i.e. 'target_setting_tool'}
}
\value{
TRUE if all jobs are deleted, otherwise, FALSE.
}
\description{
Title
}
