% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workingWithCategories.R
\name{DEs_COCs_COs_Cs}
\alias{DEs_COCs_COs_Cs}
\title{Map code list to COs}
\usage{
DEs_COCs_COs_Cs()
}
\value{
fullCodeList
}
\description{
Combine DEs and COCs from code list to COs, broken out by group.
}
