% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackOPU_PSNUxIM.R
\name{.testInvalidIndicatorCodes}
\alias{.testInvalidIndicatorCodes}
\title{Test for invalid Indicator Codes}
\usage{
.testInvalidIndicatorCodes(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Tests for invalid Indicator Codes
}
