% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{dropDuplicatedPSNUxIMColumns}
\alias{dropDuplicatedPSNUxIMColumns}
\title{Title dropDuplicatedPSNUxIMColumns}
\usage{
dropDuplicatedPSNUxIMColumns(d)
}
\arguments{
\item{d}{Datapackr d object}
}
\value{
Modified d object
}
\description{
Identify and drop any duplicated user specified mechanism
columns in the PSNUxIM tab.
}
