% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{dropInvalidMechColumns}
\alias{dropInvalidMechColumns}
\title{Title dropInvalidMechColumns}
\usage{
dropInvalidMechColumns(d, cols_to_keep, sheet = "PSNUxIM")
}
\arguments{
\item{d}{Datpackr d object}

\item{cols_to_keep}{Object of columns to keep}

\item{sheet}{Name of the sheet}
}
\value{
d
}
\description{
Removes any invalid, user-specified mechanism
columns in the PSNUxIM tab.
}
