% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptySiteToolSheetFrame.R, R/parseSiteTool.R
\name{emptySiteToolSheetFrame}
\alias{emptySiteToolSheetFrame}
\title{emptySiteToolSheetFrame(d)}
\usage{
emptySiteToolSheetFrame()

emptySiteToolSheetFrame()
}
\value{
An empty intermediate data frame when all values in a given Site
Tool sheet are completely empty.

An empty intermediate data frame when all values in a given Site
Tool sheet are completely empty.
}
\description{
emptySiteToolSheetFrame(d)

emptySiteToolSheetFrame(d)
}
