% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_indicators.R
\name{evaluate_indicators}
\alias{evaluate_indicators}
\title{Evaluate Indicators}
\usage{
evaluate_indicators(combis, values, inds)
}
\arguments{
\item{combis}{A vector of data element and category option combo UIDs
of the form #{dataelement_id.categoryoptioncombo_id}}

\item{values}{A vector of values}

\item{inds}{A dataframe consisting of indicator UIDs, name,
numerator, and denominator expression.}
}
\value{

}
\description{
Evaluate Indicators
}
