% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPackr.R
\name{exportPackr}
\alias{exportPackr}
\title{Export output.}
\usage{
exportPackr(data, output_folder, tool, datapack_name)
}
\arguments{
\item{data}{Data object to export. Can be either a dataframe or an Openxlsx
Workbook object.}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export. If left as \code{NULL}, will output to
\code{Working Directory}.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{datapack_name}{Name you would like associated with this Data Pack.
(Example: "Western Hemisphere", or "Caribbean Region", or "Kenya".)}
}
\description{
Exports a datapackr output to specified filepath.
}
