% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_uid}
\alias{extract_uid}
\alias{extract_uid_all}
\title{Extract UID.}
\usage{
extract_uid(string, bracketed = TRUE)

extract_uid_all(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible
to one.}
}
\value{
Character vector of DHIS2 11-digit UIDs found in string.
}
\description{
Extracts a DHIS2 11-digit UID from provided string.
}
