% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSNUxIM.R
\name{extractDuplicateRows}
\alias{extractDuplicateRows}
\title{Title}
\usage{
extractDuplicateRows(d, sheet = "PSNUxIM")
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{String. Name of sheet/tab within tool.}
}
\value{
Modified d object with a list of duplicated rows
}
\description{
Title
}
