% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{extractOriginalTargets}
\alias{extractOriginalTargets}
\title{Title}
\usage{
extractOriginalTargets(d, cols_to_keep, header_cols, sheet = "PSNUxIM")
}
\arguments{
\item{d}{Datapackr d object}

\item{cols_to_keep}{cols_to_keep object}

\item{header_cols}{headers_cols object}

\item{sheet}{Name of the sheet (PSNUxIM)}
}
\value{
original_targets object as a data frame.
}
\description{
Title
}
