% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRawColumnData.R
\name{extractRawColumnData}
\alias{extractRawColumnData}
\title{Title extractRawColumnData}
\usage{
extractRawColumnData(d, sheet, cols)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{String. Name of sheet/tab within tool.}

\item{cols}{Name of the columns to extract. PSNU is mandatory}
}
\value{
A data frame consisting of psnu, psnu_uid, age, sex, key_population
indicator_code and value
}
\description{
In certain situations in the checkAnalytics function, we need to extract raw
data from the DataPack sheets for use in calculations. This
utility function will help to extract raw data from the excel sheets
given a vector of columns to extract.
}
