% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSNUxIM.R
\name{extractSNUxIMCombos}
\alias{extractSNUxIMCombos}
\title{Title}
\usage{
extractSNUxIMCombos(d, p = NULL)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
Modified d object with documented and missing SNUxIM combos
}
\description{
Title
}
