% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_data_flattenDataPackModel_19}
\alias{fake_data_flattenDataPackModel_19}
\title{fake_data_flattenDataPackModel_19}
\format{
An object of class \code{list} of length 1.
}
\usage{
fake_data_flattenDataPackModel_19
}
\description{
Stripped down anonymized version of COP 19 model data file list object 
for unit testing and examples for function flattenDataPackModel_19
}
\keyword{datasets}
