% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_prioritization_table.R
\name{fetch_prioritization_table}
\alias{fetch_prioritization_table}
\title{Fetch Prioritization Table}
\usage{
fetch_prioritization_table(
  psnus,
  cop_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{psnus}{Dataframe of PSNUs to use in this function, containing at least
\code{psnu_uid}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{d2_session}{DHIS2 Session id}
}
\value{
A data frame consisting of Organisation unit, Value, prioritization
}
\description{
Fetch Prioritization Table
}
