% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{fetchMechsViewFromAPI}
\alias{fetchMechsViewFromAPI}
\title{fetchMechsViewFromAPI}
\usage{
fetchMechsViewFromAPI(
  operating_units = NULL,
  cop_year = NULL,
  uids = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{operating_units}{OUs to return data for. If not supplied, returns for all.}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{uids}{Character vector of uids to filter list by.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Mechanism List
}
\description{
fetchMechsViewFromAPI
}
