% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPrioritizationTable.R
\name{fetchPrioritizationTable}
\alias{fetchPrioritizationTable}
\title{Title}
\usage{
fetchPrioritizationTable(d, d2_session, include_no_prio = TRUE)
}
\arguments{
\item{d}{Datapackr d object}

\item{d2_session}{Datimutils session}

\item{include_no_prio}{If set to true, include No prioritization data}
}
\value{
A modified d object with d$data$memo$datim$prio
}
\description{
Utility function which retrieves prioritization table 
data from the DATIM analytics API.
}
