% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenDataPackModel.R
\name{flattenDataPackModel_19}
\alias{flattenDataPackModel_19}
\title{flattenDataPackModel_19}
\usage{
flattenDataPackModel_19(data, country_uids = NULL)
}
\arguments{
\item{data}{list object with COP 19 model data}

\item{country_uids}{string vector listing country uids at the top level of the list hierarchy.}
}
\value{
list object with one result set for each country named by country uid
}
\description{
Takes a COP 19 Data Pack Model data file and removes non-results content and flattens
the results hierarchy
}
\examples{
        
        str(flattenDataPackModel_19_fake_data)
        flattenDataPackModel_19(flattenDataPackModel_19_fake_data)

}
