% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flattenDataPackModel_19_fake_data}
\alias{flattenDataPackModel_19_fake_data}
\title{flattenDataPackModel_19_fake_data}
\format{An object of class \code{list} of length 1.}
\usage{
flattenDataPackModel_19_fake_data
}
\description{
Stripped down anonymized version of COP 19 model data file list object 
for unit testing and examples flattenDataPackModel_19_fake_data
}
\keyword{datasets}
