% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{formatSetStrings}
\alias{formatSetStrings}
\title{Title}
\usage{
formatSetStrings(vec)
}
\arguments{
\item{vec}{A vector of numbers}
}
\value{

}
\description{
Format a vector of numbers into a string of ranges
}
\note{
Lifted from https://stackoverflow.com/questions/16800803/
}
\examples{
formatSetStrings(c(1,2,3,5,6,7,8))
formatSetStrings(c(8,7,6,5,3,2,1))

}
