% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packFrame.R
\name{frameDataSheet}
\alias{frameDataSheet}
\title{writeDataSheet(wb, type = "Data Pack")}
\usage{
frameDataSheet(wb, sheet, type = "Data Pack")
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{sheet}{Name of the sheet}

\item{type}{Defaults to "Data Pack".}
}
\value{
Openxlsx workbook object with added, styled Home tab.
}
\description{
Function to write Data sheet structures
}
