% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMechanisms.R
\name{frameMechMap}
\alias{frameMechMap}
\title{Frame a mechanism mapping tool.}
\usage{
frameMechMap(datapack_uid)
}
\arguments{
\item{datapack_uid}{A unique ID specifying the PEPFAR Operating Unit 
the Site Tool belongs to.}
}
\value{
OpenXLSX workbook object.
}
\description{
Creates shell of a mechanism mapping tool from scratch and prepares it for
addition of other features and data.
}
