% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packFrame.R
\name{frameSiteTool}
\alias{frameSiteTool}
\title{Create a Site Tool from scratch and prep it for data.}
\usage{
frameSiteTool(datapack_uid)
}
\arguments{
\item{datapack_uid}{A unique ID specifying the PEPFAR Operating Unit 
the Site Tool belongs to.}
}
\value{
OpenXLSX workbook object for use in data writing functions.
}
\description{
Creates a blank Excel workbook, then packs it with non-data Site Tool
elements and features.
}
