% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fresh_cache_part}
\alias{fresh_cache_part}
\title{Return the fresh portion of a cache file}
\usage{
fresh_cache_part(cache, max_age)
}
\arguments{
\item{cache}{The cached file}

\item{max_age}{The maximum age allowed fora cache to be considered fresh}
}
\value{
A dataframe containing only the fresh portion of the cache file
}
\description{
Return the fresh portion of a cache file
}
