% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullCodeList}
\alias{fullCodeList}
\title{Full Code List specific to current COP targets dataset}
\format{
\describe{
  \item{dataelement}{}
  \item{dataelementuid}{}
  \item{categoryoptioncombo}{}
  \item{categoryoptioncombouid}{}
}
}
\usage{
fullCodeList
}
\description{
Full Code List specific to current COP targets dataset,
generated from datim.org/api/30/sqlViews/DotdxKrNZxG.
}
\keyword{datasets}
