% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_approval_memo.R
\name{generate_memo_template}
\alias{generate_memo_template}
\title{Generate Memo Template File}
\usage{
generate_memo_template(draft_memo = TRUE)
}
\arguments{
\item{draft_memo}{Boolean indicating whether the memo being written is a
draft or final memo.}
}
\value{

}
\description{
Generate Memo Template File
}
