% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateComparisonTable.R
\name{generateComparisonTable}
\alias{generateComparisonTable}
\title{Generate Comparison Table}
\usage{
generateComparisonTable(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
A comparison table
}
\description{
Generates a PSNUxIM level comparison between
the values in the (OPU) DataPack and what is currently
present in DATIM.
}
