% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{generatePSNUxIMComparison}
\alias{generatePSNUxIMComparison}
\title{Title generatePSNUxIMComparison}
\usage{
generatePSNUxIMComparison(d, original_targets)
}
\arguments{
\item{d}{Datapackr d object}

\item{original_targets}{original_targets object}
}
\value{
Modified d object with a new object d$info$psnuxim_comparison
}
\description{
Classifies discrepancies between the aggregated targets in the PSNUxIM
tab with those in the main sheets (original_targets).
}
