% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{getColumnsToKeep}
\alias{getColumnsToKeep}
\title{getColumnsToKeep(d)}
\usage{
getColumnsToKeep(d, sheet = "PSNUxIM")
}
\arguments{
\item{d}{Datapackr d object.}

\item{sheet}{Name of the sheet (PSNUxIM)}
}
\value{
A list of columns to keep.
}
\description{
Return a vector of columsn names to keep for this tool.
}
