% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{getCurrentCOPYear}
\alias{getCurrentCOPYear}
\title{Returns current COP Year}
\usage{
getCurrentCOPYear()
}
\value{
Current COP Year. (e.g., for COP19, returns 2019)
}
\description{
Returns current COP Year
}
