% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUs.R
\name{getDataPackOrgUnits}
\alias{getDataPackOrgUnits}
\title{Pull list of Org Units approved for use in Data Packs from DATIM based
on provided country UIDs.}
\usage{
getDataPackOrgUnits(
  include_mil = TRUE,
  include_DREAMS = TRUE,
  include_TSNUS = TRUE,
  additional_fields = NULL,
  use_cache = TRUE,
  cop_year = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{include_mil}{Logical. If \code{TRUE}, will also include _Military nodes
related to \code{country_uids}. Default is \code{TRUE}.}

\item{include_DREAMS}{Logical. If \code{TRUE}, will also include DREAMS
organisation units.}

\item{include_TSNUS}{Logical. If \code{TRUE}, will also include TSNUS
organisation units.}

\item{additional_fields}{Character string of any fields to return from DATIM
API other than those returned by default.}

\item{use_cache}{If \code{TRUE}, will first check to see if a cached export
of Data Pack Org Units already exists. If so, will use this cached version to
save time.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
Tibble of Data Pack Org Units.
}
\description{
Queries DATIM to extract tibble of Org Units approved for use in Data Pack
target setting.
}
