% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDataPackSchema}
\alias{getDataPackSchema}
\title{getDataPackSchema}
\usage{
getDataPackSchema(cop_year)
}
\arguments{
\item{cop_year}{cop year to pull get schema for}
}
\value{
{cop20, cop21}_data_pack_schema
}
\description{
getDataPackSchema
}
