% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataValue.R
\name{GetDataValueSet}
\alias{GetDataValueSet}
\title{GetDataValueSet}
\usage{
GetDataValueSet(keys, values, base_url = getOption("baseurl"),
  api_version = "30")
}
\arguments{
\item{keys}{character vector - data value set parameter keys (e.g. "dataSet", "period")}

\item{values}{character vector - values marching the key from keys (e.g. "Abcde123456", "2019Q1"}
}
\value{
tibble with the data requested
}
\description{
Used to read DHIS 2 data using the data value set endpoint
}
