% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataValueSets.R
\name{getDataValueSets}
\alias{getDataValueSets}
\title{getDataValueSets}
\usage{
getDataValueSets(
  keys,
  values,
  api_version = datapackr::api_version(),
  max_attempts = 3,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{keys}{character vector - data value set parameter keys (e.g. "dataSet", "period")}

\item{values}{character vector - values marching the key from keys (e.g. "Abcde123456", "2019Q1"}

\item{api_version}{- api version to use when calling DHIS2}

\item{max_attempts}{- number of times to try for a valid response}

\item{base_url}{string - base address of instance (text before api/ in URL)}
}
\value{
tibble with the data requested
}
\description{
Used to read DHIS 2 data using the data value set endpoint
}
