% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{getExistingFileS3Location}
\alias{getExistingFileS3Location}
\title{Title}
\usage{
getExistingFileS3Location(job_type, org_unit_id, period_id)
}
\arguments{
\item{job_type}{Type of job to get, i.e. 'target_setting_tool'}

\item{org_unit_id}{UID of the organization unit}

\item{period_id}{ISO8601 formatted period ID}
}
\value{
Returns the S3 file location of the existing job, otherwise NULL
}
\description{
Given a job type, organization unit ID, and period ID, this function will
return the S3 file location of the existing job. If no job is found, NULL
will be returned. If more than one job is found, a warning will be issued
and the first job will be used. If an error occurs, NULL will be returned.
}
