% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{getExistingPDAPJobs}
\alias{getExistingPDAPJobs}
\title{Title}
\usage{
getExistingPDAPJobs(org_unit_id, period_id, job_type)
}
\arguments{
\item{org_unit_id}{UID of the organization unit}

\item{period_id}{ISO8601 formatted period ID}

\item{job_type}{Type of job to get, i.e. 'target_setting_tool'}
}
\value{
Returns the response from the API. Should be a list of jobs if
successful, otherwise, an empty list. If an error occurs (e.g. 5xx), the
response will also be returned.
}
\description{
Title
}
