% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalytics.R
\name{getFY22Prioritizations}
\alias{getFY22Prioritizations}
\title{Get FY22 Prioritizations}
\usage{
getFY22Prioritizations(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
Dataframe of psnu_uid and value
}
\description{
Get FY22 Prioritizations
}
