% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullHTSModality.R
\name{getHTSModality}
\alias{getHTSModality}
\title{getHTSModality}
\usage{
getHTSModality(cop_year = getCurrentCOPYear(), dataElements = NULL)
}
\arguments{
\item{cop_year}{Specifies COP year for selection of correct modalities.}

\item{dataElements}{List of dataElements to filter against. (Optional)}
}
\value{
Dataframe of HTS modalities mapped to dataElements
}
\description{
Map HTS modality to dataElement id
}
