% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getIMPATTLevels}
\alias{getIMPATTLevels}
\title{Pull IMPATT levels from DATIM for all PEPFAR countries}
\usage{
getIMPATTLevels()
}
\value{
Dataframe of country metadata, including prioritization, planning,
country, community, and facility levels in DATIM organization hierarchy.
}
\description{
Queries DATIM to retrieve the latest version of
\code{/api/dataStore/dataSetAssignments/ous}
}
