% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMaxCores}
\alias{getMaxCores}
\title{Title}
\usage{
getMaxCores()
}
\value{
An integer number of cores to use in parallel processing
}
\description{
Determine the number of cores to be used for parallel processing
operations using the environment variable MAX_CORES. If not specified
the total number of cores will be used.
}
