% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{getMechanismView}
\alias{getMechanismView}
\title{getMechanismView(d2_session, support_files_path)}
\usage{
getMechanismView(
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  cached_mechs_path = paste0(Sys.getenv("support_files_directory"), "mechs.rds")
)
}
\arguments{
\item{d2_session}{datimutils d2Session object}

\item{cached_mechs_path}{Local file path to the cached mechanisms file.}
}
\value{
Mechs
}
\description{
Retrieves a view of mechanisms with partners and agencies
The function will attempt to read from a cached file, if defined and accessible.
Otherwise, if the user is logged in, the view
will be obtained from DATIM. Otherwise, an empty dataframe is returned.
}
