% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{getMechanismViewFromDATIM}
\alias{getMechanismViewFromDATIM}
\title{getMechanismViewFromDATIM}
\usage{
getMechanismViewFromDATIM(
  cop_year = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Mechanism List
}
\description{
getMechanismViewFromDATIM
}
