% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMechList.R
\name{getMechList}
\alias{getMechList}
\title{Compile Mechanism List for Site Tool}
\usage{
getMechList(country_uids, FY = NA)
}
\arguments{
\item{country_uids}{Character vector of DATIM country IDs. This can only
include countries. Regional Operating Unit uids will not be accepted}

\item{FY}{Numeric value of Fiscal Year to filter mechanism list by. If a FY
is not supplied, returns entire mechanism list.}
}
\description{
Extracts full list of valid mechanisms from DATIM for Fiscal Year & countries
specified and compiles these into the list as needed for use in the PEPFAR
Site Tool.
}
