% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMilitaryNodes}
\alias{getMilitaryNodes}
\title{Pull all _Military nodes from DATIM for all PEPFAR countries}
\usage{
getMilitaryNodes()
}
\value{
Dataframe of _Military names and ids, with associated Operating Units
and Countries.
}
\description{
Queries DATIM (\code{api/organisationUnits}) to retrieve the latest list of
_Military nodes for each PEPFAR country.
}
