% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getOperatingUnitFromCountryUIDs}
\alias{getOperatingUnitFromCountryUIDs}
\title{Get Operating Unit from Country UIDs}
\usage{
getOperatingUnitFromCountryUIDs(country_uids)
}
\arguments{
\item{country_uids}{List of country UIDs from the \code{d$info$country_uids} object.}
}
\value{
d
}
\description{
Takes in a set of Country UIDs and returns an Operating Unit name.
}
