% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPUDataFromDATIM.R
\name{getOPUDataFromDATIM}
\alias{getOPUDataFromDATIM}
\title{Pull data for packing OPU Data Pack}
\usage{
getOPUDataFromDATIM(
  cop_year,
  country_uids = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\value{
Returns dataset needed for use in \code{\link{packOPUDataPack}}.
}
\description{
Pulls data from DATIM API needed to pack OPU Data Pack.
}
