% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getOUFromCountryUIDs}
\alias{getOUFromCountryUIDs}
\title{Get Operating Unit from Country UIDs}
\usage{
getOUFromCountryUIDs(
  country_uids,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{Unique IDs for countries to include in the Data Pack.}

\item{d2_session}{DHIS2 Session id}
}
\value{
d
}
\description{
Takes in a set of Country UIDs and returns an Operating Unit name.
}
