% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{getPDAPJobsAPIURL}
\alias{getPDAPJobsAPIURL}
\title{Title}
\usage{
getPDAPJobsAPIURL(job = "PDAPAPIDomainName")
}
\arguments{
\item{job}{Name of the job, i.e. PDAPAPIDomainName}
}
\value{
Returns the PDAP API URL
}
\description{
Utility function to get the PDAP API URL from the AWS SSM Parameter Store.
}
