% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{getPresignedURL}
\alias{getPresignedURL}
\title{Upload DATIM Export to PDAP
Title}
\usage{
getPresignedURL(job, endpoint, query, service)
}
\arguments{
\item{job}{Name of the job, i.e. PDAPAPIDomainName}

\item{endpoint}{Path of the endpoint, i.e. /jobs/presignedurl}

\item{query}{List of query parameters}

\item{service}{Name of the service, i.e. execute-api}
}
\value{
Returns the presigned URL list (file_key, presigned_url, expiration time)
}
\description{
Upload DATIM Export to PDAP
Title
}
