% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrioritizations.R
\name{getPrioritizations}
\alias{getPrioritizations}
\title{Pull Prioritizations from DATIM}
\usage{
getPrioritizations(
  country_uids = NULL,
  cop_year = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{List of 11 digit alphanumeric DATIM codes representing
countries. If not provided, will check file for these codes. If not in file,
will flag error.}

\item{d2_session}{DHIS2 Session id}
}
\value{
prioritizations
}
\description{
Pull Prioritizations from DATIM for a given set of OUs and periods
}
