% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_import_file.R
\name{getPSNUInfo}
\alias{getPSNUInfo}
\title{getPSNUInfo}
\usage{
getPSNUInfo(
  snu_uids,
  cop_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{snu_uids}{List of UIDs corresponding to DATIM organisation units.}

\item{cop_year}{The COP year for the data.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
Tibble of orgunits mapped to orgunit name, PSNU name, & PSNU uid.
}
\description{
Certain special PSNUs (like DREAMS) are part of the
target setting process, but may exist at a level in the
organisation unit hierarchy other than the COP Prioritization level.
For organisation units which exist at the prioritization level,
their prioritization should be left as is. For organisation units
which do not exist at the level at which prioritization is set,
the parent prioritization should be used.
}
