% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPSNUs}
\alias{getPSNUs}
\title{Pull list of PSNUs from DATIM and format based on datapackr structure.}
\usage{
getPSNUs(datapack_uid)
}
\arguments{
\item{datapack_uid}{A unique ID specifying the PEPFAR Operating Unit 
the Site Tool belongs to.}
}
\value{
Data frame of PSNUs
}
\description{
Queries DATIM to extract list of PSNUs for specified Data Pack UID and adds
additional PSNUs not currently in DATIM as needed.
}
