% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getSaneName}
\alias{getSaneName}
\title{Get Sane Name for Data Pack Tool}
\usage{
getSaneName(datapack_name)
}
\arguments{
\item{datapack_name}{A string from the \code{d$info$datapack_name} object.}
}
\value{
String with the sane name.
}
\description{
Takes a Data Pack tool name and generates a
"Sane name" for the tool which has no spaces or punctuation.
}
