% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteList.R
\name{getSiteList}
\alias{getSiteList}
\title{Compile Site List for Site Tool}
\usage{
getSiteList(
  org_unit_uids,
  include_mil = TRUE,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{org_unit_uids}{Character vector of DATIM country IDs. This can only
include countries. Regional Operating Unit uids will be converted to a list of 
descendant country uids}

\item{include_mil}{When TRUE, will append _Military organisation nodes to
site list.}
}
\description{
Extracts full list of sites from DATIM, including _Military nodes where
specified, and compiles these into the list as needed for use in the PEPFAR
Site Tool.
}
