% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullTechArea.R
\name{getTechArea}
\alias{getTechArea}
\title{Get  Map of Tech Areas to Data Element IDs}
\usage{
getTechArea(
  dataElements = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{dataElements}{List of dataElements to filter against.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Dataframe of Tech Areas mapped to dataElements
}
\description{
Map Tech Area to dataElement id
}
