% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidCategoryOptions.R
\name{getValidCategoryOptions}
\alias{getValidCategoryOptions}
\title{Get Valid Category Options}
\usage{
getValidCategoryOptions(cop_year = getCurrentCOPYear())
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
Dataframe of all valid categoryOptions for specified FY
}
\description{
Returns full list of all valid categoryOptions for specified FY.
}
