% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getValidOrgUnits}
\alias{getValidOrgUnits}
\title{Return a data frame of valid organisation units
based on the COP year}
\usage{
getValidOrgUnits(cop_year = NA)
}
\arguments{
\item{cop_year}{The COP Year}
}
\value{
A data frame of organisation units along with their attributes.
}
\description{
Return a data frame of valid organisation units
based on the COP year
}
