% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleAGYW.R
\name{handleAGYW}
\alias{handleAGYW}
\title{Handle AGYW data from a Data Pack}
\usage{
handleAGYW(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Used to process AGYW sheets in a Data Pack
alerts on missing or invalid SNUs.
}
