% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlePrioritization.R
\name{handlePrioritization}
\alias{handlePrioritization}
\title{Handle Prioritization data from a Data Pack}
\usage{
handlePrioritization(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Used to process the Prioritization sheet in a Data Pack
alerts when there are invalid or blank prioritizations.
}
