% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleTX.R
\name{handleTX}
\alias{handleTX}
\title{Handle TX data from a Data Pack}
\usage{
handleTX(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Used to process TX sheets in a Data Pack (legacy).
}
