% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{handshake_file}
\alias{handshake_file}
\title{Test an input file for read access and type. Prompt if issues.}
\usage{
handshake_file(path = NA, extension)
}
\arguments{
\item{path}{Filepath to test and use.}

\item{extension}{File extension to test for. (Do not include leading period.)}
}
\value{
Character vector containing valid filepath for further use.
}
\description{
Supplied a filepath, will test that this file is readable and is of the
correct filetype. If either readability or file type is invalid, will prompt
for user selection of correct filepath via computer window.
}
