% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{headerRow}
\alias{headerRow}
\title{Tool header rows}
\usage{
headerRow(tool)
}
\arguments{
\item{tool}{"Data Pack", "Data Pack Template", "Site Tool", "Site Tool Template",
"Mechanism Map", or "Site Filter".}
}
\value{
Header row
}
\description{
Tool header rows
}
