% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_import_file.R
\name{imputePrioritizations}
\alias{imputePrioritizations}
\title{imputePrioritizations}
\usage{
imputePrioritizations(prio, psnu_import_file)
}
\arguments{
\item{prio}{Data frame consisting of orgUnit (as a uid) and prioritization
as a character.}

\item{psnu_import_file}{DHIS2 import file to convert}
}
\value{
prio
}
\description{
Utility function to handle situations where DREAMS PSNUs
do not have an explicitly defined prioritization level. The prioritization
of the parent organisation unit will be used.
}
