% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorMap}
\alias{indicatorMap}
\title{Map of Data Pack dataelements and categoryoptioncombos.}
\format{\describe{
  \item{sheet_name}{Lists the sheet/tab name as used in both the Data Pack.}
  \item{indicatorCode}{Code used in the Data Pack to uniquely
  identify each distinct programmatic area of target setting.}
  \item{typeOptions}{Either DSD or TA. The crossing of these with
  \code{indicatorCode} roughly corresponds to DATIM dataelements.}
  \item{dataelementuid}{DATIM uid used to uniquely describe the combination
  of \code{indicatorCode} and \code{typeOption}}
  \item{dataset}{Name of DATIM dataset associated with the listed
  \code{dataelementuid}, either "MER", "SUBNAT", or "IMPATT".}
  \item{validAges}{Age options allowable in the Data Pack for each
  \code{indicatorCode}}
  \item{validSexes}{Sex options allowable in the Data Pack for each
  \code{indicatorCode}}
  \item{validKPs}{Key Population options allowable in the Data Pack for each
  \code{indicatorCode}}
  \item{categoryoptioncombouid}{DATIM uid used to uniquely describe the
  combination of all disaggregates (Age, Sex, KP, etc.) for each
  \code{indicatorCode}}
}}
\usage{
indicatorMap
}
\description{
Maps from Data Pack indicators (which omit reference to DSD/TA
and are structured differently from DATIM dataelements) to DATIM dataelements
and categoryoptioncombos to allow for easy processing of Data Pack and Site
Tool data for import into and validation against DATIM.
}
\keyword{datasets}
