% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{initiatePDAPJob}
\alias{initiatePDAPJob}
\title{Title}
\usage{
initiatePDAPJob(job_type, datim_export, org_unit_id, period_id)
}
\arguments{
\item{job_type}{Type of the job, i.e. target_setting_tool}

\item{datim_export}{Location of the S3 file}

\item{org_unit_id}{UID of the organisation unit}

\item{period_id}{ISO8601 formatted period ID}
}
\value{
Returns the result of the job initiation
}
\description{
Title
}
