% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{interactive_message}
\alias{interactive_message}
\title{Uses r message instead of print if session is interactive.}
\usage{
interactive_message(x)
}
\arguments{
\item{x}{Message to print.}
}
\value{
Printed message, \code{x}.
}
\description{
Supplied a message, will share as r message() only if the session is currently interactive.
}
