% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{interactive_print}
\alias{interactive_print}
\title{Prints message if session is interactive.}
\usage{
interactive_print(x)
}
\arguments{
\item{x}{Message to print.}
}
\value{
Printed message, \code{x}.
}
\description{
Supplied a message, will print it only if the session is currently interactive.
}
