% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_fresh}
\alias{is_fresh}
\title{Checks whether cache age is fresh}
\usage{
is_fresh(time, max_age)
}
\arguments{
\item{time}{The timestamp for the cache}

\item{max_age}{The maximum age allowed for a cache to be considered fresh}
}
\value{
A logical vector
}
\description{
Checks whether cache age is fresh
}
