% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_uidish}
\alias{is_uidish}
\title{Is UID-ish}
\usage{
is_uidish(string, ish = FALSE)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible
to one.}

\item{ish}{Logical. If TRUE, looks for the UID in all parts of string, rather
than requiring the string be only the UID.}
}
\value{
A logical vector.
}
\description{
Tests whether a character string matches the regex of a DHIS2
11-digit UID.
}
