% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{listWorkbookContents}
\alias{listWorkbookContents}
\title{listWorkbookContents}
\usage{
listWorkbookContents(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
d
}
\description{
listWorkbookContents
}
