% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataPack.R
\name{loadDataPack}
\alias{loadDataPack}
\title{Load a Data Pack from a supplied filepath.}
\usage{
loadDataPack(submission_path = NULL)
}
\arguments{
\item{submission_path}{Local path to the file to import.}
}
\value{
Data Pack object
}
\description{
loadDataPack returns a Data Pack object conserving styles and
formatting of the original Data Pack .xlsx file, as well as other metadata
necessary for processing and analyzing data in the Data Pack.
}
\author{
Scott Jackson
}
