% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataPack.R
\name{loadSheets}
\alias{loadSheets}
\title{Load a DataPack from a supplied filepath.}
\usage{
loadSheets(d, sheets = NULL)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheets}{Character vector of sheet names to load. The default is NULL
which loads all sheets.}
}
\value{
DataPack object.
}
\description{
\code{loadSheets} uses \code{readSheet} to load data from specified sheets
into the DataPack \code{d} object for use in further functions.
}
\author{
Scott Jackson
}
