% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTool.R
\name{loadTool}
\alias{loadTool}
\title{loadTool}
\usage{
loadTool(
  submission_path = NULL,
  cop_year = NULL,
  tool = NULL,
  country_uids = NULL
)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}
}
\value{
d Sidecar object.
}
\description{

}
