% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEs_COCs_COs_Cs.R, R/map_COCs_to_COs.R
\name{map_COCs_to_COs}
\alias{map_COCs_to_COs}
\title{Map categoryoptioncombos to underlying categoryoptions.}
\usage{
map_COCs_to_COs()

map_COCs_to_COs()
}
\value{
dataframe of categoryoptioncombos mapped to underlying category
options

dataframe of categoryoptioncombos mapped to underlying category
options
}
\description{
maps categoryoptioncombos to underlying categoryoptions.

maps categoryoptioncombos to underlying categoryoptions.
}
