% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_DataPack_DATIM_DEs_COCs}
\alias{map_DataPack_DATIM_DEs_COCs}
\title{Map of indicators from Data Pack indicator codes to DATIM dataelements
and categoryoptioncombos}
\format{
\describe{
  \item{indicator_code}{Code used in the Data Pack to uniquely identify each
  distinct programmatic area of target setting.}
  \item{categoryoption_specified}{}
  \item{valid_ages.name}{Age disaggregate}
  \item{valid_ages.id}{Age disaggregate UID}
  \item{valid_sexes.name}{Sex disaggregate}
  \item{valid_sexes.id}{Sex disaggregate UID}
  \item{valid_kps.name}{KP disaggregate}
  \item{valid_kps.id}{KP disaggregate UID}
  \item{categoryOptions.ids}{}
  \item{support_type}{Either DSD or TA. The crossing of these with
  \code{indicatorCode} roughly corresponds to DATIM dataelements.}
  \item{dataelement}{DATIM uid for dataElements.}
  \item{hts_modality}{}
  \item{tech_area}{}
  \item{numerator_denominator}{}
  \item{dataelement.y}{}
  \item{categoryoptioncombo}{}
  \item{categoryoptioncombouid}{DATIM uid for categoryOptionCombos.}
  \item{resultstatus}{}
  \item{resultststaus_inclusive}{}
  \item{disagg_type}{}
  \item{technical_area}{}
  \item{top_level}{}
}
}
\usage{
map_DataPack_DATIM_DEs_COCs
}
\description{
Dataset that maps Data Pack indicators to dataelements and
categoryoptioncombos in DATIM, used for mapping datasets
extracted from Data Packs to DATIM import file structure.
}
\keyword{datasets}
