% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{map_DATIM_dataset_period}
\alias{map_DATIM_dataset_period}
\title{map_DATIM_dataset_period}
\usage{
map_DATIM_dataset_period(
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2_session}{DHIS2 Session id}
}
\value{
Returns a tibble of dataset uids mapped to valid periods and datastreams
}
\description{
Returns a tibble of dataset uids mapped to valid periods and datastreams
}
