% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{map_datim_join}
\alias{map_datim_join}
\title{map_datim_join}
\usage{
map_datim_join(data, second_data_set)
}
\arguments{
\item{data}{the dataset needed}

\item{second_data_set}{the datim_map used for the pack for datim function}
}
\value{
A dataframe, renamed and joined
}
\description{
Tests whether a character string matches the regex of a DHIS2
11-digit UID. Vectorized over \code{string} and \code{pattern}.
}
