% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoStructure.R
\name{memoStructure}
\alias{memoStructure}
\title{COP Memo Target Table Structure}
\usage{
memoStructure(d, d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
d object with d$memo$structure
}
\description{
COP Memo Target Table Structure
}
