% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mergeDatapack}
\alias{mergeDatapack}
\title{merge two datapacks into one d object.}
\usage{
mergeDatapack(d1 = d1, d2 = d2)
}
\arguments{
\item{d1}{the first d object}

\item{d2}{the second d object}
}
\value{
combined d object
}
\description{
If two datapacks are supplied, they will be merged into one d object.
}
