% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{import}
\name{null-default}
\alias{null-default}
\alias{\%||\%}
\title{Default value for `NULL`}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang:op-null-default]{\%||\%}}}
}}

