% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packDataPack.R
\name{packDataPack}
\alias{packDataPack}
\title{Pack a Data Pack}
\usage{
packDataPack(
  d,
  model_data = NULL,
  spectrum_data = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{model_data}{Data from DATIM needed to pack into a COP Data Pack.}

\item{spectrum_data}{Spectrum output, as a dataframe.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
Exports a Data Pack to Excel within \code{output_folder}.
}
\description{
Takes a Data Pack template, combines it with data pulled from DATIM API, and
produces a Data Pack ready for distribution.
}
