% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packDataPackSheet.R
\name{packDataPackSheet}
\alias{packDataPackSheet}
\title{Pack data into a Data Pack sheet}
\usage{
packDataPackSheet(wb, sheet, row_headers, data)
}
\arguments{
\item{wb}{datapackr list object.}

\item{sheet}{Specified sheet within wb}

\item{row_headers}{Structure to use in row_headers}

\item{data}{Data to write}
}
\value{
wb with specified sheet packed with data
}
\description{
Packs data into a specified Data Pack sheet
}
