% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSheets.R
\name{packDataPackSheets}
\alias{packDataPackSheets}
\title{Loop through and populate normal Data Pack sheets}
\usage{
packDataPackSheets(wb, country_uids, model_data)
}
\arguments{
\item{wb}{datapackr list object.}

\item{country_uids}{datapackr list object.}

\item{model_data}{}
}
\value{
wb with all sheets written except SNU x IM
}
\description{
Loops through all normally structured sheets in a submitted Data Pack
and writes data.
}
