% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packForDATIM.R
\name{packForDATIM}
\alias{packForDATIM}
\title{packForDATIM}
\usage{
packForDATIM(d, type = NULL)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{type}{Type of dataset to prep for DATIM. Choose from \code{PSNUxIM},
\code{SUBNAT_IMPATT}, \code{OPU PSNUxIM}, or \code{Undistributed MER}.}
}
\value{
d object
}
\description{
Flexible function that allows packaging of a variety of datapackr outputs as
DATIM or PAW import files. All data is packed with this function and then properly
filtered with createDATIMExport or createPAWExport for the right purpose export.
}
