% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packForDATIM_UndistributedMER.R
\name{packForDATIM_UndistributedMER}
\alias{packForDATIM_UndistributedMER}
\title{Pack Undistributed MER Data for DATIM (NOT IMPORT)}
\usage{
packForDATIM_UndistributedMER(data, cop_year)
}
\arguments{
\item{data}{Dataframe to pack for DATIM. Any mechanism attribution will be ignored.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
MER data formatted as standard DATIM import file, with all data
allocated against the default mechanism UID.
}
\description{
Packs undistributed MER data from Data Pack for use in analytics (NOT IMPORT).
}
