% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMechanisms.R
\name{packMechMap}
\alias{packMechMap}
\title{Create Mechanism Mapping tool}
\usage{
packMechMap(datapack_name, FY, output_path = NA, includeFACTS = FALSE,
  FACTSMechs_path = NA)
}
\arguments{
\item{datapack_name}{Usually the same as Operating Unit, except in rare
cases for some PEPFAR Regional Operating Units. Call
\code{datapackr::dataPackMap} to see syntax for Regional Programs.}

\item{FY}{Numeric value of COP Fiscal Year. (For COP19/FY20 targets, enter
\code{2019}.)}

\item{output_path}{A local path directing to the folder where you would like
outputs to be saved. If not supplied, will output to working directory.}

\item{includeFACTS}{Logical. If \code{TRUE}, \code{packMechMap} will corrob-
orate FY2019 DATIM mechanisms against FACTS Info Mechanisms and add any
missing mechanisms. Default is \code{FALSE}.}

\item{FACTSMechs_path}{A local file path directing to the extract of FY2019
FACTS Info Mechanisms. Default is \code{NA}.}
}
\value{
Excel workbook containing features necessary to allow various mechanism
mapping scenarios, which can be fed back into the Data Pack site distribution
process.
}
\description{
Produces an Excel tool to allow PEPFAR Country Teams to map shifting
Implementing Mechanism service patterns to allow automated distribution of
site-level data between distinct mechanisms.
}
