% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packPSNUxIM.R
\name{packPSNUxIM}
\alias{packPSNUxIM}
\title{packPSNUxIM}
\usage{
packPSNUxIM(
  wb,
  data,
  snuxim_model_data,
  cop_year = NULL,
  tool = "OPU Data Pack",
  schema = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{data}{Dataset containing totals for allocation within PSNUxIM tab,
formatted as a standard DHIS2 import file.}

\item{snuxim_model_data}{Export from DATIM needed to allocate data across
mechanisms in the PSNUxIM tab}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
r Sidecar object containing both an openxlsx Workbook and alert messages
}
\description{
Packs the PSNUxIM tab in either a COP or OPU Data Pack.
}
