% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSNUxIM.R
\name{packSNUxIM}
\alias{packSNUxIM}
\title{packSNUxIM(data)}
\usage{
packSNUxIM(d, snuxim_model_data_path)
}
\arguments{
\item{d}{Datapackr object}

\item{snuxim_model_data_path}{Filepath where SNU x IM distribution model is stored.}
}
\value{
d
}
\description{
Packs SNUxIM data prepared from unPackSNUxIM for import to DATIM.
}
