% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSNUxIM_2020.R
\name{packSNUxIM_2020}
\alias{packSNUxIM_2020}
\title{Pack SNUxIM Data (COP20 Only)}
\usage{
packSNUxIM_2020(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
d
}
\description{
Packs SNUxIM data prepared from unPackSNUxIM for import to DATIM.
Only works for COP20 Data Packs.
}
