% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSNUxIM_OPU.R
\name{packSNUxIM_OPU}
\alias{packSNUxIM_OPU}
\title{Pack SNUxIM tab for OPU Data Pack (COP20 Only)}
\usage{
packSNUxIM_OPU(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
d
}
\description{
Packs SNUxIM tab for OPU Data Packs.
}
