% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSUBNAT_IMPATT.R
\name{packSUBNAT_IMPATT}
\alias{packSUBNAT_IMPATT}
\title{packSUBNAT_IMPATT(data)}
\usage{
packSUBNAT_IMPATT(data, cop_year = getCurrentCOPYear())
}
\arguments{
\item{data}{SUBNAT/IMPATT dataframe to pack for DATIM.}

\item{cop_year}{COP Year}
}
\value{
Dataframe of SUBNAT & IMPATT data ready for DATIM ingestion.
}
\description{
Takes the outputs of the \code{\link{unPackSheets}} function and
    recompiles the dataframe containing SUBNAT and IMPATT data,
    \code{d$data$SUBNAT_IMPATT} into a standard DATIM import file.
}
