% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackCountryUIDs.R
\name{parsePSNUs}
\alias{parsePSNUs}
\title{Extract PSNUs from submitted file}
\usage{
parsePSNUs(
  submission_path,
  tool,
  cop_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Data frame of parsed PSNUs.
}
\description{
When supplied a submission path, tool type and COP year,
will return a data frame of PSNU, psnu_uid,country_name, and country_uid
If there are malformed PSNU UIDs in the file, an error will be thrown.
}
