% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{paste_oxford}
\alias{paste_oxford}
\title{Compile a list ending with different final collapse}
\usage{
paste_oxford(..., final = "and", oxford = TRUE)
}
\arguments{
\item{...}{- one or more R objects, to be converted to character vectors}

\item{final}{- conjunction to use before serial list item (usually 'and' or 'or')}

\item{oxford}{- TRUE/FALSE indicating whether to use the Oxford (i.e., serial) comma}
}
\value{
A character vector of the concatenated values.
}
\description{
Compile a list ending with different final collapse
}
