% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{periodInfo}
\alias{periodInfo}
\title{periodInfo}
\format{
\describe{
  \item{periodid}{Unique DATIM id corresponding to period.}
  \item{iso}{ISO period name.}
  \item{startdate}{Start date of period.}
  \item{enddate}{End date of period.}
  \item{periodtype}{Period type, whether Financial, Quarterly, etc.}
}
}
\usage{
periodInfo
}
\description{
Dataset of current COP period info.
}
\keyword{datasets}
