% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{pick_schema}
\alias{pick_schema}
\title{Pick correct schema}
\usage{
pick_schema(cop_year, tool)
}
\arguments{
\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{tool}{Either Data Pack or OPU Data Pack}
}
\value{
Schema file for given cop_year and tool type
}
\description{
Pick correct schema
}
