% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{pick_template_path}
\alias{pick_template_path}
\title{Pick correct template filepath}
\usage{
pick_template_path(cop_year, tool)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}
}
\value{
Template filepath for given cop_year and tool.
}
\description{
Pick correct template filepath
}
