% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackYear2Sheet.R
\name{pickUIDFromType}
\alias{pickUIDFromType}
\title{Title pickUIDFromType is a utility function used to obtain a
particular UID from a supplied list based on the type of
value we are dealing with. In the Year2 tab, values disaggregated
by AgeSex, KP and EID are co-mingled and must be separated based
on available disaggregates.}
\usage{
pickUIDFromType(type, de_uid_list)
}
\arguments{
\item{type}{Disaggregate type (e.g. AgeSex, KP, EID)}

\item{de_uid_list}{A list of possible UIDs (e.g. c("UID1","{KP}UID2" )}
}
\value{
A UID based on the supplied type. Returns NA if no UID is matched.
}
\description{
Title pickUIDFromType is a utility function used to obtain a
particular UID from a supplied list based on the type of
value we are dealing with. In the Year2 tab, values disaggregated
by AgeSex, KP and EID are co-mingled and must be separated based
on available disaggregates.
}
