% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_memo_data.R
\name{prepare_memo_data_by_agency}
\alias{prepare_memo_data_by_agency}
\title{Prepare Memo Data By Agency Level}
\usage{
prepare_memo_data_by_agency(df, memo_structure)
}
\arguments{
\item{df}{An analytics table, either d$memo$datim$analytics or
d$memo$datapack$analytics}

\item{memo_structure}{Structure of the memo d$memo$structure}
}
\value{
A dataframe of COP indicators aggregated to the prioritization level
}
\description{
Prepare Memo Data By Agency Level
}
