% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_model_data.PSNUxIM.R
\name{prepare_model_data.PSNUxIM}
\alias{prepare_model_data.PSNUxIM}
\title{packSNUxIM(data)}
\usage{
prepare_model_data.PSNUxIM(snuxim_model_data, country_uids)
}
\value{
d
}
\description{
Packs SNUxIM data prepared from unPackSNUxIM for import to DATIM.
}
