% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMemoData.R
\name{prepareMemoData}
\alias{prepareMemoData}
\title{Title}
\usage{
prepareMemoData(d, d2_session, memo_type, include_no_prio = TRUE)
}
\arguments{
\item{d}{}

\item{d2_session}{}

\item{memo_type}{One of the following: 
datapack: Create the memo based on the data in the datapack or OPU datapack
datim: Create the memo based on data currently in DATIM
comparison: Create a comparison memo with data from both DATIM and the datapack}
}
\value{

}
\description{
Title
}
