% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_memo_data.R
\name{prepareMemoData_by_agency}
\alias{prepareMemoData_by_agency}
\title{Prepare Memo Data By Agency Level}
\usage{
prepareMemoData_by_agency(df, memoStructure)
}
\arguments{
\item{df}{An analytics table, either d$memo$datim$analytics or
d$memo$datapack$analytics}

\item{memoStructure}{Structure of the memo d$memo$structure}
}
\value{
A dataframe of COP indicators aggregated to the prioritization level
}
\description{
Prepare Memo Data By Agency Level
}
