% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_memo_data.R
\name{prepareMemoData_by_partner}
\alias{prepareMemoData_by_partner}
\title{Prepare Memo Data By Partner}
\usage{
prepareMemoData_by_partner(df, memoStructure, indicators)
}
\arguments{
\item{df}{An analytics table, either d$memo$datim$analytics
or d$memo$datapack$analytics}

\item{memoStructure}{Structure of the memo d$memo$structure}

\item{indicators}{A dataframe of indicators d$memo$inds}
}
\value{
A dataframe of COP indicators aggregated to the partner level
}
\description{
Prepare Memo Data By Partner
}
