% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_memo_data.R
\name{prepareMemoData_by_prio}
\alias{prepareMemoData_by_prio}
\title{Prepare Memo Data By Prioritization Level}
\usage{
prepareMemoData_by_prio(df, memoStructure, include_no_prio = TRUE)
}
\arguments{
\item{df}{An analytics table, either d$memo$datim$analytics or
d$memo$datapack$analytics}

\item{memoStructure}{Structure of the memo d$memo$structure}

\item{include_no_prio}{If TRUE, include \code{"No Prioritiation"}
as a column in the output.}
}
\value{
A dataframe of COP indicators aggregated
to the prioritization level.
}
\description{
Prepare Memo Data By Prioritization Level
}
