% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMemoData.R
\name{prepareMemoDataByPrio}
\alias{prepareMemoDataByPrio}
\title{Title}
\usage{
prepareMemoDataByPrio(df, memo_structure, include_no_prio = TRUE)
}
\arguments{
\item{df}{}

\item{memo_structure}{}

\item{include_no_prio}{}
}
\value{

}
\description{
Title
}
