% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMemoData.R
\name{prepareMemoMetadata}
\alias{prepareMemoMetadata}
\title{Title}
\usage{
prepareMemoMetadata(d, d2_session, memo_type)
}
\arguments{
\item{d}{Datapackr d object}

\item{d2_session}{Datimutils d2_session object#'}

\item{memo_type}{One of the following: 
datapack: Create the memo based on the data in the datapack or OPU datapack
datim: Create the memo based on data currently in DATIM
comparison: Create a comparison memo with data from both DATIM and the datapack}
}
\value{

}
\description{
Title
}
