% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePartnerMemoTable.R
\name{preparePartnerMemoTable}
\alias{preparePartnerMemoTable}
\title{Title}
\usage{
preparePartnerMemoTable(
  d,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr d object}

\item{d2_session}{Datimutils sesssion object}
}
\value{
Tibble in wide format with Agency,Partner,Mechanism and all
grouped indicators along the columns
}
\description{
Title
}
