% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePrioTable.R
\name{preparePrioTable}
\alias{preparePrioTable}
\title{Title}
\usage{
preparePrioTable(d, d2_session)
}
\arguments{
\item{d}{Datapackr d object}

\item{d2_session}{}
}
\value{
Datapackr d object with d$data$memo table
}
\description{
Title
}
