% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSheetData.R
\name{prepareSheetData}
\alias{prepareSheetData}
\title{Prepare sheet-specific dataset for writing into a Data Pack sheet.}
\usage{
prepareSheetData(
  sheet,
  org_units,
  schema = datapackr::data_pack_schema,
  sheet_data,
  cop_year = getCurrentCOPYear()
)
}
\arguments{
\item{sheet}{Specified sheet within wb.}

\item{org_units}{A list of DATIM organisation units.}

\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{sheet_data}{Dataset to use as input for packing Data Pack.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
dataframe of data prepared for Data Pack
}
\description{
Prepares provided dataset for writing into specified Data Pack sheet.
}
