% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messageQueue.R
\name{printMessages.MessageQueue}
\alias{printMessages.MessageQueue}
\title{Print Message - Message Queue S3 method}
\usage{
\method{printMessages}{MessageQueue}(x)
}
\arguments{
\item{x}{A MessageQueue object}
}
\value{
Returns a formatted output to the console
}
\description{
Internal S3 method to deal with printing messages
}
