% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prioritizations}
\alias{prioritizations}
\title{Mapping of DATIM Prioritization numerals to strings.}
\format{\describe{
  \item{value}{Numeral associated with DATIM Prioritization.}
  \item{Prioritization}{Name associated with DATIM Prioritization}
}}
\usage{
prioritizations
}
\description{
Maps DATIM Prioritizations from their coded numerals (1-8) to
their more descriptive names.
}
\keyword{datasets}
