% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullCOPindicators.R
\name{pull_COPindicators}
\alias{pull_COPindicators}
\title{Extract COP indicators list from DATIM}
\usage{
pull_COPindicators(
  cop_year = datapackr::getCurrentCOPYear(),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Dataframe of COP indicators retrieved from DATIM
}
\description{
Queries DATIM to return list of COP indicators for given COP year.
}
