% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullDATIMCodeList.R
\name{pullDATIMCodeList}
\alias{pullDATIMCodeList}
\title{Pull DATIM code list.}
\usage{
pullDATIMCodeList(dataset)
}
\arguments{
\item{dataset}{Dataset uid to query.}
}
\value{
Code list as dataframe.
}
\description{
Pulls the DATIM code list for specified datasets.
}
