% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullFullCodeList.R
\name{pullFullCodeList}
\alias{pullFullCodeList}
\title{Pull & combine all MER, SUBNAT, IMPATT code lists for specified FY.}
\usage{
pullFullCodeList(
  FY = getCurrentCOPYear() + 1,
  datastream = c("mer_targets", "mer_results", "subnat_targets", "subnat_results",
    "impatt"),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{FY}{Reporting FY for which to filter active code lists.}

\item{datastream}{Specify MER, SUBNAT, or IMPATT, or omit to specify all.}

\item{d2_session}{the handle to use}
}
\value{
Combined code list as dataframe.
}
\description{
Pulls all code lists for MER, SUBNAT, and IMPATT for a specified FY and
combines these into a unique list.
}
