% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataPack.R
\name{readSheet}
\alias{readSheet}
\title{Read data from a DataPack object}
\usage{
readSheet(
  d,
  sheet = 1,
  range = NULL,
  col_names = TRUE,
  col_types = "text",
  na = "",
  guess_max = 1000,
  progress = readxl::readxl_progress(),
  .name_repair = "minimal"
)
}
\arguments{
\item{d}{DataPack object, created via \code{loadDataPack}.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{range}{A cell range to read from, as described in \link[readxl]{cell-specification}.
Includes typical Excel ranges like "B3:D87", possibly including the sheet
name like "Budget!B2:G14", and more. Interpreted strictly, even if the
range forces the inclusion of leading or trailing empty rows or columns.
Takes precedence over \code{skip}, \code{n_max} and \code{sheet}.}

\item{col_names}{\code{TRUE} to use the first row as column names, \code{FALSE} to get
default names, or a character vector giving a name for each column. If user
provides \code{col_types} as a vector, \code{col_names} can have one entry per
column, i.e. have the same length as \code{col_types}, or one entry per
unskipped column.}

\item{col_types}{Either \code{NULL} to guess all from the spreadsheet or a
character vector containing one entry per column from these options:
"skip", "guess", "logical", "numeric", "date", "text" or "list". If exactly
one \code{col_type} is specified, it will be recycled. The content of a cell in
a skipped column is never read and that column will not appear in the data
frame output. A list cell loads a column as a list of length 1 vectors,
which are typed using the type guessing logic from \code{col_types = NULL}, but
on a cell-by-cell basis.}

\item{na}{Character vector of strings to interpret as missing values. By
default, readxl treats blank cells as missing data.}

\item{guess_max}{Maximum number of data rows to use for guessing column
types.}

\item{progress}{Display a progress spinner? By default, the spinner appears
only in an interactive session, outside the context of knitting a document,
and when the call is likely to run for several seconds or more. See
\code{\link[readxl:readxl_progress]{readxl_progress()}} for more details.}

\item{.name_repair}{Handling of column names. By default, readxl ensures
column names are not empty and are unique. If the tibble package version is
recent enough, there is full support for \code{.name_repair} as documented in
\code{\link[tibble:tibble]{tibble::tibble()}}. If an older version of tibble is present, readxl falls
back to name repair in the style of tibble v1.4.2.}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Reads data from a sheet in a DataPack object. This function is
essentially a wrapper for \code{readxl}'s \code{read_excel} function, but with additional
support for selecting default parameters per DataPack setup.
}
\author{
Scott Jackson
}
