% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_approval_memo.R
\name{render_agency_table}
\alias{render_agency_table}
\title{Render Agency-Level Memo Target Table}
\usage{
render_agency_table(memo_doc, agency_table, ou_name)
}
\arguments{
\item{memo_doc}{\code{Officer} document object containing
the target memo tables.}

\item{agency_table}{A table of target data aggregated to the agency
level, either d$memo$datim$by_agency or d$memo$datapack$by_agency.}

\item{ou_name}{Name of the organisation unit.}
}
\value{

}
\description{
Render Agency-Level Memo Target Table
}
