% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_approval_memo.R
\name{render_partner_table}
\alias{render_partner_table}
\title{Render Partner-Level Memo Target Tables}
\usage{
render_partner_table(memo_doc, partners_table, memo_structure)
}
\arguments{
\item{memo_doc}{\code{Officer} document object containing
the target memo tables.}

\item{partners_table}{A table of target data aggregated to the partner and
mechanism level, either d$memo$datim$by_partner or
d$memo$datapack$by_partner.}

\item{memo_structure}{Structure of the memo d$memo$structure}
}
\value{

}
\description{
Render Partner-Level Memo Target Tables
}
