% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemo.R
\name{renderAgencyTable}
\alias{renderAgencyTable}
\title{Render Agency-Level Memo Target Table}
\usage{
renderAgencyTable(memo_doc, agency_table, ou_name, source_type)
}
\arguments{
\item{memo_doc}{\code{Officer} document object containing
the target memo tables.}

\item{agency_table}{A table of target data aggregated to the agency
level, either d$memo$datim$by_agency or d$memo$datapack$by_agency.}

\item{ou_name}{Name of the organisation unit.}

\item{source_type}{Indicates whether the data for a COP Approval Memo table
should come from the Data Pack or from DATIM. Values can be either
\code{datapack} or \code{datim}.}
}
\value{
Formatted agency-level memo target table
}
\description{
Render Agency-Level Memo Target Table
}
