% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemoDOCX.R
\name{renderPartnerTable}
\alias{renderPartnerTable}
\title{Title}
\usage{
renderPartnerTable(doc, partners_table, memo_structure)
}
\arguments{
\item{doc}{}

\item{partners_table}{}

\item{memo_structure}{}
}
\value{

}
\description{
Title
}
