% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemoDOCX.R
\name{renderPrioTable}
\alias{renderPrioTable}
\title{Title}
\usage{
renderPrioTable(prio_table, ou_name, draft = TRUE)
}
\arguments{
\item{prio_table}{}

\item{ou_name}{}
}
\value{

}
\description{
Title
}
