% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataPack.R, R/rePackSNUxIM.R
\name{rePackPSNUxIM}
\alias{rePackPSNUxIM}
\title{rePackSNUxIM(d)}
\usage{
rePackPSNUxIM(d)

rePackPSNUxIM(d)
}
\arguments{
\item{d}{datapackr list object containing at least \code{d$keychain$MER} and
\code{d$data$SNUxIM}}

\item{d}{Datapackr object}
}
\value{
A datapackr list object, \code{d}, storing at least 1 dataframe of
   data at the SNU x IM level, \code{d$data$distributedMER}.

d
}
\description{
Takes the output of the \code{\link{unPackSNUxIM}} and
    \code{\link{unPackSheets}} functions and delicatety combines these to create
    a single dataframe at the PSNU x IM level.

Takes the output of the \code{\link{unPackSNUxIM}} and
    \code{\link{unPackSheets}} functions and delicatety combines these to create
    a single dataframe at the PSNU x IM level.
}
