% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{round_trunc}
\alias{round_trunc}
\title{Round at 0.5 toward integer with highest absolute value}
\usage{
round_trunc(x)
}
\arguments{
\item{x}{A number.}
}
\value{
An integer.
}
\description{
In normal R rounding, if the first digit to be dropped is exactly 5, R uses
the standard programming convention of rounding to the nearest even number.
This can have some annoying effects.

This function rounds numbers to the nearest integer, but always rounds to the
integer with the highest absolute value when the first digit to be dropped is
exactly 5, similar to rounding in usual mathematical contexts.
}
\examples{
# If the first digit to be dropped is exactly 5, round_trunc() will round to
# integer with the highest absolute value.
round_trunc(0.5)
round_trunc(-0.5)
}
