% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rowMax}
\alias{rowMax}
\title{Take Max along row among columns matching regex}
\usage{
rowMax(df, cn, regex)
}
\arguments{
\item{df}{Dataframe}

\item{cn}{Name (character string) of Max column to create}

\item{regex}{String of regex to use in identifying columns.}
}
\value{
df
}
\description{
Take Max along row among columns matching regex
}
