% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSchema.R
\name{schema-validations}
\alias{schema-validations}
\alias{validateSchema_SkippedSheets}
\alias{validateSchema_SheetNums}
\alias{validateSchema_SheetNamesComplete}
\alias{validateSchema_InvalidDatasets}
\alias{validateSchema_InvalidColType}
\alias{validateSchema_InvalidValueType}
\alias{validateSchema_DSDSyntax}
\alias{validateSchema_TASyntax}
\alias{validateSchema_COsSyntax}
\alias{validateSchema_ValidAges}
\alias{validateSchema_ValidSexes}
\alias{validateSchema_ValidKPs}
\alias{validateSchema_Formulas}
\alias{validateSchema}
\alias{unPackSchema}
\title{Standardized Schema Checks}
\usage{
validateSchema_SkippedSheets(schema, tool, cop_year)

validateSchema_SheetNums(schema)

validateSchema_SheetNamesComplete(schema, filepath_schema)

validateSchema_InvalidDatasets(schema)

validateSchema_InvalidColType(schema)

validateSchema_InvalidValueType(schema)

validateSchema_DSDSyntax(DEs_schema, multi_uid_pattern)

validateSchema_TASyntax(DEs_schema, multi_uid_pattern)

validateSchema_COsSyntax(schema, multi_uid_pattern)

validateSchema_ValidAges(schema)

validateSchema_ValidSexes(schema)

validateSchema_ValidKPs(schema)

validateSchema_Formulas(schema)

validateSchema(schema, template_path, cop_year, tool, season)

unPackSchema(
  template_path = NULL,
  skip = NULL,
  tool = "Data Pack Template",
  cop_year = NULL
)
}
\arguments{
\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{template_path}{Local filepath to Data Pack template Excel (XLSX) file.
This file MUST NOT have any data validation formats present. If left
\code{NULL}, will select the default based on \code{cop_year} and \code{tool}.}

\item{season}{Either \code{COP} or \code{OPU}.}
}
\value{
For lower-level functions, a list of instances of failed tests. For the
higher-level \code{validateSchema}, a list object of lists of failed tests.
\itemize{
\item \code{validateSchema_SkippedSheets}:
\item \code{validateSchema_SheetNums}:
\item \code{validateSchema_SheetNamesComplete}:
\item \code{validateSchema_InvalidDatasets}:
\item \code{validateSchema_InvalidColType}:
\item \code{validateSchema_InvalidValueType}:
\item \code{validateSchema_DSDSyntax}:
\item \code{validateSchema_TASyntax}:
\item \code{validateSchema_COsSyntax} :
\item \code{validateSchema_ValidAges} :
\item \code{validateSchema_ValidSexes} :
\item \code{validateSchema_ValidKPs} :
\item \code{validateSchema_Formulas} :
\item \code{validateSchema} :
\item \code{unPackSchema} :
}
}
\description{
Standardized package functions,in terms of schema checks.
These can be run individually (e.g., \code{validateSchema_SheetNums}), or in
bulk (e.g., \code{validateSchema(template_path = "my/file/path", cop_year = 2022, tool = "Data Pack")}).
}
\concept{schema-helpers}
