% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{schema_cols}
\alias{schema_cols}
\title{Schema Columns}
\usage{
schema_cols(cop_year = getCurrentCOPYear())
}
\arguments{
\item{cop_year}{COP Year for dating as well as selection of templates.}
}
\value{
List of columns to select from template.
}
\description{
Schema Columns
}
