% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{selectOU}
\alias{selectOU}
\title{Use the console to select OU}
\usage{
selectOU()
}
\value{
An OU name, based on input selection.
}
\description{
In some cases it may be necessary to manually identify the Operating Unit
associated with a submitted Data Pack or Site Tool. This function allows
manual selection of Operating Unit using the R console.

One case where this is necessarily invoked is when OU name and OU id read
from a submitted Data Pack or Site Tool do not match one another, based on
cross-reference with DATIM organization hierarchies.
}
