% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataPack.R, R/separateDataSets.R
\name{separateDataSets}
\alias{separateDataSets}
\title{separateDataSets(d)}
\usage{
separateDataSets(d)

separateDataSets(d)
}
\arguments{
\item{d}{datapackr list object containing at least \code{d$data$targets}.}

\item{d}{Datapackr object.}
}
\value{
A datapackr list object, \code{d}, storing at least 2 dataframes of
   data extracted from submitted Data Pack or Site Tool: a \code{d$data$MER}
   dataframe containing all MER data to be distributed to site level, and/or
   \code{d$data$SUBNAT_IMPATT} containing data in the SUBNAT and IMPATT
   datasets from DATIM that can be imported into DATIM at the PSNU level.

d
}
\description{
After data has been extracted from all sheets in a Data Pack or
    Site Tool, this function separates datasets by either \code{MER} or
    \code{SUBNAT/IMPATT} and removes elements of \code{d} that are no longer
    necessary (\code{targets}, \code{extract}, and \code{sheet})

After data has been extracted from all sheets in a Data Pack,
    this function separates datasets by either \code{MER} or
    \code{SUBNAT/IMPATT} and removes elements of \code{d} that are no longer
    necessary (\code{targets}, \code{extract}, and \code{sheet})
}
