% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePrioTable.R
\name{separateIndicatorMetadata}
\alias{separateIndicatorMetadata}
\title{Separate Indicator Metadata}
\usage{
separateIndicatorMetadata(x)
}
\arguments{
\item{x}{Data frame containing indicator names which have not been separated}
}
\value{
Data frame with indicator names separated into
two columns: Indicator & Age
}
\description{
Separate Indicator Metadata
}
