% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePrioTable.R
\name{seperateIndicatorMetadata}
\alias{seperateIndicatorMetadata}
\title{Title}
\usage{
seperateIndicatorMetadata(x)
}
\arguments{
\item{x}{Data frame containing indicator names which have not been separated}
}
\value{
Data frame with indicator names separated into two columns: Indicator & Age
}
\description{
Title
}
