% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_datim_protocol}
\alias{set_datim_protocol}
\title{set_datim_protocol}
\usage{
set_datim_protocol(data)
}
\arguments{
\item{data}{the dataset needed}
}
\value{
A logical vector.
}
\description{
Tests whether a character string matches the regex of a DHIS2
11-digit UID. Vectorized over \code{string} and \code{pattern}.
}
