% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_tool_schema}
\alias{site_tool_schema}
\title{Schema describing correct structure of Site Tool template.}
\format{\describe{
  \item{sheet_num}{Lists the index value associated with the sheet name
  listed in \code{sheet_name}.}
  \item{sheet_name}{Lists the sheet/tab name as used in both the Data Pack
  and Site Tool.}
  \item{col}{Value describing the column position of each 
  \code{indicator_code}.}
  \item{col_type}{Flags whether an \code{indicator_code} is a FY20 Target
  (\code{"FY20 Target"}), row header (\code{Row Header}) or not (\code{NA}).}
  \item{tech_area}{Header text used to distinguish groups of
  \code{indicator_codes}.}
  \item{label}{String label used to describe \code{indicator_code}.}
  \item{indicator_code}{Code used in the Data Pack and Site Tool to uniquely
  identify each distinct programmatic area of target setting.}
}}
\usage{
site_tool_schema
}
\description{
This schema describes the correct structure of a Site Tool
file, generated from the template used to produce Data Packs and useful in
validating Site Tools passed through datapackr.
}
\keyword{datasets}
