% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SiteToDATIM}
\alias{SiteToDATIM}
\title{SiteToDATIM}
\format{\describe{
  \item{sheet_name}{Lists the sheet/tab name as used in the Data Pack.}
  \item{indicatorCode}{Code used in the Data Pack to uniquely identify each
  distinct programmatic area of target setting.}
  \item{typeOptions}{Either DSD or TA. The crossing of these with
  \code{indicatorCode} roughly corresponds to DATIM dataelements.}
  \item{dataelementuid}{DATIM uid for dataElements.}
  \item{dataset}{Name of DATIM dataset associated with the listed
  \code{dataelementuid}, either "MER", "SUBNAT", or "IMPATT".}
  \item{validAges}{Age disaggregate.}
  \item{validSexes}{Sex disaggregate.}
  \item{validKPs}{Key Population disaggregate.}
  \item{categoryoptioncombouid}{DATIM uid for categoryOptionCombos.}
}}
\usage{
SiteToDATIM
}
\description{
Mapping of Site Tool indicators to DATIM dataelements & categoryoptioncombos.
}
\keyword{datasets}
