% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{skip_tabs}
\alias{skip_tabs}
\title{List of tabs to skip for given tool.}
\usage{
skip_tabs(tool = "Data Pack", cop_year = getCurrentCOPYear())
}
\arguments{
\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
Character vector of tab names to skip.
}
\description{
List of tabs to skip for given tool.
}
