% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{startRow}
\alias{startRow}
\title{Tool start rows}
\usage{
startRow(tool)
}
\arguments{
\item{tool}{"Data Pack", "Data Pack Template", "Site Tool", "Site Tool Template",
"Mechanism Map", or "Site Filter".}
}
\value{
Start row
}
\description{
Tool start rows
}
