% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_wb_NAs.R
\name{strip_wb_NAs}
\alias{strip_wb_NAs}
\title{Remove errant NAs in a Data Pack}
\usage{
strip_wb_NAs(d)
}
\arguments{
\item{d}{datapackr object}
}
\value{
Updated datapackr object with NAs removed from d$tool$wb
}
\description{
Searches a Data Pack for NAs, likely introduced by opening a file with pasted
blank cell values, and replaces these with true blanks to prevent cascading
issues.
}
