% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{styleGuide}
\alias{styleGuide}
\title{Library of Openxlsx style objects to apply to both Data Pack and Site
Tool files.}
\format{
\describe{
  \item{home}{Styles for title, Data Pack Name, & PEPFAR banner.}
  \item{siteList}{Styles for site names, broken out by Community, Facility,
  National, Military, or Inactive.}
  \item{data}{Styles for title, headers, labels, and Unique ID row.}
  \item{cop21_opu}{Styles specific to cop21_opu's}
}
}
\usage{
styleGuide
}
\description{
A list object containing styles pertaining to different portions of the Data
Pack. These styles control things such as font type,font size, alignment,
border thickness, and color
}
\keyword{datasets}
