% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{supportedCOPYears}
\alias{supportedCOPYears}
\title{Returns COP Years currently supported by the package for a given tool.}
\usage{
supportedCOPYears(tool = NULL)
}
\arguments{
\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}
}
\value{
Vector of COP Years currently supported by the package for the given tool.
}
\description{
If no tool is provided, will supply COP Years supported across
all tools.
}
