% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{supportedSeasons}
\alias{supportedSeasons}
\title{Supported Seasons}
\usage{
supportedSeasons(cop_year = NULL, tool = NULL)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}
}
\value{
Character vector of seasons supported by the package. If neither
cop_year nor tool are provided, will provide list of all seasons supported.
}
\description{
Supported Seasons
}
