% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{supportedTools}
\alias{supportedTools}
\title{Supported Tools}
\usage{
supportedTools(cop_year = NULL)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
Character vector of tools supported by the package for a given cop_year.
If cop_year is not provided, will provide list of all tools supported for any
cop_year.
}
\description{
Supported Tools
}
