% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{swapColumns}
\alias{swapColumns}
\title{Swap columns between two dataframes}
\usage{
swapColumns(to, from)
}
\arguments{
\item{to}{Dataframe to pull columns into}

\item{from}{Dataframe to pull columns from}
}
\value{
A dataframe with the swapped columns
}
\description{
Replaces columns in the dataframe \code{to} with those with identical names
 in the dataframe \code{from}.
}
