% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_schema}
\alias{template_schema}
\title{Schema describing correct structure of Data Pack template.}
\format{\describe{
  \item{sheet_num}{Lists the index value associated with the sheet name
  listed in \code{sheet_name}.}
  \item{sheet_name}{Lists the sheet/tab name as used in both the Data Pack
  and Site Tool.}
  \item{col}{Value describing the column position of each 
  \code{indicatorCode}.}
  \item{indicatorCode}{Code used in the Data Pack and Site Tool to uniquely
  identify each distinct programmatic area of target setting.}
  \item{colType}{Flags whether an \code{indicatorCode} is a FY20 Target
  (\code{"FY20 Target"}) or not (\code{NA}).}
  \item{dataset}{For indicatorCodes listed as "FY20 Targets" in the 
  \code{colType} field, documents the dataset, either "MER", "IMPATT", or
  "SUBNAT"}
}}
\usage{
template_schema
}
\description{
This schema describes the correct structure of a Data Pack
file, generated from the template used to produce Data Packs and useful in
validating Data Packs passed through datapackr.
}
\keyword{datasets}
