% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{testImbalancedDistribution}
\alias{testImbalancedDistribution}
\title{Title testImbalancedDistribution}
\usage{
testImbalancedDistribution(d)
}
\arguments{
\item{d}{Datapackr d object}
}
\value{
Datapackr d object
}
\description{
Tests for any imbalanced distribution between the main DataPack sheets
and the PSNUxIM tab.
}
