% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{testInvalidDedupeValues}
\alias{testInvalidDedupeValues}
\title{testInvalidDedupeValues}
\usage{
testInvalidDedupeValues(d, header_cols)
}
\arguments{
\item{d}{Datapackr d object}

\item{header_cols}{header_cols object}
}
\value{
d object
}
\description{
Tests dedupe adjustments to determine if they are valid.
The function does not remove any data, but will flag any invalid dedupes
as an error.
}
