% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{testInvalidPSNUs}
\alias{testInvalidPSNUs}
\title{Title testInvalidPSNUs}
\usage{
testInvalidPSNUs(d)
}
\arguments{
\item{d}{Datapackr d object}
}
\value{
Datapackr d object
}
\description{
Tests for any PSNUs in the PSNUxIM sheet which are not valid.
}
