% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{testMissingDedupeRollupColumns}
\alias{testMissingDedupeRollupColumns}
\title{Title testMissingDedupeRollupColumns}
\usage{
testMissingDedupeRollupColumns(d, cols_to_keep, sheet = "PSNUxIM")
}
\arguments{
\item{d}{Datapackr d object}

\item{cols_to_keep}{cols_to_keep object}

\item{sheet}{Name of the sheet}
}
\value{
Datapackr d object
}
\description{
Title testMissingDedupeRollupColumns
}
