% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{testMissingRightSideFormulas}
\alias{testMissingRightSideFormulas}
\title{Title testMissingRightSideFormulas}
\usage{
testMissingRightSideFormulas(
  d,
  cols_to_keep,
  header_cols,
  header_row,
  blank_cols_idx,
  parsed_cells = NULL
)
}
\arguments{
\item{d}{Datapackr d object}

\item{cols_to_keep}{cols_to_keep object}

\item{header_cols}{headers_cols object}

\item{header_row}{Numeric index of the header row in the PSNUxIM tab.}

\item{blank_cols_idx}{Index of the first column which is blank in the sheet.}

\item{parsed_cells}{Primarily used for unit testing. When null,
values will be read from the PSNUxIM tab.}
}
\value{
Modified d object with data frame of columns with missing formulas.
}
\description{
Title testMissingRightSideFormulas
}
