% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{testNegativeTargetValues}
\alias{testNegativeTargetValues}
\title{Title testNegativeTargetValues}
\usage{
testNegativeTargetValues(d, header_cols)
}
\arguments{
\item{d}{Datapackr d object}

\item{header_cols}{header_cols object}
}
\value{
Datapackr d object
}
\description{
Tests for any targets values which are negative. Converts any negative
target values to numeric NAs.
}
