% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{testRoundDecimalValues}
\alias{testRoundDecimalValues}
\title{Title testRoundDecimalValues}
\usage{
testRoundDecimalValues(d)
}
\arguments{
\item{d}{Datapackr d object}
}
\value{
Datapackr d object
}
\description{
Test for any values which are not whole integers. Using the round_trunc
function, values are rounded here.
}
