% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{testRoundingDiffs}
\alias{testRoundingDiffs}
\title{Title testRoundingDiffs}
\usage{
testRoundingDiffs(d)
}
\arguments{
\item{d}{Datapackr d object}

\item{original_targets}{original_targets object}
}
\value{
Modified d object with a new object d$info$psnuxim_comparison
and if applicable, d$tests$PSNUxIM_rounding_diffs
}
\description{
Tests for rounding errors in the PSNUxIM allocation.
}
