% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{toolName_homeCell}
\alias{toolName_homeCell}
\title{Location of name of the tool on Home tab.}
\usage{
toolName_homeCell()
}
\value{
Cell reference where the name of the tool is located.
}
\description{
Location of name of the tool on Home tab.
}
