% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{uid_pattern}
\alias{uid_pattern}
\title{DHIS2 UID pattern}
\usage{
uid_pattern()
}
\value{
DHIS2 UID pattern, expressed as a regular expression in form of a
character vector.
}
\description{
Returns the DHIS2 UID pattern, expressed as a regular expression
in form of a character vector.
}
