% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackCountryUIDs.R
\name{unPackCountryUIDs}
\alias{unPackCountryUIDs}
\title{Extract country uids from submitted file}
\usage{
unPackCountryUIDs(
  submission_path,
  tool = "Data Pack",
  cop_year = getCurrentCOPYear()
)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{What type of tool is the submission file? Default is "Data Pack".}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}
}
\value{
Character vector of country_uids.
}
\description{
When supplied a submission path, will return the list of country_uids
pertaining to the file, as read from the Home tab.
}
