% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackDataChecks.R
\name{unPackDataChecks}
\alias{unPackDataChecks}
\alias{checkDupeRows}
\title{Perform various data quality checks on in-process data from Data Packs.}
\usage{
checkDupeRows(data, tool, cop_year, sheet)
}
\arguments{
\item{data}{Dataset to be validated.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{sheet}{Name of sheet from Data Pack from which \code{data} was extracted.}
}
\value{
These functions each return a list object, \code{t} containing three components:
\enumerate{
\item \code{message}: A message describing any issues encountered in the check.
\item \code{test_result}: A tibble of data flagged by the check, useful in data
quality forensics.
\item \code{has_error}: A \code{TRUE/FALSE} flag indicating whether any fatal errors
have been observed.
}
}
\description{
A series of functions to check and validate quality & integrity
of data encountered in various \code{unPack...} functions across \code{datapackr}.

\code{checkDupeRows} checks for any rows with duplicates across PSNU and other key
disaggregates.
}
