% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackDataPack.R
\name{unPackDataPack}
\alias{unPackDataPack}
\title{Unpack a submitted Data Pack}
\usage{
unPackDataPack(d, d2_session)
}
\arguments{
\item{d}{Datapackr object}

\item{d2_session}{DHIS2 Session ID}
}
\description{
Processes a submitted Data Pack (in .xlsx format) by identifying integrity
    issues, checking data against DATIM validations, and extracting data.
}
\details{
Executes the following operations in relation to a submitted Data Pack:
\enumerate{
    \item Performs integrity checks on file structure;
    \item Runs DATIM validation tests;
    \item Extracts SUBNAT and IMPATT data as a DATIM import file;
}
    
If a Data Pack is submitted as an XLSB formatted document, you must open &
re-save as an XLSX in order to process it with this function.

Once this function is called, it will present a dialog box where you can
select the file location of the Data Pack to be processed.

If Operating Unit \code{name} and \code{id} in the Data Pack's \strong{Home}
tab do not match based on cross-reference with DATIM organization
hierarchies, you will be prompted to manually select the correct Operating
Unit associated with the Data Pack to be processed. Enter this information in
the Console as directed.

The final message in the Console prints all warnings identified in the Data
Pack being processed.
}
