% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackDataPackName.R
\name{unPackDataPackName}
\alias{unPackDataPackName}
\title{Extract the name of the datapack}
\usage{
unPackDataPackName(submission_path, tool)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}
}
\value{
Character vector of the name of the data pack.
}
\description{
When supplied a submission path, will return the name of the datapack.
}
