% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSheet.R
\name{unPackDataPackSheet}
\alias{unPackDataPackSheet}
\title{Unpack a Data Pack sheet.}
\usage{
unPackDataPackSheet(
  d,
  sheet,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{Data Pack sheet.}

\item{d2_session}{DHIS2 Session id}
}
\value{
d
}
\description{
Within a submitted Data Pack (directed to by
   \code{d$keychain$submission_path}), extract data from a single sheet specified
   in \code{d$data$sheet}.
}
