% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackMechanismMap.R
\name{unPackMechanismMap}
\alias{unPackMechanismMap}
\title{Unpack Mechanism Maps for use in site distributions.}
\usage{
unPackMechanismMap(mechMap_path = NA)
}
\arguments{
\item{mechMap_path}{A local file path directing to a submitted Mechanism Map.
Default is NA. If left NA, will prompt user to select file via window.}
}
\value{
A dataframe containing mechanism mapping data ready for use as right
object in left-join with site-level data from DATIM.
}
\description{
Reads in an Excel file containing mechanism mapping from old to new
mechanisms for use in adjusting site distribution algorithms.
}
