% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackOPU_PSNUxIM.R
\name{unPackOPU_PSNUxIM}
\alias{unPackOPU_PSNUxIM}
\title{Unpack data from OPU Data Pack PSNUxIM tab.}
\usage{
unPackOPU_PSNUxIM(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Extracts data from updated targets in OPU Data Pack PSNUxIM tab.
}
