% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSchema.R
\name{unPackSchema}
\alias{unPackSchema}
\title{Extract and save schema from Data Pack template.}
\usage{
unPackSchema(filepath)
}
\arguments{
\item{path}{Local filepath for a Data Pack template (XLSX).}
}
\value{
Data Pack schema.
}
\description{
Supplied a filepath to a Data Pack template (XLSX), will extract and save a
schema based on the template.
}
