% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSiteTool.R
\name{unPackSiteTool}
\alias{unPackSiteTool}
\title{Unpacks a Site Tool}
\usage{
unPackSiteTool(d)
}
\arguments{
\item{d}{Datapackr object as passed from unPackTool.}
}
\value{
d
}
\description{
Processes a submitted Site Tool (in .xlsx format) by identifying integrity
    issues, checking data against DATIM validations, and extracting data.
}
