% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSiteTool.R
\name{unPackSiteToolData}
\alias{unPackSiteToolData}
\title{unPackSiteToolData(submission_path,output_path)}
\usage{
unPackSiteToolData(submission_path = NA, output_path = NA)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{output_path}{A local path directing to the folder where you would like
outputs to be saved. If not supplied, will output to working directory.}
}
\description{
Processes a submitted Site Tool (in .xlsx format) by identifying integrity
    issues, checking data against DATIM validations, and extracting data.
}
\details{
Executes the following operations in relation to a submitted Site Tool
\enumerate{
    \item Performs integrity checks on file structure;
    \item Runs DATIM validation tests;
}
}
