% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSheet.R
\name{unPackSiteToolSheet}
\alias{unPackSiteToolSheet}
\title{unPackSiteToolSheet(d, sheet)}
\usage{
unPackSiteToolSheet(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to unpack.}
}
\value{
d
}
\description{
Within a submitted Site Tool (directed to by
   \code{d$keychain$submission_path}), extract data from a single sheet specified
   in \code{sheet}.
}
