% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSiteTool.R
\name{unPackSiteToolSheets}
\alias{unPackSiteToolSheets}
\title{unPackSiteToolSheets(d)}
\usage{
unPackSiteToolSheets(d)
}
\arguments{
\item{d}{datapackr list object containing at least
\code{d$keychain$submission_path}.}
}
\value{
A datapackr list object, \code{d}, storing at least 2 dataframes of
   data extracted from submitted Data Pack or Site Tool: a \code{d$data$MER}
   dataframe containing all MER data to be distributed to site level, and/or
   \code{d$data$SUBNAT_IMPATT} containing data in the SUBNAT and IMPATT
   datasets from DATIM that can be imported into DATIM at the PSNU level.
}
\description{
Loops through all critical sheets in a submitted 
    Site Tool and executes \code{\link{unPackSheet}} to extract data, then
    compiles data into single flat dataframe. Also executes
    \code{\link{separateDataSets}} to separate single dataframe into at least
    two for \code{MER} and \code{SUBNAT/IMPATT}.
}
