% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackTool.R
\name{unPackTool}
\alias{unPackTool}
\title{Unpack a submitted tool}
\usage{
unPackTool(submission_path = NA, output_path = getwd(),
  tool = "Data Pack")
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{output_path}{A local path directing to the folder where you would like
outputs to be saved. If not supplied, will output to working directory.}

\item{tool}{What type of tool is the submission file? Default is "Data Pack".
Other options include "Site Tool", "Mechanism Map", and "Site Filter".}
}
\description{
Processes a submitted Data Pack, Site Tool, Mechanism Map, or Site Filter by
identifying integrity issues, checking data against DATIM validations, and
extracting data.
}
\details{
Executes the following operations in relation to a submitted Site Tool
\enumerate{
    \item Performs integrity checks on file structure;
}
}
