% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackToolSet.R
\name{unPackToolSet}
\alias{unPackToolSet}
\title{Title}
\usage{
unPackToolSet(
  d1_path = NULL,
  d2_path = NULL,
  country_uids = NULL,
  cop_year = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2_session}{}
}
\value{
Datapackr d object merged from a Datapack and standalone PSNUxIM tab
}
\description{
Title
}
