% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_de_coc_co_map.R
\name{update_de_coc_co_map}
\alias{update_de_coc_co_map}
\title{Generate a map from Data Pack indicators to Data Pack DEs, COCs, & COs}
\usage{
update_de_coc_co_map(
  cop_year = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
dp_map
}
\description{
For a given COP Year, produces a dataframe mapping from Data Pack
indicators to DATIM dataElements, categoryOptionCombos, & categoryOptions.
}
