% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{updated_indicator_codes}
\alias{updated_indicator_codes}
\title{Data Pack Updated Indicator Codes}
\format{
\describe{
  \item{indicator_code}{Old Data Pack indicator code for COP20.}
  \item{indicator_code_updated}{New Data Pack indicator code for COP21.}
  \item{finalized}{Binary of 0 or 1 indicating if the new indicator code
  has been finalized or may be subject to further revisions.}
}
}
\usage{
updated_indicator_codes
}
\description{
Data frame showing the relationship between COP20 and COP21
Data Pack indicator codes.
}
\keyword{datasets}
