% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateExistingPrioritization.R
\name{updateExistingPrioritization}
\alias{updateExistingPrioritization}
\title{Update Existing Prioritization}
\usage{
updateExistingPrioritization(prios, analytics_table)
}
\value{

}
\description{
Provided a data frame of priorization levels and an analytics
table, replace the existing prioritizations. This is normally
used with an OPU DataPack.
}
