% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{uploadDATIMExportToPDAP}
\alias{uploadDATIMExportToPDAP}
\title{Upload DATIM Export to PDAP}
\usage{
uploadDATIMExportToPDAP(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
Returns the response from the upload
}
\description{
Extracts current COP year data from the d object,
formats it, and uploads the data to PDAP.
}
