% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_PSNUs}
\alias{valid_PSNUs}
\title{List of valid PSNUs used for generating Data Packs.}
\format{\describe{
  \item{ou}{}
  \item{ou_id}{}
  \item{country_name}{}
  \item{country_uid}{}
  \item{snu1}{}
  \item{snu1_uid}{}
  \item{psnu}{}
  \item{psnu_uid}{}
  \item{psnu_type}{}
  \item{lastUpdated}{}
  \item{ancestors}{}
  \item{organisationUnitGroups}{}
}}
\usage{
valid_PSNUs
}
\description{
List of valid PSNUs used for generating Data Packs. Must be
synced and saved manually.
}
\keyword{datasets}
