% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginToDATIM.R
\name{ValidateConfig}
\alias{ValidateConfig}
\title{Check login credentials}
\usage{
ValidateConfig(dhis_config)
}
\arguments{
\item{dhis_config}{List of DATIM login credentials, including username, 
password, and login URL.}
}
\description{
Validates login credentials to make sure they've been provided correctly.
}
