% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationSummary.R
\name{validationSummary}
\alias{validationSummary}
\title{Validation Summary}
\usage{
validationSummary(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
Data frame consisting of ou, ou_id, country_name, country_uid, validation_issue_category, count
}
\description{
Produce an output of tests.
}
