% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validDPDisaggs.R
\name{validDPDisaggs}
\alias{validDPDisaggs}
\title{List Valid Data Pack Disaggs}
\usage{
validDPDisaggs()
}
\value{
Data list object of valid Data Pack disaggs, by sheet name.
}
\description{
Lists valid Data Pack disaggs
}
