% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSiteTool.R
\name{write_site_level_sheet}
\alias{write_site_level_sheet}
\title{Write Site Tool sheet}
\usage{
write_site_level_sheet(wb, sheet, d)
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{sheet}{Name of sheet to write.}

\item{d}{A datapackr list object.}
}
\description{
Writes an individual Site Tool sheet with OU-level summary row, subtotal
formulas, and site-level distributed dataset.
}
