% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{writeFxColumnwise}
\alias{writeFxColumnwise}
\title{Write character vector of Excel formulas into Openxlsx Workbook object
horizontally, rather than vertically.}
\usage{
writeFxColumnwise(wb, sheet, x, xy)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet.}

\item{sheet}{The worksheet to write to. Can be the worksheet index or name.}

\item{x}{A character vector of Excel formulas.}

\item{xy}{A vector of the form \code{c(start column, start row)}.}
}
\description{
Takes a character vector of Excel formulas (\code{x}) and writes these into
an Openxlsx Workbook object (\code{wb}) horizontally, beginning at the cell
position demarcated by \code{xy}. This function augments the existing
function in the Openxlsx package \code{\link[openxlsx]{writeFormula}} by
overcoming its strict ability to write formulas only rowwise.
}
