% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packFrame.R
\name{writeHomeTab}
\alias{writeHomeTab}
\title{writeHomeTab(wb, data_pack_uid, type = "Data Pack")}
\usage{
writeHomeTab(wb, data_pack_uid, type = "Data Pack")
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{data_pack_uid}{Unique ID designating which Data Pack the Home tab should
list.}

\item{type}{Either "Data Pack" or "Site Tool". Defaults to "Data Pack".}
}
\value{
Openxlsx workbook object with added, styled Home tab.
}
\description{
Function to write Home tab details into either Data Pack or Site Tool, as
specified.
}
