% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{writePDAPExportCSV}
\alias{writePDAPExportCSV}
\title{Title}
\usage{
writePDAPExportCSV(d, job_type)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{job_type}{The type of job to upload the data to. Currently only
target_setting_tool or year_two_targets are supported. An invalid job
type will throw an error.}
}
\value{
Returns the raw binary data of the CSV file
}
\description{
Given a DataPack object, this function will extract the necessary data
and format it for the PDAP API. The data will be written to a CSV file
and returned as raw binary data. The job type will determine the format
of the data. Currently, only target_setting_tool and year_two_targets
are supported.
}
