% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePSNUxIM.R
\name{writePSNUxIM}
\alias{writePSNUxIM}
\title{Write PSNUxIM Tab}
\usage{
writePSNUxIM(
  d,
  snuxim_model_data_path = NULL,
  output_folder = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{snuxim_model_data_path}{Filepath where SNU x IM distribution model is stored.}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export. If left as \code{NULL}, will output to
\code{Working Directory}.}

\item{d2_session}{DHIS2 Session id}
}
\value{
d
}
\description{
Checks a Data Pack for need of new or appended PSNUxIM data, then
writes this into the Data Pack supplied. unPackTool must be run as prerequisite.
}
