% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSpectrumData.R
\name{writeSpectrumData}
\alias{writeSpectrumData}
\title{Title writeSpectrumData}
\usage{
writeSpectrumData(wb, spectrum_data)
}
\arguments{
\item{wb}{Openxlsx worbook object}

\item{spectrum_data}{Spectrum data}
}
\value{
Modified openxlsx worbook object with Spectrum data written to the
Spectrum tab.
}
\description{
Utility function mostly for testing which will
write Spectrum data in a defined format to the Spectrum tab.
}
