% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemo.R
\name{zerosToDashes}
\alias{zerosToDashes}
\title{Zeroes to Dashes}
\usage{
zerosToDashes(x)
}
\arguments{
\item{x}{Target value to be formatted.}
}
\value{
Formatted target value
}
\description{
Zeroes to Dashes
}
