% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUs.R
\name{add_dp_label}
\alias{add_dp_label}
\title{Modify Data Pack Org Unit list to add datapackr IDs.}
\usage{
add_dp_label(orgunits, cop_year)
}
\arguments{
\item{orgunits}{Data frame of Data Pack org units produced by \code{\link{getDataPackOrgUnits}}.}

\item{cop_year}{COP Year. For COP years less than 2023,
the organisation unit type will be added to tbe DP label}
}
\value{
Data frame of Data Pack Org units with added Data Pack label, \code{dp_label}.
}
\description{
Adds Org Unit label used in Data Packs.
}
