% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUs.R
\name{add_dp_psnu}
\alias{add_dp_psnu}
\title{Modify PSNU list to add datapackr IDs.}
\usage{
add_dp_psnu(PSNUs, country_uids)
}
\arguments{
\item{PSNUs}{Data frame of PSNUs produced by \code{\link{getPSNUs}}.}

\item{country_uids}{Character vector of DATIM country IDs. This can only
include countries. Regional Operating Unit UIDs will not be accepted}
}
\value{
Data frame of PSNUs
}
\description{
Queries DATIM to extract list of PSNUs and adds additional PSNUs not
currently in DATIM as needed.
}
