% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_import_file.R
\name{adorn_import_file}
\alias{adorn_import_file}
\title{Convert PSNU-level DATIM import file into analytics-friendly object}
\usage{
adorn_import_file(
  psnu_import_file,
  cop_year = getCurrentCOPYear(),
  psnu_prioritizations = NULL,
  filter_rename_output = TRUE,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{psnu_import_file}{DHIS2 import file to convert}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{psnu_prioritizations}{List of orgUnit, value containing prioritization
values for each PSNU. If not included, blank prioritizations shown.}

\item{filter_rename_output}{T/F Should this function output the final data in
the new, more complete format?}

\item{d2_session}{DHIS2 Session id}
}
\value{
data
}
\description{
Convert PSNU-level DATIM import file into analytics-friendly
object, similar to the MER Structured Datasets
}
