% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_import_file.R
\name{adorn_import_file}
\alias{adorn_import_file}
\title{Convert a 'PSNU-level' DATIM import file into an analytics-friendly
 object.}
\usage{
adorn_import_file(
  psnu_import_file,
  cop_year = NULL,
  psnu_prioritizations = NULL,
  map_des_cocs = NULL,
  filter_rename_output = TRUE,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  include_default = FALSE
)
}
\arguments{
\item{psnu_import_file}{DHIS2 import file to convert}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{psnu_prioritizations}{List of orgUnit, value containing prioritization
values for each PSNU.}

\item{map_des_cocs}{Can be used to supply a specific data element/COC map.
If not specified, results will be obtained from default methods.}

\item{filter_rename_output}{Should this function output the final data in
the new, more complete format?}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}

\item{include_default}{Should default mechanisms be included?}
}
\value{
psnu_import_file
}
\description{
Convert a 'PSNU-level' DATIM import file into an
 analytics-friendly object, similar to the MER Structured Datasets
}
