% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{adornMechanisms}
\alias{adornMechanisms}
\title{Adorn Mechanisms}
\usage{
adornMechanisms(
  data,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{data}{Dataset to adorn, typically d$data$analytics}

\item{d2_session}{DHIS2 Session id}
}
\value{
Modified data object
}
\description{
Join analytical dimensions with d$data$analtyics related
to partner, agency and mechanism information.
}
