% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{adornMechanisms}
\alias{adornMechanisms}
\title{adornMechanisms(data)}
\usage{
adornMechanisms(data)
}
\arguments{
\item{data}{Dataset to adorn}
}
\value{
Modified data object
}
\description{
Join analytical dimensions with d$data$analtyics related
to partner, agency and mechanism information.
}
