% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornPSNUs.R
\name{adornPSNUs}
\alias{adornPSNUs}
\title{Adorn PSNUs}
\usage{
adornPSNUs(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
Modified d$data$distributedMER
object with a DATIM compatible data frame for import id d$datim$MER
}
\description{
Creates an analtyical dimension for planning prioritization from SUBNAT data
contained in the datapack
}
