% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnalytics.R
\name{analyze_pmtctknownpos}
\alias{analyze_pmtctknownpos}
\title{Check Data Pack data for PMTCT retention < 98\% or >100\%}
\usage{
analyze_pmtctknownpos(data)
}
\arguments{
\item{data}{Analytics object to analyze}
}
\value{
a
}
\description{
Check data gathered from Data Pack to identify cases where
retention of PMTCT patients is less than the standard of 98\% or >100\%.
}
