% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnalytics.R
\name{analyze_tbknownpos}
\alias{analyze_tbknownpos}
\title{Check Data Pack data for TB retention < 98\% or >100\%}
\usage{
analyze_tbknownpos(data)
}
\arguments{
\item{data}{Analytics object to analyze}
}
\value{
a
}
\description{
Check data gathered from Data Pack to identify cases where
retention of TB patients is less than the standard of 98\% or >100\%.
}
