% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R
\name{api_filter}
\alias{api_filter}
\title{Filter a DATIM API query.}
\usage{
api_filter(api_call, field, operation, match)
}
\arguments{
\item{field}{Endpoint field aginst which to filter.}

\item{operation}{Operation to apply as filter. See
\href{https://docs.dhis2.org/2.22/en/developer/html/ch01s08.html}{DHIS2 Web API documentation}
for valid operators.}

\item{match}{Text to match using \code{operator}.}
}
\value{
Web-encoded URL for DATIM API query.
}
\description{
Adds filter to DATIM API query and encodes for web.
}
