% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiUtilities.R
\name{api_sql_call}
\alias{api_sql_call}
\title{Query DATIM SQL View.}
\usage{
api_sql_call(
  sqlView,
  var = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{sqlView}{uid of sqlView table to query.}

\item{var}{Variable to substitute into SQL query. Only supply if SQL view is
of type query.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Web-encoded URL for DATIM API query.
}
\description{
Queries a DATIM SQL View and returns data object.
}
