% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handshakeFile.R
\name{canReadFile}
\alias{canReadFile}
\title{Test for file read access}
\usage{
canReadFile(path = NULL)
}
\arguments{
\item{path}{Filepath to test. Default is NA.}
}
\value{
Logical. \code{TRUE} if file can be read, \code{FALSE} if not.
}
\description{
Given a file path, will test whether the file can be read.
}
