% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnalytics.R
\name{checkAnalytics}
\alias{checkAnalytics}
\title{Check Data Pack data for analytics concerns}
\usage{
checkAnalytics(
  d,
  model_data_path,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{model_data_path}{Filepath to model data produced from DATIM.}

\item{d2_session}{DHIS2 Session id}
}
\value{
d
}
\description{
Check data gathered from Data Pack to identify
validation concerns at the PSNU level.
}
