% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAnalytics.R
\name{checkAnalytics}
\alias{checkAnalytics}
\title{Check Data Pack data for analytics concerns}
\usage{
checkAnalytics(d, model_data_path)
}
\arguments{
\item{d}{Datapackr object.}
}
\value{
d
}
\description{
Check data gathered from Data Pack to identify validation concerns
at the PSNU level.
}
