% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkColStructure.R
\name{checkColStructure}
\alias{checkColStructure}
\title{checkColStructure(d)}
\usage{
checkColStructure(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Checks structural integrity of columns on critical sheets for
   submitted Data Pack.
}
