% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataPack.R
\name{checkColStructure}
\alias{checkColStructure}
\title{checkColStructure(d)}
\usage{
checkColStructure(d)
}
\arguments{
\item{d}{datapackr list object containing at least d$data$extract,
d$data$sheet & d$info$warningMsg.}
}
\value{
A datapackr list object, \code{d}, storing a warning message of all
   issues related to Data Pack or Site Tool columns names or order.
}
\description{
Checks structural integrity of columns on critical sheets for
   submitted Data Pack or Site Tool.
}
