% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConnections.R
\name{checkConnections}
\alias{checkConnections}
\title{Detect potentially troublesome external connections.}
\usage{
checkConnections(d)
}
\arguments{
\item{d}{datapackr object}
}
\value{
Logical whether there exists a connection in a specified Excel file
}
\description{
Searches an Excel workbook for any external connections.
}
