% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDecimalValues.R
\name{checkDecimalValues}
\alias{checkDecimalValues}
\title{Check for invalid Decimal Values}
\usage{
checkDecimalValues(d, sheet, quiet = T)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}

\item{quiet}{Logical. Should warning messages be printed? Default is TRUE.}
}
\value{
d
}
\description{
Checks if there are invalid decimal values.
}
