% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDuplicateRows.R
\name{checkDuplicateRows}
\alias{checkDuplicateRows}
\title{checkDuplicateRows(d)}
\usage{
checkDuplicateRows(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check}
}
\value{
d
}
\description{
Checks data pulled from a single sheet in a Data Pack and
alerts where there are duplicate rows.
}
