% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSNUxIM.R
\name{checkHasPSNUxIM}
\alias{checkHasPSNUxIM}
\title{Title}
\usage{
checkHasPSNUxIM(d)
}
\value{
d object list with additional flags for PSNUxIM state.
}
\description{
Internal function to determine if a PSNUxIM tab exists
}
