% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInvalidOrgUnits.R
\name{checkInvalidOrgUnits}
\alias{checkInvalidOrgUnits}
\title{Check Invalid Org Units}
\usage{
checkInvalidOrgUnits(d, sheet)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheet}{Specified sheet within wb.}
}
\value{
d
}
\description{
Checks data pulled from a single sheet in a Data Pack and
alerts where there are unallowed org units based on current DATIM PSNU level.
}
