% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMissingMetadata.R
\name{checkMissingMetadata}
\alias{checkMissingMetadata}
\title{checkMissingMetadata(d)}
\usage{
checkMissingMetadata(d, sheet, quiet = T)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet to check.}

\item{quiet}{If true, function runs without displaying messages while processing.}
}
\value{
d
}
\description{
Checks data pulled from a single sheet in a Data Pack and
alerts where there are NAs instead of valid metadata.
}
