% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataPack.R
\name{checkOUinfo}
\alias{checkOUinfo}
\title{checkOUinfo(d)}
\usage{
checkOUinfo(d)
}
\arguments{
\item{d}{datapackr list object containing at least d$keychain$submission_path.}
}
\value{
A datapackr list object, \code{d}, storing a unique UID and Name for
   the PEPFAR Operating Unit related to the submitted Data Pack or Site Tool.
}
\description{
Cross-checks and updates PEPFAR Operating Unit name and id as
 read from Data Pack or Site Tool submission file.
}
