% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{checkPSNUxIMDisaggs}
\alias{checkPSNUxIMDisaggs}
\title{Title checkPSNUxIMDisaggs}
\usage{
checkPSNUxIMDisaggs(d)
}
\arguments{
\item{d}{Datapackr d object}
}
\value{
d Datapackr d object
}
\description{
Compares Age/Sex/KeyPop disaggs found in the PSNUxIM tab
with those found in the DE/COC map for a given year.
}
