% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStructure.R
\name{checkStructure}
\alias{checkStructure}
\title{Check tab structure of submitted tool.=}
\usage{
checkStructure(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Checks structural integrity of sheets for submitted tool.
}
