% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapackVsDatim.R
\name{compareData_DatapackVsDatim}
\alias{compareData_DatapackVsDatim}
\title{compareData_DatapackVsDatim}
\usage{
compareData_DatapackVsDatim(
  d,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  datim_data = NULL,
  datastreams = c("mer_targets", "subnat_targets", "impatt")
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}

\item{datim_data}{A data frame resulting from datimutils::getDataValueSets. If null, the data will be fetched
from DATIM.}

\item{datastreams}{Data stream or streams. One or more of \code{mer_targets},
\code{mer_results}, \code{subnat_targets}, \code{subnat_results}, or
\code{impatt}. If not specified, then all data streams
are returned.}
}
\value{
list object of diff result $psnu_x_im_wo_dedup, $psnu_w_dedup,
$updates (import to bring DATIM up to date with datapack), $deletes
(import to bring DATIM up to date with datapack)
}
\description{
Compares the data in a parsed data pack that would be destined for DATIM with target data in in DATIM.
}
