% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapackVsDatim.R
\name{compareData_OpuDatapackVsDatim}
\alias{compareData_OpuDatapackVsDatim}
\title{compareData_OpuDatapackVsDatim}
\usage{
compareData_OpuDatapackVsDatim(d, base_url = getOption("baseurl"))
}
\arguments{
\item{d}{list object - parsed data pack object}

\item{base_url}{string - base address of instance (text before api/ in URL)}
}
\value{
list object of diff result $psnu_x_im_wo_dedup, $psnu_w_dedup,
$updates (import to bring DATIM up to date with datapack), $deletes
(import to bring DATIM up to date with datapack)
}
\description{
Compares the data in a parsed data pack that would be destined for DATIM with target data in in DATIM.
}
