% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cop20_data_pack_schema}
\alias{cop20_data_pack_schema}
\title{Schema describing correct structure of COP20 Data Pack template.}
\format{
\describe{
  \item{sheet_num}{Lists the index value associated with the sheet name
  listed in \code{sheet_name}.}
  \item{sheet_name}{Lists the sheet/tab name as used in both the Data Pack.}
  \item{data_structure}{}
  \item{col}{Value describing the column position of each
  \code{indicator_code}.}
  \item{indicator_code}{Code used in the Data Pack to uniquely
  identify each distinct programmatic area of target setting.}
  \item{dataset}{For \code{indicator_codes} listed as "\code{Targets}"
  in the \code{col_type} field, documents the dataset, either \code{MER},
  \code{IMPATT}, or \code{SUBNAT}.}
  \item{col_type}{Flags whether an \code{indicator_code} is a Target
  (\code{"Target"}), row header (\code{Row Header}) or not (\code{NA}).}
  \item{value_type}{}
  \item{dataelement_dsd}{}
  \item{dataelement_ta}{}
  \item{categoryoption_specified}{}
  \item{valid_ages}{}
  \item{valid_sexes}{}
  \item{valid_kps}{}
  \item{formula}{Excel formula defined for \code{indicator_code}.}
}
}
\usage{
cop20_data_pack_schema
}
\description{
This schema describes the correct structure of a COP20 Data Pack
file, generated from the template used to produce Data Packs and useful in
validating Data Packs passed through datapackr.
}
\keyword{datasets}
