% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COP21_datapacks_countries}
\alias{COP21_datapacks_countries}
\title{Datapack country groupings}
\format{
\describe{
  \item{datapack_name}{Country name on home tab of datapack}
  \item{country_uids}{Country's UIDs listed on the home tab}
}
}
\usage{
COP21_datapacks_countries
}
\description{
Tibble of data pack country names and their UIDs.
}
\keyword{datasets}
