% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalytics.R
\name{createAnalytics}
\alias{createAnalytics}
\title{createAnalytics(d)}
\usage{
createAnalytics(d, d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d}{Datapackr object}

\item{d2_session}{R6 datimutils object which handles authentication with DATIM}
}
\value{
Modified d object with d$data$analytics
}
\description{
Wrapper function for creation of d$data$analytics object
which is suitable for export to external analytics sytems.
}
