% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalytics.R
\name{createAnalytics}
\alias{createAnalytics}
\title{createAnalytics(d)}
\usage{
createAnalytics(d)
}
\arguments{
\item{d}{}
}
\value{
Modified d object with d$data$analtyics
}
\description{
Wrapper function for creation of d$data$analtyics object 
which is suitable for export to external analytics sytems.
}
