% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKeychainInfo.R
\name{createKeychainInfo}
\alias{createKeychainInfo}
\title{Create Keychain Info for use in DataPack object.}
\usage{
createKeychainInfo(
  submission_path = NULL,
  tool = NULL,
  country_uids = NULL,
  cop_year = NULL
)
}
\value{
DataPack object
}
\description{
Creates Keychain info needed for use across many \code{datapackr} functions.
}
