% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestDataset.R
\name{createTestDataset}
\alias{createTestDataset}
\title{Create a test Targets dataset}
\usage{
createTestDataset(
  country_uids,
  cop_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{d2_session}{DHIS2 Session id}
}
\value{
wb with specified sheet packed with data
}
\description{
Compiles a test DATIM import file as would be output by a Data Pack.
}
