% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datapack_cogs}
\alias{datapack_cogs}
\title{Datapack Category option groups}
\format{
\describe{
  \item{name}{Name of the Category Option Group for example "01-04 Only"}
  \item{id}{Category Option Group UID}
}
}
\usage{
datapack_cogs
}
\description{
Data frame of category option groups (id and name)
along with their individual category options (id and name) as a
nested data frame.
}
\keyword{datasets}
