% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{datapackr_params}
\alias{datapackr_params}
\title{Standardized package function parameter definitions}
\usage{
datapackr_params(
  model_data,
  snuxim_model_data,
  SNUxIM,
  MER,
  PSNUxIM_combos,
  datapack_name,
  country_uids,
  template_path,
  submission_path,
  cached_mechs_path,
  cached_psnus_path,
  cop_year,
  output_folder,
  results_archive,
  d2_session,
  d,
  datastreams,
  include_mil,
  include_DREAMS,
  additional_fields,
  schema,
  wb,
  PSNUs,
  psnus,
  tool,
  season,
  sheet,
  draft_memo,
  memo_type,
  memo_doc,
  memo_structure,
  memoStructure,
  source_type,
  prios,
  include_no_prio,
  remove_empty_columns,
  ...
)
}
\arguments{
\item{model_data}{Data from DATIM needed to pack into a COP Data Pack.}

\item{snuxim_model_data}{Export from DATIM needed to allocate data across
mechanisms in the PSNUxIM tab}

\item{SNUxIM}{SNUxIM dataset extract from unPackSNUxIM}

\item{MER}{MER dataset extract from unPackSheets}

\item{PSNUxIM_combos}{Dataset extract from unPackSNUxIM that shows data
missing from the PSNUxIM tab.}

\item{datapack_name}{Name you would like associated with this Data Pack.
(Example: "Western Hemisphere", or "Caribbean Region", or "Kenya".)}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.}

\item{template_path}{Local filepath to Data Pack template Excel (XLSX) file.
This file MUST NOT have any data validation formats present. If left
\code{NULL}, will select the default based on \code{cop_year} and \code{tool}.}

\item{submission_path}{Local path to the file to import.}

\item{cached_mechs_path}{Local file path to an RDS file containing
a cached copy of the mechanisms SQL view.}

\item{cached_psnus_path}{Local file path to an RDS file containing
a cached copy of PSNUs from DATIM.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export.}

\item{results_archive}{If TRUE, will export compiled results of all tests and
processes to output_folder.}

\item{d2_session}{DHIS2 Session id}

\item{d}{Datapackr sidecar object}

\item{datastreams}{Data stream or streams. One or more of \code{mer_targets},
\code{mer_results}, \code{subnat_targets}, \code{subnat_results}, or
\code{impatt}. If not specified, then all data streams
are returned.}

\item{include_mil}{Logical. If \code{TRUE}, will also include _Military nodes
related to \code{country_uids}. Default is \code{TRUE}.}

\item{include_DREAMS}{If \code{TRUE} will also include DREAMS organisation units.}

\item{additional_fields}{Character string of any fields to return from DATIM
API other than those returned by default.}

\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{wb}{Openxlsx workbook object.}

\item{PSNUs}{Dataframe of PSNUs to use in this function, containing at least
\code{psnu_uid}.}

\item{psnus}{Dataframe of PSNUs to use in this function, containing at least
\code{psnu_uid}.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{season}{Either \code{COP} or \code{OPU}.}

\item{sheet}{Data Pack sheet.}

\item{draft_memo}{Boolean indicating whether the memo being written is a
draft or final memo.}

\item{memo_type}{memo_type One of the following:
datapack: Create the memo based on the data in the datapack or OPU datapack
datim: Create the memo based on data currently in DATIM
comparison: Create a comparison memo with data from both DATIM and datapack}

\item{memo_doc}{\code{Officer} document object containing
the target memo tables.}

\item{memo_structure}{Structure of the memo d$memo$structure}

\item{memoStructure}{Structure of the memo d$memo$structure}

\item{source_type}{Indicates whether the data for a COP Approval Memo table
should come from the Data Pack or from DATIM. Values can be either
\code{datapack} or \code{datim}.}

\item{prios}{Data frame of prioritization levels.}

\item{include_no_prio}{If TRUE, include \code{"No Prioritiation"}
as a column in the output.}

\item{remove_empty_columns}{Should empty columns be removed from memos?}

\item{...}{Additional arguments to pass.}
}
\value{
list of all paramaters of this constructor function
}
\description{
Standardized package function parameter definitions
}
\seealso{
Other parameter-helpers: 
\code{\link{parameter-checks}}
}
\concept{parameter-helpers}
