% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_approval_memo.R
\name{defaultMemoStyleHeader}
\alias{defaultMemoStyleHeader}
\title{Default Memo Header Style}
\usage{
defaultMemoStyleHeader()
}
\value{
Officer list object defining default style for memo table headers
}
\description{
Default Memo Header Style
}
