% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunctDisaggs.R
\name{defunctDisaggs}
\alias{defunctDisaggs}
\title{defunctDisaggs(d)}
\usage{
defunctDisaggs(d, sheet)
}
\arguments{
\item{d}{Datapackr object.}

\item{sheet}{Sheet name.}
}
\value{
d
}
\description{
Checks data extracted from a sheet in a submitted Data Pack
   identify cases where invalid Disaggregate combinations have
   been used.
}
