% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataPack.R
\name{defunctDisaggs}
\alias{defunctDisaggs}
\title{defunctDisaggs(d)}
\usage{
defunctDisaggs(d)
}
\arguments{
\item{d}{datapackr list object containing at least d$data$extract of data
from a single sheet in submission file.}
}
\value{
A datapackr list object, \code{d}, storing a dataframe of defunct
   disaggs used in a specific sheet.
}
\description{
Checks data extracted from a sheet in a submitted Data Pack or
   Site Tool to identify cases where invalid Disaggregate combinations have
   been used.
}
