% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunctDisaggs.R, R/parseDataPack.R
\name{defunctDisaggs}
\alias{defunctDisaggs}
\title{defunctDisaggs(d)}
\usage{
defunctDisaggs(d, type = "Data Pack")

defunctDisaggs(d, type = "Data Pack")
}
\arguments{
\item{d}{Datapackr object.}

\item{type}{Either "Data Pack" or "Site Tool".}

\item{sheet}{Sheet name.}

\item{d}{datapackr list object containing at least d$data$extract of data
from a single sheet in submission file.}
}
\value{
d

A datapackr list object, \code{d}, storing a dataframe of defunct
   disaggs used in a specific sheet.
}
\description{
Checks data extracted from a sheet in a submitted Data Pack or
   Site Tool to identify cases where invalid Disaggregate combinations have
   been used.

Checks data extracted from a sheet in a submitted Data Pack or
   Site Tool to identify cases where invalid Disaggregate combinations have
   been used.
}
