% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDistributedMERToDATIM.R
\name{exportDistributedDataToDATIM}
\alias{exportDistributedDataToDATIM}
\title{exportDistributedDataToDATIM(data)}
\usage{
exportDistributedDataToDATIM(d, keep_dedup = FALSE)
}
\arguments{
\item{d}{Datapackr object}

\item{keep_dedup}{T/F as to whether to include Dedupe values in export file.}
}
\value{
Modified d object with  a DATIM compatible data frame for import id d$datim$MER
}
\description{
Packs distributed MER data prepared from unPackSNUxIM for import to DATIM.
}
