% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{exportPackr}
\alias{exportPackr}
\title{Export output.}
\usage{
exportPackr(data, output_path, type, datapack_name)
}
\arguments{
\item{data}{Data object to export. Can be either a dataframe or an Openxlsx
Workbook object.}

\item{output_path}{A local path directing to the folder where you would like
outputs to be saved. If not supplied, will output to working directory.}

\item{type}{File prefix to be applied in output filename, as follows:
  \describe{
    \item{Site Tool}{Openxlsx Workbook object containing Site Tool.}
    \item{Data Pack}{Openxlsx Workbook object containing Data Pack.}
    \item{FAST Export}{Data frame containing FAST export data.}
    \item{SUBNAT IMPATT}{Data frame containing SUBNAT/IMPATT data.}
    \item{Results Archive}{List object containing results archive.}
}}

\item{datapack_name}{Country name or OU name to be listed in output filename.}
}
\description{
Exports a datapackr output to specified filepath.
}
