% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{exportPackr}
\alias{exportPackr}
\title{Export output.}
\usage{
exportPackr(data, output_path, type, datapack_name)
}
\arguments{
\item{data}{Data object to export. Can be either a dataframe or an Openxlsx
Workbook object.}

\item{output_path}{A local path directing to the folder where you would like
outputs to be saved. If not supplied, will output to working directory.}

\item{type}{File prefix to be applied in output filename.}

\item{datapack_name}{Country name or OU name to be listed in output filename.}
}
\description{
Exports a datapackr output to specified filepath.
}
