% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSubnatToDatim.R
\name{exportSubnatToDATIM}
\alias{exportSubnatToDATIM}
\title{exportSubnatToDATIM(d)}
\usage{
exportSubnatToDATIM(d, datim_map)
}
\arguments{
\item{d}{Datapackr object}

\item{datim_map}{datim map object for the cop year}
}
\value{
Datapackr d object
}
\description{
Takes the outputs of the \code{\link{unPackSheets}} function and
    adds  a dataframe containing SUBNAT and IMPATT data,
    \code{d$data$SUBNAT_IMPATT} into a standard DATIM import file.
}
