% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{extractDuplicateRows}
\alias{extractDuplicateRows}
\title{Title extractDuplicateRows}
\usage{
extractDuplicateRows(d, sheet = "PSNUxIM")
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
Modified d object with a list of duplicated rows
}
\description{
Identifies Duplicate rows in the PSNUxIM tab.
}
