% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{extractSNUxIMCombos}
\alias{extractSNUxIMCombos}
\title{Title extractSNUxIMCombos for a given datapack (d)
and a given PSNUxIM object (p) extract all of the missing
combos which exist in in the DataPack but do not exist
in the PSNUxIM tab. Note that a combo in this context actually refers
to a combination of PSNU, indicator_code, Age, Sex and KeyPop
so does not contain any IM identifier.}
\usage{
extractSNUxIMCombos(d, p = NULL)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{p}{An optional PSNUxIM object.}
}
\value{
Modified d object with documented and missing SNUxIM combos
}
\description{
Title extractSNUxIMCombos for a given datapack (d)
and a given PSNUxIM object (p) extract all of the missing
combos which exist in in the DataPack but do not exist
in the PSNUxIM tab. Note that a combo in this context actually refers
to a combination of PSNU, indicator_code, Age, Sex and KeyPop
so does not contain any IM identifier.
}
