% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTForward.R
\name{FASTforward}
\alias{FASTforward}
\title{FASTforward(d)}
\usage{
FASTforward(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Takes the output of the \code{\link{unPackSheets}} function and
    recompiles this to contain only data relevant for and in same structure
    as the PEPFAR FAST Tool.
}
