% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTForward.R, R/parseDataPack.R
\name{FASTforward}
\alias{FASTforward}
\title{FASTforward(d)}
\usage{
FASTforward(d)

FASTforward(d)
}
\arguments{
\item{d}{Datapackr object}

\item{d}{datapackr list object containing at least
\code{d$data$distributedMER}.}
}
\value{
d

A datapackr list object, \code{d}, storing at least 1 dataframe of
   data structured for ingestion into the PEPFAR FAST Tool,
   \code{d$data$FAST}.
}
\description{
Takes the output of the \code{\link{unPackSheets}} function and
    recompiles this to contain only data relevant for and in same structure
    as the PEPFAR FAST Tool.

Takes the output of the \code{\link{unPackSheets}} function and
    recompiles this to contain only data relevant for and in same structure
    as the PEPFAR FAST Tool.
}
