% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{fetchMechsViewFromAPI}
\alias{fetchMechsViewFromAPI}
\title{fetchMechsViewFromAPI}
\usage{
fetchMechsViewFromAPI(
  operating_units = NULL,
  cop_year = NULL,
  uids = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{Numeric value of COP Fiscal Year to filter mechanism list by.
Ex: For mechanisms active in FY 2020, pertaining to COP 2019, enter
\code{2019}. If a FY is not supplied, returns entire mechanism list.}

\item{uids}{Character vector of uids to filter list by.}

\item{d2_session}{}
}
\value{
Mechanism List
}
\description{
fetchMechsViewFromAPI
}
