% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPrioritizationTable.R
\name{fetchPrioritizationTable}
\alias{fetchPrioritizationTable}
\title{Title}
\usage{
fetchPrioritizationTable(psnus, cop_year, d2_session)
}
\arguments{
\item{psnus}{A list of PSNUs}

\item{cop_year}{The cop year}

\item{d2_session}{Datimutils d2_session object}
}
\value{
A data frame consisting of Organisation unit, Value, prioritization
}
\description{
Title
}
