% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemoDOCX.R
\name{generateApprovalMemo}
\alias{generateApprovalMemo}
\title{Title}
\usage{
generateApprovalMemo(d, draft = TRUE)
}
\arguments{
\item{d}{}

\item{draft}{Should a draft watermark be added to the document?}
}
\value{
Returns a COP Memo DOCX object suitable for download
}
\description{
Title
}
