% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemo.R
\name{generateApprovalMemo}
\alias{generateApprovalMemo}
\title{Generate COP Approval Memo Target Tables (DOCX Format)}
\usage{
generateApprovalMemo(
  d,
  memo_type,
  draft_memo = TRUE,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{memo_type}{memo_type One of the following:
datapack: Create the memo based on the data in the datapack or OPU datapack
datim: Create the memo based on data currently in DATIM
comparison: Create a comparison memo with data from both DATIM and datapack}

\item{draft_memo}{Boolean indicating whether the memo being written is a
draft or final memo.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
Returns a COP Memo DOCX object suitable for download
}
\description{
Generate COP Approval Memo Target Tables (DOCX Format)
}
