% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemoDOCX.R
\name{generateApprovalMemo}
\alias{generateApprovalMemo}
\title{Generate COP Approval Memo Target Tables (DOCX Format)}
\usage{
generateApprovalMemo(
  d,
  d2_sesssion,
  memo_type,
  draft = TRUE,
  remove_empty_columns = TRUE,
  include_no_prio = TRUE
)
}
\arguments{
\item{d}{}

\item{draft}{Should a draft watermark be added to the document?}

\item{remove_empty_columns}{Should empty columns be removed?}
}
\value{
Returns a COP Memo DOCX object suitable for download
}
\description{
Generate COP Approval Memo Target Tables (DOCX Format)
}
