% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOPDataFromDATIM.R
\name{getCOPDataFromDATIM}
\alias{getCOPDataFromDATIM}
\title{Grab all raw data in DATIM for a country for the COP data sets for a given COP Year.}
\usage{
getCOPDataFromDATIM(
  country_uids,
  cop_year = getCurrentCOPYear(),
  base_url = getOption("baseurl")
)
}
\arguments{
\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{base_url}{base to append api calls normally defined beforehand}

\item{country_uid}{country_uid}
}
\value{
Raw data in DATIM for a country for the COP data sets for a given COP Year.
}
\description{
Grab all raw data in DATIM for a country for the COP data sets for a given COP Year.
}
