% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOPDataFromDATIM.R
\name{getCOPDataFromDATIM}
\alias{getCOPDataFromDATIM}
\title{Grab all raw data in DATIM for a country for the COP data sets for a given COP Year.}
\usage{
getCOPDataFromDATIM(
  country_uids,
  cop_year,
  streams = c("mer_targets", "subnat_targets", "impatt"),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{d2_session}{DHIS2 Session id}

\item{country_uid}{country_uid}
}
\value{
Raw data in DATIM for a country for the COP data sets for a given COP Year.
}
\description{
Grab all raw data in DATIM for a country for the COP data sets for a given COP Year.
}
\examples{
\dontrun{getCOPDataFromDATIM(country_uid = d$info$country_uids, cop_year = d$info$cop_year)}

}
