% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCopDataFromDatim}
\alias{getCopDataFromDatim}
\title{grab all raw data in DATIM for a country for the COP data sets for a given fiscal year.}
\usage{
getCopDataFromDatim(
  country_uid,
  fiscal_year_yyyy,
  base_url = getOption("baseurl")
)
}
\arguments{
\item{country_uid}{country_uid}

\item{fiscal_year_yyyy}{fiscal_year}

\item{base_url}{base to append api calls normally defined beforehand}
}
\value{
raw data in DATIM for a country for the COP data sets for a given fiscal year.
}
\description{
grab all raw data in DATIM for a country for the COP data sets for a given fiscal year.
}
