% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOPDataFromDATIM.R
\name{getCOPDataFromDATIM}
\alias{getCOPDataFromDATIM}
\title{Grab all raw data in DATIM for a country for the COP data sets for a given COP Year.}
\usage{
getCOPDataFromDATIM(
  country_uids,
  cop_year,
  streams = c("mer_targets", "subnat_targets", "impatt"),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{streams}{data stream or streams. One or more of "mer_targets", "subnat_targets", "impatt".
If not specified, then all data streams are returned.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Raw data in DATIM for a country for the COP data sets for a given COP Year.
}
\description{
Grab all raw data in DATIM for a country for the COP data sets for a given COP Year.
}
\examples{
\dontrun{getCOPDataFromDATIM(country_uid = d$info$country_uids, cop_year = d$info$cop_year)}

}
