% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOPDataFromDATIM.R
\name{getCOPDataFromDATIM}
\alias{getCOPDataFromDATIM}
\title{Grab all raw data in DATIM for a country for the COP data sets for a given COP Year.}
\usage{
getCOPDataFromDATIM(
  country_uids,
  cop_year,
  datastreams = c("mer_targets", "subnat_targets", "impatt"),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{Unique IDs for countries to include in the Data Pack.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{datastreams}{Data stream or streams. One or more of \code{mer_targets},
\code{mer_results}, \code{subnat_targets}, \code{subnat_results}, or
\code{impatt}. If not specified, then all data streams
are returned.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Raw data in DATIM for a country for the COP data sets for a given COP Year.
}
\description{
Grab all raw data in DATIM for a country for the COP data sets for a given COP Year.
}
\examples{
\dontrun{getCOPDataFromDATIM(country_uid = d$info$country_uids, cop_year = d$info$cop_year)}

}
