% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCOPDatasetUids}
\alias{getCOPDatasetUids}
\title{getCOPDatasetUids}
\usage{
getCOPDatasetUids(cop_year, datastreams)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{datastreams}{Data stream or streams. One or more of \code{mer_targets},
\code{mer_results}, \code{subnat_targets}, \code{subnat_results}, or
\code{impatt}. If not specified, then all data streams
are returned.}

\item{datastream}{character vector - one or more of:
{"mer_targets", "mer_results", "subnat_targets", "subnat_results", "impatt"}}
}
\value{
returns a character vector of the related dataset uids
}
\description{
returns character vector of dataset uids for a given FY:
{"2019", "2020", ... , "2023"}
and type {"mer_targets", "mer_results", "subnat_targets", "subnat_results",
"impatt"}
}
