% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCopDatasetUids}
\alias{getCopDatasetUids}
\title{Pull & combine all  UIDS for specified data_stream types and FY.}
\usage{
getCopDatasetUids(FY = NULL, types = NULL)
}
\arguments{
\item{FY}{Reporting FY for which to filter active code lists, numeric, if null current year is used.}

\item{types}{Specify MER, SUBNAT, or IMPATT, or omit to specify all.}
}
\value{
a character vector with the requested uids respecting the selection in the type parameter and FY parameter.
}
\description{
Pulls all uids for types argument for a specified FY.
}
