% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCOPDataFromDATIM.R
\name{getCOPDatasetUIDs}
\alias{getCOPDatasetUIDs}
\title{Pull & combine all  UIDS for specified data_stream types and FY.}
\usage{
getCOPDatasetUIDs(
  cop_year = getCurrentCOPYear(),
  types = c("MER", "SUBNAT", "IMPATT")
)
}
\arguments{
\item{cop_year}{Specifies COP year. Remember, COP20 = FY21 targets.}

\item{types}{Specify MER, SUBNAT, or IMPATT, or omit to specify all.}
}
\value{
a character vector with the requested uids respecting the selection in the type parameter and FY parameter.
}
\description{
Pulls all uids for types argument for a specified FY.
}
