% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCountries}
\alias{getCountries}
\title{Pull list of Countries from DATIM.}
\usage{
getCountries(
  datapack_uid = NA,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{datapack_uid}{A unique ID specifying the PEPFAR Operating Unit or
specific Data Pack country grouping. If left unspecified, will pull all
Country Names.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Data frame of Countries
}
\description{
Queries DATIM to extract list of Countries for specified Data Pack UID and
adds additional Countries not currently in DATIM as needed.
}
