% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDatasetUids}
\alias{getDatasetUids}
\title{Get Dataset UIDs}
\usage{
getDatasetUids(
  fiscal_year,
  type = c("mer_targets", "mer_results", "subnat_targets", "subnat_results", "impatt")
)
}
\arguments{
\item{fiscal_year}{character - one of: {"2019", "2020", "2021", "2022",
"2023"}}

\item{type}{character vector - one or more of:
{"mer_targets", "mer_results", "subnat_targets", "subnat_results", "impatt"}}
}
\value{
returns a character vector of the related dataset uids
}
\description{
returns character vector of dataset uids for a given FY:
{"2019", "2020", ... , "2023"}
and type {"mer_targets", "mer_results", "subnat_targets", "subnat_results",
"impatt"}
}
