% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDatasetUids}
\alias{getDatasetUids}
\title{getDatasetUids}
\usage{
getDatasetUids(fiscal_year, type = c("targets", "results", "subnat", "impatt"))
}
\arguments{
\item{fiscal_year}{character - one of {"2019", "2020", "2021"}}

\item{type}{character vector - one or more of {"targets", "results", "subnat_impatt"}}
}
\value{
returns a character vector of the related dataset uids
}
\description{
returns character vector of dataset uids for a given FY {"2019", "2020", "2021"}
and type {"targets", "results","subnat_impatt"}
}
