% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataValue.R
\name{GetDataValueSet}
\alias{GetDataValueSet}
\title{GetDataValueSet}
\usage{
GetDataValueSet(keys, values, base_url = getOption("baseurl"),
  api_version = "30", max_attempts = 3)
}
\arguments{
\item{keys}{character vector - data value set parameter keys (e.g. "dataSet", "period")}

\item{values}{character vector - values marching the key from keys (e.g. "Abcde123456", "2019Q1"}

\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{api_version}{- api version to use when calling DHIS2}

\item{max_attempts}{- number of times to try for a valid response}
}
\value{
tibble with the data requested
}
\description{
Used to read DHIS 2 data using the data value set endpoint
}
