% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataValueSets.R
\name{getDataValueSets}
\alias{getDataValueSets}
\title{getDataValueSets}
\usage{
getDataValueSets(
  keys,
  values,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{keys}{character vector - data value set parameter keys (e.g. "dataSet", "period")}

\item{values}{character vector - values marching the key from keys (e.g. "Abcde123456", "2019Q1"}

\item{d2_session}{R6 datimutils object which handles authentication with DATIM}
}
\value{
tibble with the data requested
}
\description{
Used to read DHIS 2 data using the data value set endpoint
}
