% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalytics.R
\name{getFY22Prioritizations}
\alias{getFY22Prioritizations}
\title{getFY22Prioritizations(d)}
\usage{
getFY22Prioritizations(
  d,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id}
}
\value{
Dataframe of psnu_uid and value
}
\description{
getFY22Prioritizations(d)
}
