% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalytics.R
\name{getFY22Prioritizations}
\alias{getFY22Prioritizations}
\title{getFY22Prioritizations(d)}
\usage{
getFY22Prioritizations(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
Dataframe of psnu_uid and value
}
\description{
getFY22Prioritizations(d)
}
