% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackPSNUxIM_API.R
\name{getHeaderColumns}
\alias{getHeaderColumns}
\title{getHeaderColumns returns a vector of header columns
provided an input vector of columns to keep.}
\usage{
getHeaderColumns(cols_to_keep, sheet = "PSNUxIM")
}
\arguments{
\item{sheet}{}
}
\value{
A vector of header columns.
}
\description{
getHeaderColumns returns a vector of header columns
provided an input vector of columns to keep.
}
