% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullHTSModality.R
\name{getHTSModality}
\alias{getHTSModality}
\title{getHTSModality}
\usage{
getHTSModality(
  cop_year = getCurrentCOPYear(),
  dataElements = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{dataElements}{List of dataElements to filter against.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Dataframe of HTS modalities mapped to dataElements
}
\description{
Map HTS modality to dataElement id
}
