% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHTSModality.R
\name{getHTSModality}
\alias{getHTSModality}
\title{Get HTS Modalities map}
\usage{
getHTSModality(
  cop_year = getCurrentCOPYear(),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
Dataframe of HTS modalities mapped to dataElements
}
\description{
Map HTS modality to dataElement id
}
