% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getIMPATTLevels}
\alias{getIMPATTLevels}
\title{Pull IMPATT levels from DATIM for all PEPFAR countries}
\usage{
getIMPATTLevels(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{DHIS session id.}
}
\value{
Dataframe of country metadata, including prioritization, planning,
country, community, and facility levels in DATIM organization hierarchy.
}
\description{
Queries DATIM to retrieve the latest version of
\code{/api/dataStore/dataSetAssignments/ous}
}
