% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMapDataPack_DATIM_DEs_COCs}
\alias{getMapDataPack_DATIM_DEs_COCs}
\title{get_Map_DataPack_DATIM_DEs_COCs}
\usage{
getMapDataPack_DATIM_DEs_COCs(cop_year)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
{cop20, cop21}_map_DataPack_DATIM_DEs_COCs
}
\description{
get_Map_DataPack_DATIM_DEs_COCs
}
