% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getMapDataPack_DATIM_DEs_COCs}
\alias{getMapDataPack_DATIM_DEs_COCs}
\title{get_Map_DataPack_DATIM_DEs_COCs}
\usage{
getMapDataPack_DATIM_DEs_COCs(cop_year, datasource = NULL)
}
\arguments{
\item{cop_year}{cop year to pull get map for}

\item{datasource}{Type of datasource (Data Pack, OPU Data Pack, DATIM)}
}
\value{
{cop21, cop22}_map_DataPack_DATIM_DEs_COCs
}
\description{
get_Map_DataPack_DATIM_DEs_COCs
}
