% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{getMechanismView}
\alias{getMechanismView}
\title{getMechanismView(d2_session, support_files_path)}
\usage{
getMechanismView(
  country_uids = NULL,
  cop_year = NULL,
  include_dedupe = FALSE,
  include_MOH = FALSE,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  cached_mechs_path = paste0(Sys.getenv("support_files_directory"), "mechs.rds")
)
}
\arguments{
\item{country_uids}{Character vector of DATIM country IDs. This can only
include countries. Regional Operating Unit uids will not be accepted. If not
supplied, returns entire mechanism list, trimmed to user's DATIM permissions.}

\item{cop_year}{Numeric value of COP Fiscal Year to filter mechanism list by.
Ex: For mechanisms active in FY 2020, pertaining to COP 2019, enter
\code{2019}. If a FY is not supplied, returns entire mechanism list.}

\item{include_dedupe}{Logical. If TRUE will include deduplication mechanisms.
Default is FALSE.}

\item{include_MOH}{Logical. If TRUE will include MOH mechanisms. Default is
FALSE.}

\item{d2_session}{datimutils d2Session object}

\item{cached_mechs_path}{Local file path to the cached mechanisms file.}
}
\value{
Mechs
}
\description{
Retrieves a view of mechanisms with partners and agencies
The function will attempt to read from a cached file, if defined and accessible.
Otherwise, if the user is logged in, the view
will be obtained from DATIM. Otherwise, an empty dataframe is returned.
}
