% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{getMechanismView}
\alias{getMechanismView}
\title{getMechanismView}
\usage{
getMechanismView(
  country_uids = NULL,
  cop_year = NULL,
  include_dedupe = FALSE,
  include_MOH = FALSE,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  cached_mechs_path = paste0(Sys.getenv("support_files_directory"), "mechs.rds"),
  update_stale_cache = FALSE
)
}
\arguments{
\item{country_uids}{Character vector of DATIM country IDs. This can only
include countries. Regional Operating Unit uids will not be accepted. If not
supplied, returns entire mechanism list, trimmed to user's DATIM permissions.}

\item{cop_year}{Numeric value of COP Fiscal Year to filter mechanism list by.
Ex: For mechanisms active in FY 2020, pertaining to COP 2019, enter
\code{2019}. If a FY is not supplied, returns entire mechanism list.}

\item{include_dedupe}{Logical. If TRUE will include deduplication mechanisms.
Default is FALSE.}

\item{include_MOH}{Logical. If TRUE will include MOH mechanisms. Default is
FALSE.}

\item{d2_session}{DHIS2 Session id}

\item{cached_mechs_path}{Filepath to an RDS file containing a cached copy of the
SQL view used defined via a envionment variable.}

\item{update_stale_cache}{If the cached_mechs_path file is outdated or unreadable,
should a new cache be saved?}
}
\value{
Mechs
}
\description{
Retrieves a view of mechanisms with partners and agencies
The function will attempt to read from a cached file, if defined in
the support_files_directory option has been set, and the mechs.rds file
is available to be read. Otherwise, if the user is logged in, the view
will be obtained from DATIM. Otherwise, an empty dataframe is returned.
}
