% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{getMechanismViewFromDATIM}
\alias{getMechanismViewFromDATIM}
\title{Title}
\usage{
getMechanismViewFromDATIM(
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2_session}{}
}
\value{

}
\description{
Title
}
