% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adornMechanisms.R
\name{getMechanismViewFromDATIM}
\alias{getMechanismViewFromDATIM}
\title{getMechanismViewFromDATIM}
\usage{
getMechanismViewFromDATIM(
  operating_units = NULL,
  cop_year = NULL,
  uids = NULL,
  include_dedupe = FALSE,
  include_MOH = FALSE,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{Numeric value of COP Fiscal Year to filter mechanism list by.
Ex: For mechanisms active in FY 2020, pertaining to COP 2019, enter
\code{2019}. If a FY is not supplied, returns entire mechanism list.}

\item{uids}{Character vector of uids to filter list by.}

\item{include_dedupe}{Logical. If TRUE will include deduplication mechanisms.
Default is FALSE.}

\item{include_MOH}{Logical. If TRUE will include MOH mechanisms. Default is
FALSE.}

\item{d2_session}{}
}
\value{
Mechanism List
}
\description{
getMechanismViewFromDATIM
}
