% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMemoIndicators.R
\name{getMemoIndicators}
\alias{getMemoIndicators}
\title{Title}
\usage{
getMemoIndicators(
  cop_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{COP Year}

\item{d2_session}{Datimutils session object}
}
\value{
A dataframe consisting of indicator UIDs,name,numerator and denominator expression.
If a given COP year does not exist, NULL is returned
}
\description{
Title
}
