% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumeratorDenominator.R
\name{getNumeratorDenominator}
\alias{getNumeratorDenominator}
\title{getNumeratorDenominator}
\usage{
getNumeratorDenominator(
  dataElements = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{dataElements}{List of dataElements to filter against. (Optional)}
}
\value{
Dataframe of Numerator and Denominators mapped to dataElements
}
\description{
Map Numerator and Denominator to dataElement id
}
