% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPUDataFromDATIM.R
\name{getOPUDataFromDATIM}
\alias{getOPUDataFromDATIM}
\title{Pull data for packing OPU Data Pack}
\usage{
getOPUDataFromDATIM(
  cop_year,
  country_uids = NULL,
  country_names = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Returns dataset needed for use in \code{\link{packOPUDataPack}}.
}
\description{
Pulls data from DATIM API needed to pack OPU Data Pack.
}
