% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOPUDataFromDATIM.R
\name{getOPUDataFromDATIM}
\alias{getOPUDataFromDATIM}
\title{Pull data for packing OPU Data Pack}
\usage{
getOPUDataFromDATIM(
  cop_year,
  country_uids = NULL,
  country_names = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{cop_year}{COP Year. Remember, COP22 = FY23 targets.}

\item{country_uids}{DATIM UIDs of Countries to return data for. Can supply either
this or Country Names.}

\item{country_names}{Names of Countries to return data for. Can supply either this
or DATIM Country UIDs.}

\item{d2_session}{R6 datimutils object which handles authentication with DATIM}
}
\value{
Returns dataset needed for use in \code{\link{packOPUDataPack}}.
}
\description{
Pulls data from DATIM API needed to pack OPU Data Pack.
}
