% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalytics.R
\name{getPrioritizations}
\alias{getPrioritizations}
\title{getPrioritizations(d)}
\usage{
getPrioritizations(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
Dataframe of psnu_uid and value
}
\description{
getPrioritizations(d)
}
