% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUs.R
\name{getPSNUs}
\alias{getPSNUs}
\title{Pull list of PSNUs from DATIM and format based on datapackr structure.}
\usage{
getPSNUs(country_uids = NA, include_mil = TRUE)
}
\arguments{
\item{country_uids}{Character vector of DATIM country IDs. This can only
include countries. Regional Operating Unit uids will not be accepted}

\item{include_mil}{Logical. If \code{TRUE}, will also include _Military nodes
related to \code{country_uids}. Default is \code{TRUE}.}
}
\value{
Data frame of PSNUs
}
\description{
Queries DATIM to extract list of PSNUs for specified Data Pack UID and adds
additional PSNUs not currently in DATIM as needed.
}
