% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUs.R
\name{getPSNUs}
\alias{getPSNUs}
\title{Pull list of PSNUs from DATIM based on provided country UIDs.}
\usage{
getPSNUs(
  country_uids = NULL,
  include_mil = TRUE,
  include_DREAMS = TRUE,
  additional_fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\value{
Data frame of PSNUs
}
\description{
Queries DATIM to extract list of PSNUs and adds additional PSNUs not
currently in DATIM as needed.
}
