% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPSNUs}
\alias{getPSNUs}
\title{Pull list of PSNUs from DATIM and format based on datapackr structure.}
\usage{
getPSNUs(country_uids = NA)
}
\arguments{
\item{country_uids}{Character vector of DATIM country IDs. This can only
include countries. Regional Operating Unit uids will not be accepted}
}
\value{
Data frame of PSNUs
}
\description{
Queries DATIM to extract list of PSNUs for specified Data Pack UID and adds
additional PSNUs not currently in DATIM as needed.
}
