% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSNUs.R
\name{getPSNUs}
\alias{getPSNUs}
\title{Pull list of PSNUs from DATIM based on provided country UIDs.}
\usage{
getPSNUs(
  country_uids = NULL,
  include_mil = TRUE,
  include_DREAMS = TRUE,
  additional_fields = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{include_mil}{Logical. If \code{TRUE}, will also include _Military nodes
related to \code{country_uids}. Default is \code{TRUE}.}

\item{include_DREAMS}{If \code{TRUE} will also include DREAMS
organisation units.}

\item{additional_fields}{Character string of any fields to return from DATIM
API other than those returned by default: \code{name}, \code{id}, \code{ancestors},
& \code{organisationUnitGroups}.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Data frame of PSNUs
}
\description{
Queries DATIM to extract list of PSNUs and adds additional PSNUs not
currently in DATIM as needed.
}
