% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteList.R
\name{getSiteList}
\alias{getSiteList}
\title{Compile Site List for Site Tool}
\usage{
getSiteList(country_uids, include_mil = TRUE)
}
\arguments{
\item{country_uids}{Character vector of DATIM country IDs. This can only
include countries. Regional Operating Unit uids will not be accepted}

\item{include_mil}{When TRUE, will append _Military organisation nodes to
site list.}
}
\description{
Extracts full list of sites from DATIM, including _Military nodes where
specified, and compiles these into the list as needed for use in the PEPFAR
Site Tool.
}
