% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullTechArea.R
\name{getTechArea}
\alias{getTechArea}
\title{getTechArea}
\usage{
getTechArea(dataElements = NULL)
}
\arguments{
\item{dataElements}{List of dataElements to filter against. (Optional)}
}
\value{
Dataframe of Tech Areas mapped to dataElements
}
\description{
Map Tech Area to dataElement id
}
