% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidCategoryOptions.R
\name{getValidCategoryOptions}
\alias{getValidCategoryOptions}
\title{getValidCategoryOptions}
\usage{
getValidCategoryOptions(cop_year = getCurrentCOPYear())
}
\arguments{
\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}
}
\value{
Dataframe of all valid categoryOptions for specified FY
}
\description{
Returns full list of all valid categoryOptions for specified FY.
}
