% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handshakeFile.R
\name{handshakeFile}
\alias{handshakeFile}
\title{Test an input file for read access and type. Prompt if issues.}
\usage{
handshakeFile(path = NA, tool = "Data Pack")
}
\arguments{
\item{path}{Filepath to test and use.}

\item{tool}{What type of tool is the submission file? Default is "Data Pack".
Other options include "Site Tool", "Mechanism Map", and "Site Filter".}

\item{extension}{File extension to test for. (Do not include leading period.)}
}
\value{
Character vector containing valid filepath for further use.
}
\description{
When supplied a filepath, will test that the file is readable and is of the
correct filetype. If either readability or file type is invalid, will prompt
for user selection of correct filepath via computer window.
}
