% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handshakeFile.R
\name{handshakeFile}
\alias{handshakeFile}
\title{Test an input file for read access and type. Prompt if issues.}
\usage{
handshakeFile(path = NULL, tool = NULL)
}
\arguments{
\item{path}{Filepath to test and use.}

\item{tool}{What type of tool is the submission file? Options include 
"Data Pack", "Data Pack Template", and "OPU Data Pack Template".}
}
\value{
Character vector containing valid filepath for further use.
}
\description{
When supplied a filepath, will test that the file is readable and is of the
correct filetype. If either readability or file type is invalid, will prompt
for user selection of correct filepath via computer window.
}
