% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{headerRow}
\alias{headerRow}
\title{Tool header rows}
\usage{
headerRow(tool, cop_year = getCurrentCOPYear())
}
\arguments{
\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
Header row
}
\description{
Tool header rows
}
