% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{headerRow}
\alias{headerRow}
\title{Tool to assist with formatting the header rows}
\usage{
headerRow(tool, cop_year)
}
\arguments{
\item{tool}{"Data Pack", "Data Pack Template",""OPU Data Pack".}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}
}
\value{
Header row
}
\description{
Tool to assist with formatting the header rows
}
