% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{headerRow}
\alias{headerRow}
\title{Tool header rows}
\usage{
headerRow(tool, cop_year = getCurrentCOPYear())
}
\arguments{
\item{tool}{"Data Pack", "Data Pack Template", "Site Tool", "Site Tool Template".}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}
}
\value{
Header row
}
\description{
Tool header rows
}
