% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isLoggedIn}
\alias{isLoggedIn}
\title{Determine whether the user is logged into DATIM or not.}
\usage{
isLoggedIn(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{R6 datimutils object which handles authentication with DATIM}
}
\value{
TRUE or FALSE
}
\description{
Determine whether the user is logged into DATIM or not.
}
