% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_COCs_to_COs.R
\name{map_COCs_to_COs}
\alias{map_COCs_to_COs}
\title{Map categoryoptioncombos to underlying categoryoptions.}
\usage{
map_COCs_to_COs(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{R6 datimutils object which handles authentication with DATIM}
}
\value{
dataframe of categoryoptioncombos mapped to underlying category
options
}
\description{
maps categoryoptioncombos to underlying categoryoptions.
}
