% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_COCs_to_COs.R
\name{map_COCs_to_COs}
\alias{map_COCs_to_COs}
\title{Map categoryoptioncombos to underlying categoryoptions.}
\usage{
map_COCs_to_COs()
}
\value{
dataframe of categoryoptioncombos mapped to underlying category
options
}
\description{
maps categoryoptioncombos to underlying categoryoptions.
}
