% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_Cs_to_COs.R
\name{map_Cs_to_COs}
\alias{map_Cs_to_COs}
\title{Map categoryoptioncombos to underlying categoryoptions.}
\usage{
map_Cs_to_COs(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{DHIS2 Session id}
}
\value{
dataframe of categoryoptioncombos mapped to underlying category
options
}
\description{
maps categoryoptioncombos to underlying categoryoptions.
}
