% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoStructure.R
\name{memoStructure}
\alias{memoStructure}
\title{COP Memo Target Table Structure}
\usage{
memoStructure(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
d object with d$memo$structure
}
\description{
COP Memo Target Table Structure
}
