% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packDataPack.R
\name{packDataPack}
\alias{packDataPack}
\title{Pack a Data Pack}
\usage{
packDataPack(datapack_uid = NA, country_uids = NA,
  template_path = NA, output_folder = getwd())
}
\arguments{
\item{datapack_uid}{Unique ID to assign Data Pack. Usually the same as
Operating Unit ID (level 3 ID). For full list of these IDs, see
\code{datapackr::dataPackMap}.}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::dataPackMap}.}

\item{template_path}{Local filepath to Data Pack template Excel (XLSX) file.
This file MUST NOT have any data validation formats present. If left
\code{NA}, will prompt for file selection via window.}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export. If left as \code{NA}, will output to
\code{Working Directory}.}
}
\value{
Exports a Data Pack to Excel within \code{output_folder}.
}
\description{
Takes a Data Pack template, combines it with data pulled from DATIM API, and
produces a Data Pack ready for distribution.
}
