% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packDataPack.R
\name{packDataPack}
\alias{packDataPack}
\title{Pack a Data Pack}
\usage{
packDataPack(model_data, datapack_name, country_uids,
  template_path = NULL, cop_year = cop_year(),
  output_folder = getwd(), results_archive = TRUE)
}
\arguments{
\item{model_data}{Data from DATIM needed to pack into Data Pack}

\item{datapack_name}{Name you would like associated with this Data Pack.
(Example: "Western Hemisphere", or "Caribbean Region", or "Kenya".)}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::dataPackMap}.}

\item{template_path}{Local filepath to Data Pack template Excel (XLSX) file.
This file MUST NOT have any data validation formats present. If left
\code{NA}, will prompt for file selection via window.}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export. If left as \code{NA}, will output to
\code{Working Directory}.}

\item{results_archive}{If TRUE, will export compiled results of all tests and
processes to output_folder.}
}
\value{
Exports a Data Pack to Excel within \code{output_folder}.
}
\description{
Takes a Data Pack template, combines it with data pulled from DATIM API, and
produces a Data Pack ready for distribution.
}
