% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packDataPackSheet.R
\name{packDataPackSheet}
\alias{packDataPackSheet}
\title{Pack data into a Data Pack sheet}
\usage{
packDataPackSheet(wb, sheet, org_units,
  schema = datapackr::data_pack_schema, sheet_data)
}
\arguments{
\item{wb}{datapackr list object.}

\item{sheet}{Specified sheet within wb}

\item{org_units}{}

\item{schema}{Defaults to standard Data Pack schema, but allows for provision
of custom schema if needed.}

\item{sheet_data}{Dataset to use as input for packing Data Pack. If left NULL,
will produce a Data Pack with orgUnits and disagg specifications, but no data.}
}
\value{
wb with specified sheet packed with data
}
\description{
Packs data into a specified Data Pack sheet
}
