% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packDataPackSheet.R
\name{packDataPackSheet}
\alias{packDataPackSheet}
\title{Pack data into a Data Pack sheet}
\usage{
packDataPackSheet(
  wb,
  sheet,
  org_units,
  schema = datapackr::data_pack_schema,
  sheet_data,
  cop_year = getCurrentCOPYear()
)
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{sheet}{Specified sheet within wb.}

\item{org_units}{A list of DATIM organisation units.}

\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{sheet_data}{Dataset to use as input for packing Data Pack. If left NULL,
will produce a Data Pack with orgUnits and disagg specifications, but no data.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
wb with specified sheet packed with data
}
\description{
Packs data into a specified Data Pack sheet
}
