% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSheets.R
\name{packDataPackSheets}
\alias{packDataPackSheets}
\title{Loop through and populate normal Data Pack sheets}
\usage{
packDataPackSheets(
  wb,
  country_uids,
  ou_level = "Prioritization",
  org_units = NULL,
  model_data = NULL,
  schema = datapackr::data_pack_schema,
  sheets = NULL,
  cop_year = getCurrentCOPYear()
)
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{ou_level}{Level in DATIM hierarchy to pull orgUnits from. Choose from:
"Prioritization", "Community", "Facility", or the numbers 4 through 7.}

\item{org_units}{A list of DATIM organisation units.}

\item{model_data}{Data from DATIM needed to pack into a COP Data Pack.}

\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{sheets}{Specified sheets within wb.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
wb with all sheets written except SNU x IM
}
\description{
Loops through all normally structured sheets in a submitted Data Pack
and writes data.
}
