% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSheets.R
\name{packDataPackSheets}
\alias{packDataPackSheets}
\title{Loop through and populate normal Data Pack sheets}
\usage{
packDataPackSheets(
  wb,
  country_uids,
  ou_level = "Prioritization",
  org_units = NULL,
  model_data = NULL,
  schema = datapackr::data_pack_schema,
  sheets = NULL,
  cop_year = getCurrentCOPYear()
)
}
\arguments{
\item{wb}{datapackr list object.}

\item{country_uids}{Character vector of Country UIDs from DATIM.}

\item{ou_level}{Level in DATIM hierarchy to pull orgUnits from. Choose from:
"Prioritization", "Community", "Facility", or the numbers 4 through 7.}

\item{org_units}{Allows for specification of custom list of orgUnits to include
in Data Pack sheets.}

\item{model_data}{Dataset to use as input for packing Data Pack. If left NULL,
will produce a Data Pack with orgUnits and disagg specifications, but no data.}

\item{schema}{Defaults to standard Data Pack schema, but allows for provision
of custom schema if needed.}

\item{sheets}{Sheets to pack. Defaults to all those available in \code{wb},
minus the first few front-matter/summary tabs.}

\item{cop_year}{COP year for dating as well as selection of
templates.}
}
\value{
wb with all sheets written except SNU x IM
}
\description{
Loops through all normally structured sheets in a submitted Data Pack
and writes data.
}
