% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSheets.R
\name{packDataPackSheets}
\alias{packDataPackSheets}
\title{Pack data into a Data Pack sheet}
\usage{
packDataPackSheets(d)
}
\arguments{
\item{d}{datapackr list object.}

\item{wb}{Openxlsx workbook object built from template.}
}
\value{
wb with all sheets written except SNU x IM
}
\description{
Loops through all normally structured sheets in a submitted Data Pack
and writes data.
}
