% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packForDATIM_OPU.R
\name{packForDATIM_OPU}
\alias{packForDATIM_OPU}
\title{packForDATIM_OPU(d)}
\usage{
packForDATIM_OPU(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
Modified d object with a DATIM compatible data frame for import id d$datim$OPU
}
\description{
Packs extracted PSNUxIM data from OPU Data Pack for DATIM import.
}
