% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packForDATIM_UndistributedMER.R
\name{packForDATIM_UndistributedMER}
\alias{packForDATIM_UndistributedMER}
\title{packForDATIM_UndistributedMER(d)}
\usage{
packForDATIM_UndistributedMER(d, datim_map)
}
\arguments{
\item{d}{Datapackr object}

\item{datim_map}{datim map object for the cop year}
}
\value{
MER data formatted as standard DATIM import file, with all data
allocated against the default mechanism UID.
}
\description{
Packs undistributed MER data from Data Pack for use in analytics (NOT IMPORT).
}
