% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packForPAW.R
\name{packForPAW}
\alias{packForPAW}
\title{packForPAW}
\usage{
packForPAW(d, type = NULL)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{type}{Type of dataset to prep for PAW. Choose from \code{PSNUxIM},
\code{SUBNAT_IMPATT}, \code{OPU PSNUxIM}, or \code{Undistributed MER}.}
}
\value{
d object
}
\description{
Function built to specifically export data to PAW
}
