% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packForPAW.R
\name{packForPAW}
\alias{packForPAW}
\title{packForPAW(data)}
\usage{
packForPAW(d, type = NA)
}
\arguments{
\item{d}{Datapackr object.}

\item{type}{Type of dataset to prep for PAW. Choose from \code{PSNUxIM} or
\code{PSNU}.}
}
\value{
Data frame ready for use in PAW
}
\description{
Prepare dataset for use in PAW.
}
