% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packForPAW.R
\name{packForPAW}
\alias{packForPAW}
\title{packForPAW}
\usage{
packForPAW(d)
}
\arguments{
\item{d}{Datapackr object.}
}
\value{
Data frame ready for use in PAW
}
\description{
Prepare dataset for use in PAW.
}
