% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packFrame.R
\name{packFrame}
\alias{packFrame}
\title{Create a Tool from scratch and prep it for data.}
\usage{
packFrame(datapack_uid, type = "Data Pack")
}
\arguments{
\item{datapack_uid}{A unique ID specifying the PEPFAR Operating Unit
the Tool belongs to.}

\item{type}{Either "Data Pack".}
}
\value{
OpenXLSX workbook object for use in data writing functions.
}
\description{
Creates a blank Excel workbook, then packs it.
}
