% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapMechanisms.R
\name{packMechMap}
\alias{packMechMap}
\title{Create Mechanism Mapping tool}
\usage{
packMechMap(datapack_uid, FY, output_path = NA)
}
\arguments{
\item{datapack_uid}{A unique ID specifying the PEPFAR Operating Unit 
the Site Tool belongs to.}

\item{FY}{Numeric value of COP Fiscal Year. (For COP19/FY20 targets, enter
\code{2019}.)}

\item{output_path}{A local path directing to the folder where you would like
outputs to be saved. If not supplied, will output to working directory.}
}
\value{
Excel workbook containing needed features to allow various mechanism
mapping scenarios, which can be fed back into the Data Pack site distribution
process.
}
\description{
Produces an Excel tool to allow PEPFAR Country Teams to map shifting
Implementing Mechanism service patterns to allow automated distribution of
site-level data between distinct mechanisms.
}
