% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packOPUDataPack.R
\name{packOPUDataPack}
\alias{packOPUDataPack}
\title{Pack an OPU Data Pack}
\usage{
packOPUDataPack(
  d,
  undistributed_mer_data = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{undistributed_mer_data}{Data from the \code{d$data$UndistributedMER}
dataset that can be provided while generating an OPU tool such that the
targets to be distributed will be sourced from this file.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Exports an OPU Data Pack to Excel within \code{output_folder}.
}
\description{
Takes an OPU Data Pack template, combines it with data pulled from DATIM
API, and produces an OPU Data Pack ready for distribution.
}
