% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSNUxIM.R
\name{packSNUxIM}
\alias{packSNUxIM}
\title{packSNUxIM(data)}
\usage{
packSNUxIM(data)
}
\arguments{
\item{data}{SNUxIM dataframe to pack for DATIM.}
}
\value{
Dataframe of SNUxIM data ready for DATIM import.
}
\description{
Packs SNUxIM data prepared from unPackSNUxIM for import to DATIM.
}
