% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSNUxIM.R
\name{packSNUxIM}
\alias{packSNUxIM}
\title{packSNUxIM(data)}
\usage{
packSNUxIM(d, d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Packs SNUxIM data prepared from unPackSNUxIM for import to DATIM.
}
