% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataPack.R
\name{packSUBNAT_IMPATT}
\alias{packSUBNAT_IMPATT}
\title{packSUBNAT_IMPATT(d)}
\usage{
packSUBNAT_IMPATT(d)
}
\arguments{
\item{d}{datapackr list object containing at least
\code{d$data$SUBNAT_IMPATT}.}
}
\value{
A datapackr list object, \code{d}, storing at least 1 dataframe of
   SUBNAT & IMPATT data (\code{d$datim$SUBNAT_IMPATT}) ready for DATIM
   ingestion.
}
\description{
Takes the outputs of the \code{\link{unPackSheets}} function and
    recompiles the dataframe containing SUBNAT and IMPATT data,
    \code{d$data$SUBNAT_IMPATT} into a standard DATIM import file.
}
