% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packSUBNAT_IMPATT.R, R/parseDataPack.R
\name{packSUBNAT_IMPATT}
\alias{packSUBNAT_IMPATT}
\title{packSUBNAT_IMPATT(data)}
\usage{
packSUBNAT_IMPATT(d)

packSUBNAT_IMPATT(d)
}
\arguments{
\item{d}{Datapackr list object}

\item{data}{SUBNAT/IMPATT dataframe to pack for DATIM.}
}
\value{
Dataframe of SUBNAT & IMPATT data ready for DATIM ingestion.

Dataframe of SUBNAT & IMPATT data ready for DATIM ingestion.
}
\description{
Takes the outputs of the \code{\link{unPackSheets}} function and
    recompiles the dataframe containing SUBNAT and IMPATT data,
    \code{d$data$SUBNAT_IMPATT} into a standard DATIM import file.

Takes the outputs of the \code{\link{unPackSheets}} function and
    recompiles the dataframe containing SUBNAT and IMPATT data,
    \code{d$data$SUBNAT_IMPATT} into a standard DATIM import file.
}
