% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataPack.R
\name{packSUBNAT_IMPATT}
\alias{packSUBNAT_IMPATT}
\title{packSUBNAT_IMPATT(d)}
\usage{
packSUBNAT_IMPATT(data)
}
\arguments{
\item{d}{datapackr list object containing at least
\code{d$data$SUBNAT_IMPATT}.}
}
\value{
Dataframe of SUBNAT & IMPATT data ready for DATIM ingestion.
}
\description{
Takes the outputs of the \code{\link{unPackSheets}} function and
    recompiles the dataframe containing SUBNAT and IMPATT data,
    \code{d$data$SUBNAT_IMPATT} into a standard DATIM import file.
}
