% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_maybe_number}
\alias{parse_maybe_number}
\title{Parse a value to numeric}
\usage{
parse_maybe_number(x, default = NULL)
}
\arguments{
\item{x}{Value to test and coerce}

\item{default}{Default value to assign to x if not a character string,
\code{NA_character_}, or a factor.}
}
\value{
x parsed as numeric, if possible
}
\description{
If x is character, attempts to parse the first occurence of a
sub-string that looks like a number.
}
