% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackCountryUIDs.R
\name{parsePSNUs}
\alias{parsePSNUs}
\title{Extract PSNUs from submitted file}
\usage{
parsePSNUs(submission_path, tool, cop_year)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{What type of tool is the submission file? Default is "Data Pack".}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}
}
\value{
Data frame of parsed PSNUs.
}
\description{
When supplied a submission path, tool type and COP year,
will return a data frame of PSNU, psnu_uid,country_name, and country_uid
If there are malformed PSNU UIDs in the file, an error will be thrown.
}
