% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{pick_schema}
\alias{pick_schema}
\title{Pick correct schema}
\usage{
pick_schema(cop_year, tool)
}
\arguments{
\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}
}
\value{
Schema file for given cop_year and tool type
}
\description{
Pick correct schema
}
