% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{pick_template_path}
\alias{pick_template_path}
\title{Pick correct template filepath}
\usage{
pick_template_path(cop_year = getCurrentCOPYear(), tool = "Data Pack")
}
\arguments{
\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{tool}{Either Data Pack or OPU Data Pack}
}
\value{
Template filepath for given cop_year and tool.
}
\description{
Pick correct template filepath
}
