% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_model_data.PSNUxIM.R
\name{prepare_model_data.PSNUxIM}
\alias{prepare_model_data.PSNUxIM}
\title{Prepare PSNUxIM Model Data}
\usage{
prepare_model_data.PSNUxIM(snuxim_model_data, country_uids)
}
\arguments{
\item{snuxim_model_data}{Export from DATIM needed to allocate data across
mechanisms in the PSNUxIM tab}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}
}
\value{
d
}
\description{
Prepare PSNUxIM Model Data
}
