% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_memo_data.R
\name{prepareExistingDataAnalytics}
\alias{prepareExistingDataAnalytics}
\title{Prepare Existing Data Analytics}
\usage{
prepareExistingDataAnalytics(
  d,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id}
}
\value{
Datapackr d object
}
\description{
Get existing PSNUxIM raw data from DATIM
and store in d$memo$datim$analytics
}
