% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMemoData.R
\name{prepareMemoData}
\alias{prepareMemoData}
\title{Prepare Memo Data}
\usage{
prepareMemoData(
  d,
  memo_type,
  include_no_prio = TRUE,
  n_cores = getMaxCores(),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{memo_type}{memo_type One of the following:
datapack: Create the memo based on the data in the datapack or OPU datapack
datim: Create the memo based on data currently in DATIM
comparison: Create a comparison memo with data from both DATIM and datapack}

\item{include_no_prio}{If TRUE, include \code{"No Prioritiation"}
as a column in the output.}

\item{n_cores}{Indicates how many cores should be used.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Datapackr d object
}
\description{
A list of tables is returned in d$memo$datapack
or d$memo$datim depending on the requested type.

by_psnu: Dataframe of indicators aggregated to the PSNU level
by_prio: Dataframe of indicators aggregated to the prioritization level
by_partner: Dataframe of indicators aggregate to the partner level
}
