% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_memo_data.R
\name{prepareMemoDataByAgency}
\alias{prepareMemoDataByAgency}
\title{Prepare Memo Data By Agency Level}
\usage{
prepareMemoDataByAgency(df, memo_structure)
}
\arguments{
\item{df}{An analytics table, either d$memo$datim$analytics or
d$memo$datapack$analytics}
}
\value{
A dataframe of COP indicators aggregated to the prioritization level
}
\description{
Prepare Memo Data By Agency Level
}
