% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMemoData.R
\name{prepareMemoDataByPartner}
\alias{prepareMemoDataByPartner}
\title{Title}
\usage{
prepareMemoDataByPartner(df, memo_structure, indicators, partners_table)
}
\arguments{
\item{df}{}

\item{memo_structure}{}

\item{indicators}{}

\item{partners_table}{}
}
\value{

}
\description{
Title
}
