% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMemoData.R
\name{prepareMemoDataByPrio}
\alias{prepareMemoDataByPrio}
\title{Prepare Memo Data By Prioritization Level}
\usage{
prepareMemoDataByPrio(df, memo_structure, include_no_prio = TRUE)
}
\arguments{
\item{df}{An analytics table, either d$memo$datim$analytics or
d$memo$datapack$analytics}

\item{include_no_prio}{If TRUE, include \code{"No Prioritiation"}
as a column in the output.}
}
\value{
A dataframe of COP indicators aggregated
to the prioritization level.
}
\description{
Prepare Memo Data By Prioritization Level
}
