% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMemoData.R
\name{prepareMemoDataByPSNU}
\alias{prepareMemoDataByPSNU}
\title{Title}
\usage{
prepareMemoDataByPSNU(analytics, memo_type)
}
\arguments{
\item{analytics}{}

\item{memo_type}{}
}
\value{

}
\description{
Title
}
