% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMemoData.R
\name{prepareMemoMetadata}
\alias{prepareMemoMetadata}
\title{Populate necessary metadata for the memo}
\usage{
prepareMemoMetadata(
  d,
  memo_type,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{memo_type}{memo_type One of the following:
datapack: Create the memo based on the data in the datapack or OPU datapack
datim: Create the memo based on data currently in DATIM
comparison: Create a comparison memo with data from both DATIM and datapack}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\value{
Datapackr d object
}
\description{
Populate necessary metadata for the memo
}
