% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparePrioTable.R
\name{preparePrioTable}
\alias{preparePrioTable}
\title{Prepare COP Memo Prioritization Target Table Data}
\usage{
preparePrioTable(d, d2_session)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id}
}
\value{
Datapackr d object with d$data$memo table
}
\description{
Prepare COP Memo Prioritization Target Table Data
}
