% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSheetData.R
\name{prepareSheetData}
\alias{prepareSheetData}
\title{Loop through and populate normal Data Pack sheets}
\usage{
prepareSheetData(sheet, row_headers, data)
}
\arguments{
\item{sheet}{Specified sheet within wb}

\item{row_headers}{Structure to use in row_headers}

\item{data}{Data from DATIM}
}
\value{
dataframe of data prepared for Data Pack
}
\description{
Loops through all normally structured sheets in a submitted Data Pack
and writes data.
}
