% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullCOPindicators.R
\name{pull_COPindicators}
\alias{pull_COPindicators}
\title{Extract COP indicators list from DATIM}
\usage{
pull_COPindicators(cop_year = datapackr::getCurrentCOPYear())
}
\arguments{
\item{cop_year}{Year of COP for which to return indicator list. (e.g., for 
COP19 enter 19.) If left blank, will use COP Year as stored in datapackr.}
}
\value{
Dataframe of COP indicators retrieved from DATIM
}
\description{
Queries DATIM to return list of COP indicators for given COP year.
}
