% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullDATIMCodeList.R
\name{pullDATIMCodeList}
\alias{pullDATIMCodeList}
\title{Pull DATIM code list.}
\usage{
pullDATIMCodeList(
  dataset,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{dataset}{Dataset uid to query.}

\item{d2_session}{DHIS2 Session id}
}
\value{
Code list as dataframe.
}
\description{
Pulls the DATIM code list for specified datasets.
}
