% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullFullCodeList.R
\name{pullFullCodeList}
\alias{pullFullCodeList}
\title{Pull & combine all MER, SUBNAT, IMPATT code lists for specified FY.}
\usage{
pullFullCodeList(FY = currentFY(), datastream = c("MER", "SUBNAT", "IMPATT"))
}
\arguments{
\item{FY}{Reporting FY for which to filter active code lists.}

\item{datastream}{Specify MER, SUBNAT, or IMPATT, or omit to specify all.}
}
\value{
Combined code list as dataframe.
}
\description{
Pulls all code lists for MER, SUBNAT, and IMPATT for a specified FY and
combines these into a unique list.
}
