% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullFullCodeList.R
\name{pullFullCodeList}
\alias{pullFullCodeList}
\title{Pull & combine all MER, SUBNAT, IMPATT code lists for specified FY.}
\usage{
pullFullCodeList(
  FY = getCurrentCOPYear() + 1,
  datastreams = c("mer_targets", "mer_results", "subnat_targets", "subnat_results",
    "impatt"),
  datasets,
  expanded = FALSE,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{FY}{Reporting FY for which to filter active code lists.}

\item{datastreams}{One or more of "mer_targets", "mer_results",
"subnat_targets", "subnat_results", or "impatt".}

\item{datasets}{Character vector of dataSet IDs.}

\item{expanded}{If TRUE, will add dataset, period, period_dataset, and
targets_results, related categoryOption metadata,  as additional columns}

\item{d2_session}{DHIS2 Session id}
}
\value{
Combined code list as dataframe.
}
\description{
Pulls all code lists for MER, SUBNAT, and IMPATT for a specified FY and
combines these into a unique list. If no datastream is specified, then
all datastreams are returned.
}
