% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemo.R
\name{renderPartnerTable}
\alias{renderPartnerTable}
\title{Render Partner-Level Memo Target Tables}
\usage{
renderPartnerTable(memo_doc, partners_table, memoStructure, source_type)
}
\arguments{
\item{memo_doc}{\code{Officer} document object containing
the target memo tables.}

\item{partners_table}{A table of target data aggregated to the partner and
mechanism level, either d$memo$datim$by_partner or
d$memo$datapack$by_partner.}

\item{memoStructure}{Structure of the memo d$memo$structure}
}
\value{

}
\description{
Render Partner-Level Memo Target Tables
}
