% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateApprovalMemo.R
\name{renderPrioTable}
\alias{renderPrioTable}
\title{Render Prioritization-Level Memo Target Table}
\usage{
renderPrioTable(memo_doc, prio_table, ou_name, source_type)
}
\arguments{
\item{memo_doc}{\code{Officer} document object containing
the target memo tables.}

\item{prio_table}{A table of target data aggregated to the Prioritization
level, either d$memo$datim$by_prio or d$memo$datapack$by_prio.}

\item{ou_name}{Name of organisation unit.}
}
\value{

}
\description{
Render Prioritization-Level Memo Target Table
}
