% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rePackSNUxIM.R
\name{rePackPSNUxIM}
\alias{rePackPSNUxIM}
\title{rePackSNUxIM(d)}
\usage{
rePackPSNUxIM(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Takes the output of the \code{\link{unPackSNUxIM}} and
    \code{\link{unPackSheets}} functions and delicatety combines these to create
    a single dataframe at the PSNU x IM level.
}
