% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rowMax}
\alias{rowMax}
\title{Extracts the desired columns for analysis via regular expression, then
 takes the maximum value row-wise. Ultimately resulting in a new column
 containing the max values.}
\usage{
rowMax(df, cn, regex)
}
\arguments{
\item{df}{The dataframe to be analyzed.}

\item{cn}{The column name (character string) of the Max column that is
created after execution of this function.}

\item{regex}{A regular expression used in identifying the columns of
interest.}
}
\value{
df
}
\description{
Extracts the desired columns for analysis via regular expression, then
 takes the maximum value row-wise. Ultimately resulting in a new column
 containing the max values.
}
