% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSchema.R
\name{schema-validations}
\alias{schema-validations}
\alias{checkSchema_SkippedSheets}
\alias{checkSchema_SheetNums}
\alias{checkSchema_SheetNames}
\alias{checkSchema_InvalidDatasets}
\alias{checkSchema_InvalidColType}
\alias{checkSchema_InvalidValueType}
\alias{checkSchema_DataElementSyntax}
\alias{checkSchema_COsSyntax}
\alias{checkSchema_ValidAges}
\alias{checkSchema_ValidSexes}
\alias{checkSchema_ValidKPs}
\alias{checkSchema_Formulas}
\alias{checkSchema}
\alias{unPackSchema}
\title{Standardized Schema Checks}
\usage{
checkSchema_SkippedSheets(schema, tool, cop_year)

checkSchema_SheetNums(schema)

checkSchema_SheetNames(schema, filepath_schema)

checkSchema_InvalidDatasets(schema, tool, cop_year)

checkSchema_InvalidColType(schema, tool, cop_year)

checkSchema_InvalidValueType(schema, tool, cop_year)

checkSchema_DataElementSyntax(schema)

checkSchema_COsSyntax(schema)

checkSchema_ValidAges(schema)

checkSchema_ValidSexes(schema)

checkSchema_ValidKPs(schema)

checkSchema_Formulas(schema)

checkSchema(schema, template_path, cop_year, tool)

unPackSchema(
  template_path = NULL,
  skip = NULL,
  tool = "Data Pack Template",
  cop_year = NULL
)
}
\arguments{
\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{filepath_schema}{Path to the schema file.}

\item{template_path}{Local filepath to Data Pack template Excel (XLSX) file.
This file MUST NOT have any data validation formats present. If left
\code{NULL}, will select the default based on \code{cop_year} and \code{tool}.}

\item{skip}{Character vector of Sheet Names to label for skipping in schema.}
}
\value{
For lower-level functions, a list of instances of failed tests. For the
higher-level \code{checkSchema}, a list object of lists of failed tests.
\itemize{
\item \code{checkSchema_SkippedSheets}:
\item \code{checkSchema_SheetNums}:
\item \code{checkSchema_SheetNames}:
\item \code{checkSchema_InvalidDatasets}:
\item \code{checkSchema_InvalidColType}:
\item \code{checkSchema_InvalidValueType}:
\item \code{checkSchema_DSDSyntax}:
\item \code{checkSchema_TASyntax}:
\item \code{checkSchema_COsSyntax} :
\item \code{checkSchema_ValidAges} :
\item \code{checkSchema_ValidSexes} :
\item \code{checkSchema_ValidKPs} :
\item \code{checkSchema_Formulas} :
\item \code{checkSchema} :
\item \code{unPackSchema} :
}
}
\description{
Standardized package functions,in terms of schema checks.
These can be run individually (e.g., \code{checkSchema_SheetNums}), or in
bulk (e.g., \code{checkSchema(template_path = "my/file/path", cop_year = 2022, tool = "Data Pack")}).
}
\concept{schema-helpers}
