% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separateDataSets.R
\name{separateDataSets}
\alias{separateDataSets}
\title{separateDataSets(d)}
\usage{
separateDataSets(d)
}
\arguments{
\item{d}{Datapackr object.}
}
\value{
d
}
\description{
After data has been extracted from all sheets in a Data Pack,
    this function separates datasets by either \code{MER} or
    \code{SUBNAT/IMPATT} and removes elements of \code{d} that are no longer
    necessary (\code{targets}, \code{extract}, and \code{sheet})
}
