% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{skip_tabs}
\alias{skip_tabs}
\title{List of tabs to skip for given tool.}
\usage{
skip_tabs(tool = "Data Pack")
}
\arguments{
\item{tool}{"Data Pack", "Data Pack Template", "Site Tool", "Site Tool Template",
"Mechanism Map", or "Site Filter".}
}
\value{
Character vector of tab names to skip.
}
\description{
List of tabs to skip for given tool.
}
