% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageSetup.R
\name{skip_tabs}
\alias{skip_tabs}
\title{List of tabs to skip for given tool.}
\usage{
skip_tabs(tool = "Data Pack", cop_year = getCurrentCOPYear())
}
\arguments{
\item{tool}{"Data Pack", "Data Pack Template", "Site Tool", "Site Tool Template".}

\item{cop_year}{COP year for dating as well as selection of
templates.}
}
\value{
Character vector of tab names to skip.
}
\description{
List of tabs to skip for given tool.
}
