% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{styleGuide}
\alias{styleGuide}
\title{Library of Openxlsx style objects to apply to both Data Pack and Site
Tool files.}
\format{\describe{
  \item{home}{Styles for title, Data Pack Name, & PEPFAR banner.}
  \item{siteList}{Styles for site names, broken out by Community, Facility,
  National, Military, or Inactive.}
  \item{data}{Styles for title, headers, labels, and Unique ID row.}
}}
\usage{
styleGuide
}
\description{
A list object containing styles pertaining to different portions of the Data
Pack.
}
\keyword{datasets}
