% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{swapColumns}
\alias{swapColumns}
\title{Swap columns between dataframes}
\usage{
swapColumns(to, from)
}
\arguments{
\item{to}{Dataframe to pull columns into}

\item{from}{Data frame to pull columns from}
}
\value{
dataframe with swapped columns
}
\description{
Replaces columns in \code{to} with those with identical names in \code{from}.
}
