% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackCountryUIDs.R
\name{unPackCountryUIDs}
\alias{unPackCountryUIDs}
\title{Extract country uids from submitted file}
\usage{
unPackCountryUIDs(
  submission_path,
  tool = "Data Pack",
  cop_year = getCurrentCOPYear()
)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
Character vector of country_uids.
}
\description{
When supplied a submission path, will return the list of country_uids
pertaining to the file, as read from the Home tab.
}
