% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackCountryUIDs.R
\name{unPackCountryUIDs}
\alias{unPackCountryUIDs}
\title{Extract country uids from submitted file}
\usage{
unPackCountryUIDs(submission_path = NA, tool = "Data Pack")
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{What type of tool is the submission file? Default is "Data Pack".
Other options include "Site Tool", and "Site Filter".}
}
\value{
Character vector of country_uids.
}
\description{
When supplied a submission path, will return the list of country_uids
pertaining to the file, as read from the Home tab.
}
