% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackData.R
\name{unPackData}
\alias{unPackData}
\title{Unpack a submitted Data Pack}
\usage{
unPackData(output_path = NA, export_FAST = FALSE,
  archive_results = FALSE, export_SUBNAT_IMPATT = FALSE)
}
\arguments{
\item{output_path}{A local path directing to the folder where you would like
outputs to be saved. If not supplied, will output to working directory.}

\item{export_FAST}{If TRUE, will extract and output to \code{output_path} a
CSV file of data needed for the PEPFAR FAST Tool.}

\item{archive_results}{If TRUE, will output to \code{output_path} a compiled
\code{datapack} list object containing all results and warning messages from
processing the selected Data Pack or Site Tool.}

\item{export_SUBNAT_IMPATT}{If TRUE, will extract and output to 
\code{output_path} a DATIM import file containing all SUBNAT and IMPATT data
from the selected Data Pack.}
}
\description{
Processes a submitted Data Pack (in .xlsx format) by identifying integrity
    issues, checking data against DATIM validations, and extracting data.
}
\details{
Executes the following operations in relation to a submitted Data Pack:
\enumerate{
    \item Performs integrity checks on file structure;
    \item Runs DATIM validation tests;
    \item Extracts SUBNAT and IMPATT data as a DATIM import file;
    \item Extracts FAST data for use by the FAST Tool; and
}
    
If a Data Pack is submitted as an XLSB formatted document, you must open &
re-save as an XLSX in order to process it with this function.

Once this function is called, it will present a dialog box where you can
select the file location of the Data Pack to be processed.

If Operating Unit \code{name} and \code{id} in the Data Pack's \strong{Home}
tab do not match based on cross-reference with DATIM organization
hierarchies, you will be prompted to manually select the correct Operating
Unit associated with the Data Pack to be processed. Enter this information in
the Console as directed.

FAST and SUBNAT/IMPATT extracts are saved to \code{output_path} as CSV files.

The final message in the Console prints all warnings identified in the Data
Pack being processed.
}
