% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackingChecks.R
\name{unPackDataChecks}
\alias{unPackDataChecks}
\alias{checkDupeRows}
\alias{checkColumnStructure}
\alias{checkToolStructure}
\alias{checkToolComments}
\alias{checkToolConnections}
\alias{checkNonNumeric}
\alias{checkMissingMetadata}
\alias{checkNegativeValues}
\alias{checkDecimalValues}
\alias{checkInvalidOrgUnits}
\alias{checkInvalidPrioritizations}
\alias{checkFormulas}
\alias{checkDisaggs}
\alias{checkSheetData}
\title{Perform various data quality checks on in-process data from Data Packs.}
\usage{
checkDupeRows(d, sheet, quiet = TRUE)

checkColumnStructure(d, sheet, quiet = TRUE)

checkToolStructure(d, quiet = TRUE)

checkToolComments(d, quiet = TRUE)

checkToolConnections(d, quiet = TRUE)

checkNonNumeric(d, sheet, quiet = TRUE)

checkMissingMetadata(d, sheet, quiet = T)

checkNegativeValues(d, sheet, quiet = T)

checkDecimalValues(d, sheet, quiet = TRUE)

checkInvalidOrgUnits(d, sheet, quiet = TRUE)

checkInvalidPrioritizations(d, sheet, quiet = T)

checkFormulas(d, sheet, quiet = TRUE)

checkDisaggs(d, sheet, quiet = TRUE)

checkSheetData(d, sheets = NULL, quiet = TRUE, ...)
}
\arguments{
\item{d}{DataPack object loaded via \code{loadDataPack}.}

\item{sheet}{String. Name of DataPack sheet to check data from. Default is
first sheet.}

\item{quiet}{Logical. Should warning messages be printed? Default is TRUE.}
}
\value{
A DataPack object, with updated tests and warnings.
}
\description{
A series of functions to check and validate quality & integrity
of data encountered in various \code{unPack...} functions across \code{datapackr}. Note
that these functions do not attempt to correct issues identified, but only to
identify them.

\code{checkToolStructure} checks structural integrity of sheets for submitted tool.

\code{checkToolComments} searches Data Pack for any comments that cause
corruption when executing openxlsx::saveWorkbook.

\code{checkToolConnections} detects the presence of any external links in a Tool.

\code{checkDupeRows} checks for any rows with duplicates across PSNU and other key
disaggregates.

\code{checkColumnStructure} checks structural integrity of columns on critical sheets for
a submitted Data Pack.

\code{checkNonNumeric} alerts to non-numeric values instead of valid data.

\code{checkMissingMetadata} alerts to missing Age, Sex, or KeyPop.

\code{checkNegativeValues} alerts to negative values.

\code{checkDecimalValues} alerts to decimal values where these are unallowed.

\code{checkInvalidOrgUnits} alerts to invalid org units.

\code{checkInvalidPrioritizations} alerts to invalid Prioritizations.

\code{checkFormulas} checks formulas in a specified sheet in a submitted Data Pack
to make sure they are up to date and have not been tampered with.

\code{checkDisaggs} alerts to invalid disaggs (Age, Sex, KeyPop).

Some functions (\code{checkToolStructure}, \code{checkToolComments}, &
\code{checkToolConnections}) are designed to check across entire tools at
once. All others are specific to single sheet.
}
