% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackingChecks.R
\name{unPackDataChecks}
\alias{unPackDataChecks}
\alias{checkToolStructure}
\alias{checkToolComments}
\alias{checkToolConnections}
\alias{checkDupeRows}
\alias{checkMissingCols}
\alias{checkDupeCols}
\alias{checkOutOfOrderCols}
\alias{checkNonNumeric}
\alias{checkNegativeValues}
\alias{checkDecimalValues}
\alias{checkInvalidOrgUnits}
\alias{checkInvalidPrioritizations}
\alias{checkFormulas}
\alias{checkDisaggs}
\alias{checkToolEmptySheets}
\alias{checkSheetData}
\title{Perform various data quality checks on in-process data from Data Packs.}
\usage{
checkToolStructure(d, quiet = TRUE)

checkToolComments(d, quiet = TRUE)

checkToolConnections(d, quiet = TRUE)

checkDupeRows(sheets, d, quiet = TRUE)

checkMissingCols(sheets, d, quiet = TRUE)

checkDupeCols(sheets, d, quiet = TRUE)

checkOutOfOrderCols(sheets, d, quiet = TRUE)

checkNonNumeric(sheets, d, quiet = TRUE)

checkNegativeValues(sheets, d, quiet = TRUE)

checkDecimalValues(sheets, d, quiet = TRUE)

checkInvalidOrgUnits(sheets, d, quiet = TRUE)

checkInvalidPrioritizations(sheets, d, quiet = TRUE)

checkFormulas(sheets, d, quiet = TRUE)

checkDisaggs(sheets, d, quiet = TRUE)

checkToolEmptySheets(d, sheets, quiet = TRUE)

checkSheetData(d, sheets = NULL, quiet = TRUE, ...)
}
\arguments{
\item{d}{DataPack object loaded via \code{loadDataPack}.}

\item{quiet}{Logical. Should warning messages be printed? Default is TRUE.}

\item{sheet}{String. Name of DataPack sheet to check data from. Default is
first sheet.}
}
\value{
A DataPack object, with updated tests and warnings.
}
\description{
A series of functions to check and validate quality & integrity
of data encountered in various \code{unPack...} functions across \code{datapackr}. Note
that these functions do not attempt to correct issues identified, but only to
identify them.

\code{checkToolStructure} checks structural integrity of sheets for submitted tool.

\code{checkToolComments} searches Data Pack for any comments that cause
corruption when executing openxlsx::saveWorkbook.

\code{checkToolConnections} detects the presence of any external links in a Tool.

\code{checkToolEmptySheets} detects whether a sheet is essentially empty (no data
in rows, or no data in row 14 column headers).

\code{checkDupeRows} checks for any rows with duplicates across PSNU and other key
disaggregates.

\code{checkColumnStructure} checks structural integrity of columns on critical sheets for
a submitted Data Pack.

\code{checkNonNumeric} alerts to non-numeric values instead of valid data.

\code{checkMissingMetadata} alerts to missing Age, Sex, or KeyPop.

\code{checkNegativeValues} alerts to negative values.

\code{checkDecimalValues} alerts to decimal values where these are unallowed.

\code{checkInvalidOrgUnits} alerts to invalid org units.

\code{checkInvalidPrioritizations} alerts to invalid Prioritizations.

\code{checkFormulas} checks formulas in a specified sheet in a submitted Data Pack
to make sure they are up to date and have not been tampered with.

\code{checkDisaggs} alerts to invalid disaggs (Age, Sex, KeyPop).

Some functions (\code{checkToolStructure}, \code{checkToolComments}, &
\code{checkToolConnections}) are designed to check across entire tools at
once. All others are specific to single sheet.
}
