% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackingChecks.R
\name{unPackDataChecks}
\alias{unPackDataChecks}
\alias{checkDupeRows}
\alias{checkColumnStructure}
\alias{checkToolStructure}
\alias{checkToolComments}
\alias{checkConnections}
\alias{checkNumeric}
\alias{checkSheetData}
\title{Perform various data quality checks on in-process data from Data Packs.}
\usage{
checkDupeRows(d, sheet, quiet = TRUE)

checkColumnStructure(d, sheet, quiet = TRUE)

checkToolStructure(d, quiet = TRUE)

checkToolComments(d, quiet = TRUE)

checkConnections(d, quiet = TRUE)

checkNumeric(d, sheet, quiet = TRUE)

checkSheetData(d, sheets = NULL, quiet = TRUE, ...)
}
\arguments{
\item{d}{DataPack object loaded via \code{loadDataPack}.}

\item{sheet}{String. Name of DataPack sheet to check data from. Default is
first sheet.}

\item{quiet}{Logical. Should warning messages be printed? Default is TRUE.}

\item{compile}{Logical. Should results be compiled back into the d object, or
returned as a test results object containing \code{test_results}, \code{messages},
and \code{error} flag? Default is TRUE}
}
\value{
If \code{compile = TRUE}, a DataPack object, with updated tests and warnings.
If \code{compile = FALSE}, a test results object containing \code{test_results},
\code{messages}, and \code{error} flag.
}
\description{
A series of functions to check and validate quality & integrity
of data encountered in various \code{unPack...} functions across \code{datapackr}. Note
that these functions do not attempt to correct issues identified, but only to
identify them.

\code{checkDupeRows} checks for any rows with duplicates across PSNU and other key
disaggregates.

\code{checkColumnStructure} checks structural integrity of columns on critical sheets for
a submitted Data Pack.

\code{checkToolStructure} checks structural integrity of sheets for submitted tool.

\code{checkToolComments} searches Data Pack for any comments that cause
corruption when executing openxlsx::saveWorkbook.

\code{checkConnections} detects the presence of any external links in a Tool.

\code{checkNumeric} alerts to non-numeric values instead of valid data.
}
