% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackDataPackName.R
\name{unPackDataPackName}
\alias{unPackDataPackName}
\title{Extract the name of the datapack}
\usage{
unPackDataPackName(submission_path = NA, tool = "Data Pack")
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{What type of tool is the submission file? Default is "Data Pack".
Other options include "Site Tool", and "Site Filter".
javascript:;}
}
\value{
Character vector of the name of the data pack.
}
\description{
When supplied a submission path, will return the name of the datapack.
}
