% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSheet.R
\name{unPackDataPackSheet}
\alias{unPackDataPackSheet}
\title{Unpack a Data Pack sheet.}
\usage{
unPackDataPackSheet(d, sheet)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
d
}
\description{
Within a submitted Data Pack (directed to by
   \code{d$keychain$submission_path}), extract data from a single sheet specified
   in \code{d$data$sheet}.
}
