% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSheet.R
\name{unPackDataPackSheet}
\alias{unPackDataPackSheet}
\title{Unpack a Data Pack sheet.}
\usage{
unPackDataPackSheet(
  d,
  sheet,
  clean_orgs = TRUE,
  clean_disaggs = TRUE,
  clean_values = TRUE
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{clean_orgs}{Logical. If TRUE, drops rows with missing or invalid org
units.}

\item{clean_disaggs}{Logical. If TRUE, drops rows with missing or incorrect
Age, Sex, or KeyPop.}

\item{clean_values}{Logical. If TRUE, will convert values to numeric type,
drop non-numerics, NAs, & negatives; and aggregate all data across duplicate
rows.}
}
\value{
d
}
\description{
Within a submitted Data Pack (directed to by
   \code{d$keychain$submission_path}), extract data from a single sheet specified
   in \code{d$data$sheet}.
}
