% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackOPUDataPack.R
\name{unPackOPUDataPack}
\alias{unPackOPUDataPack}
\title{Unpack a submitted OPU Data Pack}
\usage{
unPackOPUDataPack(
  d,
  pzns = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{datapackr sidecar}

\item{d2_session}{DHIS2 Session id}
}
\description{
Processes a submitted OPU Data Pack (in .xlsx format) by identifying integrity
    issues, checking data against DATIM validations, and extracting data.
}
\details{
Once this function is called, it will present a dialog box where you can
select the file location of the Data Pack to be processed.

If Operating Unit \code{name} and \code{id} in the Data Pack's \strong{Home}
tab do not match based on cross-reference with DATIM organization
hierarchies, you will be prompted to manually select the correct Operating
Unit associated with the Data Pack to be processed. Enter this information in
the Console as directed.

The final message in the Console prints all warnings identified in the Data
Pack being processed.
}
