% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSchema.R
\name{unPackSchema_datapack}
\alias{unPackSchema_datapack}
\title{Extract and save schema from Data Pack template.}
\usage{
unPackSchema_datapack(
  filepath = NULL,
  skip = NULL,
  tool = "Data Pack Template",
  cop_year = getCurrentCOPYear(),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{filepath}{Local filepath for a Data Pack template (XLSX).}

\item{skip}{Character vector of Sheet Names to label for skipping in schema.}

\item{tool}{Type of tool to unpack.}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{d2_session}{R6 datimutils object which handles authentication with DATIM}
}
\value{
Data Pack schema.
}
\description{
Supplied a filepath to a Data Pack template (XLSX), will extract and save a
schema based on the template.
}
