% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSchema.R
\name{unPackSchema_datapack}
\alias{unPackSchema_datapack}
\title{Extract and save schema from Data Pack template.}
\usage{
unPackSchema_datapack(
  filepath = NULL,
  skip = NULL,
  tool = "Data Pack Template",
  cop_year = getCurrentCOPYear()
)
}
\arguments{
\item{filepath}{Local filepath for a Data Pack template (XLSX).}

\item{skip}{Character vector of Sheet Names to label for skipping in schema.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
Data Pack schema.
}
\description{
Supplied a filepath to a Data Pack template (XLSX), will extract and save a
schema based on the template.
}
