% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSchema.R
\name{unPackSchema_datapack}
\alias{unPackSchema_datapack}
\title{Extract and save schema from Data Pack template.}
\usage{
unPackSchema_datapack(filepath, skip = NA)
}
\arguments{
\item{skip}{Character vector of Sheet Names to label for skipping in schema.}

\item{path}{Local filepath for a Data Pack template (XLSX).}
}
\value{
Data Pack schema.
}
\description{
Supplied a filepath to a Data Pack template (XLSX), will extract and save a
schema based on the template.
}
