% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSchema.R
\name{unPackSchema_datapack}
\alias{unPackSchema_datapack}
\title{Extract and save schema from Data Pack template.}
\usage{
unPackSchema_datapack(filepath = NA, skip = NA,
  cop_year = cop_year())
}
\arguments{
\item{filepath}{Local filepath for a Data Pack template (XLSX).}

\item{skip}{Character vector of Sheet Names to label for skipping in schema.}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}
}
\value{
Data Pack schema.
}
\description{
Supplied a filepath to a Data Pack template (XLSX), will extract and save a
schema based on the template.
}
