% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSheets.R
\name{unPackSheets}
\alias{unPackSheets}
\title{Unpack data from Data Pack sheets.}
\usage{
unPackSheets(d, sheets = NULL, check_sheets = TRUE, separate_datasets = TRUE)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{sheets}{Character vector. Names of sheets/tabs within tool.}

\item{check_sheets}{Logical. Should sheet data be validated?}

\item{separate_datasets}{Logical. Should datasets be separated?}
}
\value{
d
}
\description{
Loops through all critical sheets in a submitted Data Pack
and extracts data, then compiles into single flat dataframe.
}
