% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSheets.R
\name{unPackSheets}
\alias{unPackSheets}
\title{Unpack data from a submission file.}
\usage{
unPackSheets(d)
}
\arguments{
\item{d}{Datapackr object.}
}
\value{
d
}
\description{
Loops through all critical sheets in a submitted 
    Tool and extracts data, then compiles data into single flat dataframe.
}
