% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSheets.R
\name{unPackSheets}
\alias{unPackSheets}
\title{Unpack data from Data Pack sheets.}
\usage{
unPackSheets(d, d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id}
}
\value{
d
}
\description{
Loops through all critical sheets in a submitted Data Pack
and extracts data, then compiles into single flat dataframe.
}
