% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSheets.R
\name{unPackSheets}
\alias{unPackSheets}
\title{Unpack data from Data Pack or Site Tool sheets.}
\usage{
unPackSheets(d)
}
\arguments{
\item{d}{Datapackr object}
}
\value{
d
}
\description{
Loops through all critical sheets in a submitted Data Pack or
Site Tool and extracts data, then compiles into single flat dataframe.
}
