% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSiteTool.R
\name{unPackSiteToolSheet}
\alias{unPackSiteToolSheet}
\title{unPackSiteToolSheet(d)}
\usage{
unPackSiteToolSheet(d)
}
\arguments{
\item{d}{datapackr list object containing at least
\code{d$keychain$submission_path},
\code{d$data$sheet}, & \code{d$info$warningMsg}.}
}
\value{
A datapackr list object, \code{d}, storing a dataframe of extracted
   data (\code{d$data$extract}) and a warning message
   (\code{d$info$warningMsg}) compiled with errors discovered while
   extracting data from sheet specified in \code{d$data$sheet}.
}
\description{
Within a submitted Site Tool (directed to by
   \code{d$keychain$submission_path}), extract data from a single sheet specified
   in \code{d$data$sheet}.
}
