% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDataPack.R
\name{unPackSNUxIM}
\alias{unPackSNUxIM}
\title{unPackSNUxIM(d)}
\usage{
unPackSNUxIM(d)
}
\arguments{
\item{d}{datapackr list object containing at least
\code{d$keychain$submission_path}.}
}
\value{
A datapackr list object, \code{d}, storing at least 1 dataframe of
   data extracted from the \code{SNU x IM} tab in a submitted Data Pack or
   Site Tool, as well as a running Warning Message queue string,
   \code{d$info$warningMsg}
}
\description{
Looks inside submitted Data Pack to extract SNU x IM data from
    \code{SNU x IM} tab and restructure this to be ready for cross-
    pollination with PSNU-level MER data coming from
    \code{\link{unPackSheets}}. This data is also analyzed to identify
    structural or data anomalies and print any issues into running Warning
    Message queue.
}
