% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackTool.R
\name{unPackTool}
\alias{unPackTool}
\title{Unpack a submitted tool}
\usage{
unPackTool(
  submission_path = NULL,
  tool = "Data Pack",
  country_uids = NULL,
  cop_year = getCurrentCOPYear(),
  snuxim_model_data_path,
  output_folder = getwd()
)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{What type of tool is the submission file? Default is "Data Pack".
Other options include "Site Tool", "Mechanism Map", and "Site Filter".}

\item{country_uids}{List of 11 digit alphanumeric DATIM codes representing
countries. If not provided, will check file for these codes. If not in file,
will flag error.}

\item{cop_year}{Specifies COP year for dating as well as selection of
templates.}

\item{snuxim_model_data_path}{Filepath where SNU x IM distribution model is stored.}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export.}
}
\description{
Processes a submitted Data Pack, Site Tool, Mechanism Map, or Site Filter by
identifying integrity issues, checking data against DATIM validations, and
extracting data.
}
\details{
Executes the following operations in relation to a submitted Site Tool
\enumerate{
    \item Performs integrity checks on file structure;
}
}
