% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackTool.R
\name{unPackTool}
\alias{unPackTool}
\title{Unpack a submitted tool}
\usage{
unPackTool(
  submission_path = NULL,
  tool = NULL,
  country_uids = NULL,
  cop_year = NULL,
  pzns = NULL,
  mer_data = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{submission_path}{Local path to the file to import.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_OrgUnits}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{pzns}{A object containing prioiritization information by organisation unit
to be added to the DataPack object.}

\item{mer_data}{If prior MER data (for instance from DATIM), is passed to this
parameter, it will be used. Otherwise, data from the existing set of tools
will be used.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}
}
\description{
Processes a submitted Data Pack by identifying integrity issues, checking
data against DATIM validations, and extracting data.
}
\details{
Executes the following operations in relation to a submitted tool
\enumerate{
    \item Performs integrity checks on file structure;
}
}
