% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateExistingPrioritization.R
\name{updateExistingPrioritization}
\alias{updateExistingPrioritization}
\title{Title}
\usage{
updateExistingPrioritization(d, d2_session)
}
\arguments{
\item{d}{}

\item{d2_session}{}
}
\value{

}
\description{
Title
}
