% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_existing_prioritization.R
\name{updateExistingPrioritization}
\alias{updateExistingPrioritization}
\title{Update Existing Prioritization}
\usage{
updateExistingPrioritization(
  d,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{d2_session}{DHIS2 Session id}
}
\value{

}
\description{
Update Existing Prioritization
}
