% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePSNUxIMTargetValues.R
\name{updatePSNUxIMTargetValues}
\alias{updatePSNUxIMTargetValues}
\title{Update PSNUxIM Target Values}
\usage{
updatePSNUxIMTargetValues(d)
}
\arguments{
\item{d}{Datapackr sidecar object}
}
\value{
d object with a PSNUxIM workbook object.
Target values in the "Data Pack Target" column will be updated with values
from the main DataPack tabs.
}
\description{
Update PSNUxIM Target Values
}
