% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsForPDAP.R
\name{uploadDATIMExportToPDAP}
\alias{uploadDATIMExportToPDAP}
\title{Upload DATIM Export to PDAP}
\usage{
uploadDATIMExportToPDAP(raw_file, job_type, content_type = "text/csv")
}
\arguments{
\item{raw_file}{Raw binary data of the CSV file to upload}

\item{job_type}{The type of job to upload the data to. Currently only
target_setting_tool or year_two_targets are supported.}
}
\value{
Returns the S3 file location of the uploaded file if successful,
otherwise NULL
}
\description{
Extracts current COP year data from the d object,
formats it, and uploads the data to PDAP.
}
