% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_category_options}
\alias{valid_category_options}
\title{List of  valid categoryoptions based on current COP Target Code List.}
\format{
\describe{
  \item{name}{}
  \item{id}{}
  \item{categoryoptiongroup}{}
  \item{datapack_disagg}{}
  \item{datapack_schema_group}{}
}
}
\usage{
valid_category_options
}
\description{
List of valid categoryoptions based on current COP Target Code
List. Must be synced and saved manually.
}
\keyword{datasets}
