% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unPackSchema.R
\name{validateSchema}
\alias{validateSchema}
\title{Validate a `datapackr` Schema object}
\usage{
validateSchema(schema, template_path, cop_year, tool, season)
}
\arguments{
\item{schema}{Which datapackr schema to use in guiding this function. If left
\code{NULL} will select the default based on \code{cop_year} and \code{tool}.}

\item{template_path}{Local filepath to Data Pack template Excel (XLSX) file.
This file MUST NOT have any data validation formats present. If left
\code{NULL}, will select the default based on \code{cop_year} and \code{tool}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{season}{Either \code{COP} or \code{OPU}.}
}
\value{
List object with any failed Schema validation checks.
}
\description{
Useful in validating a `datapackr` schema object and investigating any issues.
}
\seealso{
Other schema-helpers: 
\code{\link{unPackSchema_datapack}()}
}
\concept{schema-helpers}
