% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePacks.R
\name{writeComparisonWorkbook}
\alias{writeComparisonWorkbook}
\title{Write a comparison workbook}
\usage{
writeComparisonWorkbook(d, filename)
}
\arguments{
\item{d}{List of data frames from comparePacks}

\item{filename}{Filename where you want Comparison workbooks written.}
}
\description{
Writes an XLSX workbook from a object from comparePacks
}
