% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHomeTab.R
\name{writeHomeTab}
\alias{writeHomeTab}
\title{Write Home Tab}
\usage{
writeHomeTab(
  wb = NULL,
  datapack_name = NULL,
  country_uids,
  cop_year = getCurrentCOPYear(),
  tool = "Data Pack"
)
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{datapack_name}{Name you would like associated with this Data Pack.
(Example: "Western Hemisphere", or "Caribbean Region", or "Kenya".)}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.
For full list of these IDs, see \code{datapackr::valid_PSNUs}.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}
}
\value{
Openxlsx workbook object with added, styled Home tab.
}
\description{
Function to write Home tab details into Data Pack as specified.
}
