% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHomeTab.R
\name{writeHomeTab}
\alias{writeHomeTab}
\title{writeHomeTab(wb, datapack_uid, tool = "Data Pack")}
\usage{
writeHomeTab(
  wb = NULL,
  datapack_name = NULL,
  country_uids,
  cop_year = getCurrentCOPYear(),
  tool = "Data Pack",
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{datapack_name}{Name you would like associated with this Data Pack.
(Example: "Western Hemisphere", or "Caribbean Region", or "Kenya".)}

\item{country_uids}{Unique IDs for countries to include in the Data Pack.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}

\item{tool}{Type of tool this function will create or interact with. Either
\code{OPU Data Pack} or \code{Data Pack}}

\item{d2_session}{DHIS2 Session id}
}
\value{
Openxlsx workbook object with added, styled Home tab.
}
\description{
Function to write Home tab details into Data Pack as specified.
}
