% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHomeTab.R
\name{writeHomeTab}
\alias{writeHomeTab}
\title{writeHomeTab(wb, datapack_uid, tool = "Data Pack")}
\usage{
writeHomeTab(
  wb,
  datapack_name,
  country_uids,
  cop_year = getCurrentCOPYear(),
  tool = "Data Pack"
)
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{datapack_name}{Name you would like associated with this Data Pack.
(Example: "Western Hemisphere", or "Caribbean Region", or "Kenya".)}

\item{country_uids}{Character vector of 11 digit alphanumeric DATIM codes
representing countries.}

\item{cop_year}{COP Year in format YYYY.}

\item{tool}{Defaults to "Data Pack".}
}
\value{
Openxlsx workbook object with added, styled Home tab.
}
\description{
Function to write Home tab details into Data Pack as specified.
}
