% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePSNUxIM.R
\name{writePSNUxIM}
\alias{writePSNUxIM}
\title{Write PSNUxIM Tab}
\usage{
writePSNUxIM(
  d,
  snuxim_model_data_path = NULL,
  output_folder = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE),
  append = TRUE,
  use_template = FALSE
)
}
\arguments{
\item{d}{Datapackr sidecar object}

\item{snuxim_model_data_path}{Export from DATIM needed to allocate data
across mechanisms in the PSNUxIM tab}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export.}

\item{d2_session}{DHIS2 Session id. R6 datimutils object which handles
authentication with DATIM.}

\item{append}{If TRUE append rows to the existing DataPack otherwise,
output a Missing PSNUxIM targets workbook.}

\item{use_template}{If TRUE, use a template to create the object, otherwise
use the existing workbook object which is part of the main d object.}
}
\value{
d
}
\description{
Checks a Data Pack for need of new or appended PSNUxIM data, then
writes this into the Data Pack supplied. unPackTool must be run as prerequisite.
}
