% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePSNUxIM.R
\name{writePSNUxIM}
\alias{writePSNUxIM}
\title{writePSNUxIM(d)}
\usage{
writePSNUxIM(d, snuxim_model_data_path = NULL, output_folder = NULL)
}
\arguments{
\item{d}{Datapackr object}

\item{snuxim_model_data_path}{Filepath where SNU x IM distribution model is stored.}

\item{output_folder}{Local folder where you would like your Data Pack to be
saved upon export.}
}
\value{
d
}
\description{
Checks a Data Pack for need of new or appended PSNUxIM data, then
writes this into the Data Pack supplied. unPackTool must be run as prerequisite.
}
