% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSpectrumData.R
\name{writeSpectrumData}
\alias{writeSpectrumData}
\title{writeSpectrumData}
\usage{
writeSpectrumData(wb, spectrum_data, cop_year = NULL)
}
\arguments{
\item{wb}{Openxlsx workbook object.}

\item{spectrum_data}{Spectrum output, as a dataframe.}

\item{cop_year}{COP Year to use for tailoring functions. Remember,
FY22 targets = COP21.}
}
\value{
Modified openxlsx workbook object with Spectrum data written to the
Spectrum tab.
}
\description{
Utility function mostly for testing which will
write Spectrum data in a defined format to the Spectrum tab.
}
